% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{baseballr-package}
\alias{baseballr}
\alias{baseballr-package}
\title{baseballr: Acquiring and Analyzing Baseball Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides numerous utilities for acquiring and analyzing baseball data from online sources such as 'Baseball Reference' \url{https://www.baseball-reference.com/}, 'FanGraphs' \url{https://www.fangraphs.com/}, and the 'MLB Stats' API \url{https://www.mlb.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://billpetti.github.io/baseballr/}
  \item \url{https://github.com/BillPetti/baseballr}
  \item Report bugs at \url{https://github.com/BillPetti/baseballr/issues}
}

}
\author{
\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com} (\href{https://orcid.org/0000-0002-7194-9067}{ORCID})

Authors:
\itemize{
  \item Saiem Gilani \email{saiem.gilani@gmail.com} (\href{https://orcid.org/0000-0002-7194-9067}{ORCID})
  \item Bill Petti \email{billpetti@gmail.com}
}

Other contributors:
\itemize{
  \item Ben Baumer \email{ben.baumer@gmail.com} [contributor]
  \item Ben Dilday \email{ben.dilday.phd@gmail.com} [contributor]
  \item Robert Frey \email{rfrey22@gmail.com} [contributor]
  \item Camden Kay \email{camden.kay23@gmail.com} [contributor]
  \item baseballr authors [copyright holder]
}

}
\keyword{internal}
