% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_batter_game_logs.R
\name{batter_game_logs_fg}
\alias{batter_game_logs_fg}
\title{\strong{(legacy) Scrape Batter Game Logs from FanGraphs}}
\usage{
batter_game_logs_fg(playerid, year)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\value{
A data frame of batter game logs.
}
\description{
\strong{(legacy) Scrape Batter Game Logs from FanGraphs}
}
\keyword{legacy}
