% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref.R
\name{bref}
\alias{bref}
\title{\strong{Baseball Reference Functions Overview}}
\description{
\describe{
\item{\code{bref_daily_batter()}}{Scrape Batter Performance Data Over a Custom Time Frame}
\item{\code{bref_daily_pitcher()}}{Scrape Pitcher Performance Data Over a Custom Time Frame}
\item{\code{bref_standings_on_date()}}{Scrape MLB Standings on a Given Date}
\item{\code{bref_team_results()}}{Scrape Team Results}
}
}
\details{
\subsection{\strong{Scrape Batter Performance Data Over a Custom Time Frame}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  bref_daily_batter("2015-05-10", "2015-06-20")
}\if{html}{\out{</div>}}
}

\subsection{\strong{Scrape Pitcher Performance Data Over a Custom Time Frame}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  bref_daily_batter("2015-05-10", "2015-06-20")
}\if{html}{\out{</div>}}
}

\subsection{\strong{Scrape MLB Standings on a Given Date}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  bref_standings_on_date(date = "2015-08-04", division = "AL East")
}\if{html}{\out{</div>}}
}

\subsection{\strong{Scrape Team Results}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  bref_team_results("NYM", 2015)
  bref_team_results(Tm="TBR", year=2008)
}\if{html}{\out{</div>}}
}

\subsection{\strong{Team Level Consistency}}{

Uses \code{bref_team_results()} to calculate team consistency metrics\if{html}{\out{<div class="sourceCode r">}}\preformatted{  team_consistency(year=2015)
}\if{html}{\out{</div>}}
}
}
