% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick.R
\name{chadwick}
\alias{chadwick}
\title{\strong{Chadwick Bureau Register Player Lookup}}
\description{
\describe{
\item{\code{chadwick_player_lu()}:}{ Directly download the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.}
\item{\code{playername_lookup()}:}{ Look up Baseball Player Name.}
\item{\code{playerid_lookup()}:}{ Look up Baseball Player IDs.}
}
}
\details{
\subsection{\strong{Directly download the Chadwick Bureau's public register of baseball players.}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  chadwick_player_lu()
}\if{html}{\out{</div>}}
}

\subsection{\strong{Look up baseball player name by ID}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  playername_lookup(4885)
  playername_lookup("kaaihki01")
}\if{html}{\out{</div>}}
}

\subsection{\strong{Look up baseball player IDs by player name}}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{  playerid_lookup("Garcia", "Karim")
}\if{html}{\out{</div>}}
}
}
