% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick_playerid_lookup.R
\name{playerid_lookup}
\alias{playerid_lookup}
\title{\strong{Look up Baseball Player IDs by Player Name}}
\usage{
playerid_lookup(last_name = NULL, first_name = NULL)
}
\arguments{
\item{last_name}{A text string used to return results for players with that string in their last name.}

\item{first_name}{A text string used to return results for players with that string in their first name.}
}
\value{
A data frame of baseball players and the various IDs associated with them in different systems of record.\tabular{lll}{
   col_name \tab types \tab description \cr
   first_name \tab character \tab Player first name. \cr
   last_name \tab character \tab Player last name. \cr
   given_name \tab character \tab Player full given (legal) name. \cr
   name_suffix \tab character \tab Name suffix (e.g. Jr., Sr., III). \cr
   nick_name \tab character \tab Player nickname. \cr
   birth_year \tab integer \tab Year of birth. \cr
   mlb_played_first \tab integer \tab First MLB season as a player. \cr
   mlbam_id \tab integer \tab MLB Advanced Media (MLBAM) player ID. \cr
   retrosheet_id \tab character \tab Retrosheet player ID. \cr
   bbref_id \tab character \tab Baseball-Reference player ID. \cr
   fangraphs_id \tab integer \tab FanGraphs player ID. \cr
}
}
\description{
This function allows you to query the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.
}
\examples{
\donttest{
  try(playerid_lookup("Garcia", "Karim"))
}
}
