% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_code_barrel.R
\name{code_barrel}
\alias{code_barrel}
\title{\strong{Helper for determining whether a batted ball is a "barrel"}}
\usage{
code_barrel(df)
}
\arguments{
\item{df}{A dataframe generated by baseballsavant.mlb.com that must include the following variables: \code{launch_angle} and \code{launch_speed}.}
}
\value{
Returns a tibble with the additional column, \code{barrel}.
}
\description{
This function allows you to code a batted ball as a barrel as defined by the Statcast research team using data as provided by baseballsavant.mlb.com.
}
