% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{column_structure_draft_mlb}
\alias{column_structure_draft_mlb}
\title{Column Structure of MLB Draft data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 72 columns.
}
\usage{
column_structure_draft_mlb
}
\description{
A tibble
}
\keyword{datasets}
