% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csv_from_url}
\alias{csv_from_url}
\title{\strong{Load .csv / .csv.gz file from a remote connection}}
\usage{
csv_from_url(...)
}
\arguments{
\item{...}{passed to data.table::fread}
}
\description{
This is a thin wrapper on data.table::fread
}
\keyword{internal}
