% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_daily_pitcher.R
\name{daily_pitcher_bref}
\alias{daily_pitcher_bref}
\title{\strong{(legacy) Scrape Pitcher Performance Data Over a Custom Time Frame}}
\usage{
daily_pitcher_bref(t1, t2)
}
\arguments{
\item{t1}{First date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}

\item{t2}{Last date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}
}
\value{
Returns a tibble of pitcher performance
}
\description{
\strong{(legacy) Scrape Pitcher Performance Data Over a Custom Time Frame}
}
\seealso{
\code{bref_daily_pitcher()}
}
\keyword{legacy}
