% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_espn.R
\name{.resp_text}
\alias{.resp_text}
\title{Extract an httr2 response body as UTF-8 text}
\usage{
.resp_text(resp)
}
\arguments{
\item{resp}{An httr2 response object.}
}
\value{
Character string of the response body.
}
\description{
Extract an httr2 response body as UTF-8 text
}
\keyword{internal}
