% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_code.R
\name{edge_code}
\alias{edge_code}
\title{Edge Percentage Scrape}
\usage{
edge_code(df)
}
\arguments{
\item{df}{A dataframe that, at a minimum, includes the following columns: batter height (b_height), vertical location of the pitch (pz), and then horizontal location of the pitch (pz)}
}
\description{
This function allows you to classify individual pitches based on the various categories from the Edge% metric. The dataframe passed to the function must include the batter's handedness, the px and pz coordinates from the PITCHf/x system, and the batter's height.
}
\examples{
\dontrun{edge_code(df)}
}
\keyword{MLB,}
\keyword{PITCHf/x}
\keyword{sabermetrics,}

