% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_edge_scrape.R
\name{edge_scrape}
\alias{edge_scrape}
\title{\strong{Edge Percentage Scrape}}
\usage{
edge_scrape(start, end, group)
}
\arguments{
\item{start}{First date in your date range. Must be a character string in the format "yyyy-mm-dd".}

\item{end}{Last date in your date range. Must be a character string in the format "yyyy-mm-dd".}

\item{group}{Character string indicating whether to group the output by pitchers or batters. Options are "pitcher" or "batter".}
}
\description{
This function allows you to calculate the percent of pitches thrown to different edges of the strike zone over any custom time frame. Data is acquired by scraping the GameDay application from MLBAM using Carson Sievert's pitchRx package. The data can take a while to query, particular for large date ranges.
}
