% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_college_baseball_groups.R
\name{espn_college_baseball_conferences}
\alias{espn_college_baseball_conferences}
\title{\strong{Get ESPN College Baseball Conferences}}
\usage{
espn_college_baseball_conferences()
}
\value{
A \code{baseballr_data} tibble with one row per conference:\tabular{lll}{
   col_name \tab types \tab description \cr
   group_id \tab integer \tab Group identifier (e.g. conference group_id). \cr
   conference_short_name \tab character \tab Conference short name (e.g. 'ACC'). \cr
   conference_uid \tab character \tab ESPN universal conference identifier. \cr
   conference_name \tab character \tab Full conference name. \cr
   conference_logo \tab character \tab Logo image URL for conference. \cr
   parent_group_id \tab integer \tab Unique identifier for parent group. \cr
   conference_id \tab integer \tab Conference identifier. \cr
}
}
\description{
\strong{Get ESPN College Baseball Conferences}

\strong{Get ESPN College Baseball Conferences}
}
\examples{
\donttest{
  try(espn_college_baseball_conferences())
}
}
\seealso{
Other ESPN College Baseball Functions:
\code{\link[=espn_college_baseball_athletes_index]{espn_college_baseball_athletes_index()}},
\code{\link[=espn_college_baseball_calendar]{espn_college_baseball_calendar()}},
\code{\link[=espn_college_baseball_coach]{espn_college_baseball_coach()}},
\code{\link[=espn_college_baseball_coach_record]{espn_college_baseball_coach_record()}},
\code{\link[=espn_college_baseball_coach_season]{espn_college_baseball_coach_season()}},
\code{\link[=espn_college_baseball_coaches]{espn_college_baseball_coaches()}},
\code{\link[=espn_college_baseball_game_all]{espn_college_baseball_game_all()}},
\code{\link[=espn_college_baseball_game_broadcasts]{espn_college_baseball_game_broadcasts()}},
\code{\link[=espn_college_baseball_game_official_detail]{espn_college_baseball_game_official_detail()}},
\code{\link[=espn_college_baseball_game_officials]{espn_college_baseball_game_officials()}},
\code{\link[=espn_college_baseball_game_play]{espn_college_baseball_game_play()}},
\code{\link[=espn_college_baseball_game_play_personnel]{espn_college_baseball_game_play_personnel()}},
\code{\link[=espn_college_baseball_game_player_box]{espn_college_baseball_game_player_box()}},
\code{\link[=espn_college_baseball_game_rosters]{espn_college_baseball_game_rosters()}},
\code{\link[=espn_college_baseball_game_situation]{espn_college_baseball_game_situation()}},
\code{\link[=espn_college_baseball_game_team_leaders]{espn_college_baseball_game_team_leaders()}},
\code{\link[=espn_college_baseball_game_team_linescores]{espn_college_baseball_game_team_linescores()}},
\code{\link[=espn_college_baseball_game_team_records]{espn_college_baseball_game_team_records()}},
\code{\link[=espn_college_baseball_game_team_roster]{espn_college_baseball_game_team_roster()}},
\code{\link[=espn_college_baseball_game_team_roster_entry]{espn_college_baseball_game_team_roster_entry()}},
\code{\link[=espn_college_baseball_game_team_score]{espn_college_baseball_game_team_score()}},
\code{\link[=espn_college_baseball_game_team_statistics]{espn_college_baseball_game_team_statistics()}},
\code{\link[=espn_college_baseball_leaders]{espn_college_baseball_leaders()}},
\code{\link[=espn_college_baseball_news]{espn_college_baseball_news()}},
\code{\link[=espn_college_baseball_pbp]{espn_college_baseball_pbp()}},
\code{\link[=espn_college_baseball_player_awards]{espn_college_baseball_player_awards()}},
\code{\link[=espn_college_baseball_player_box]{espn_college_baseball_player_box()}},
\code{\link[=espn_college_baseball_player_career_stats]{espn_college_baseball_player_career_stats()}},
\code{\link[=espn_college_baseball_player_eventlog]{espn_college_baseball_player_eventlog()}},
\code{\link[=espn_college_baseball_player_gamelog]{espn_college_baseball_player_gamelog()}},
\code{\link[=espn_college_baseball_player_info]{espn_college_baseball_player_info()}},
\code{\link[=espn_college_baseball_player_overview]{espn_college_baseball_player_overview()}},
\code{\link[=espn_college_baseball_player_seasons]{espn_college_baseball_player_seasons()}},
\code{\link[=espn_college_baseball_player_splits]{espn_college_baseball_player_splits()}},
\code{\link[=espn_college_baseball_player_statisticslog]{espn_college_baseball_player_statisticslog()}},
\code{\link[=espn_college_baseball_scoreboard]{espn_college_baseball_scoreboard()}},
\code{\link[=espn_college_baseball_season_group]{espn_college_baseball_season_group()}},
\code{\link[=espn_college_baseball_season_group_children]{espn_college_baseball_season_group_children()}},
\code{\link[=espn_college_baseball_season_group_teams]{espn_college_baseball_season_group_teams()}},
\code{\link[=espn_college_baseball_season_groups]{espn_college_baseball_season_groups()}},
\code{\link[=espn_college_baseball_season_info]{espn_college_baseball_season_info()}},
\code{\link[=espn_college_baseball_season_leaders]{espn_college_baseball_season_leaders()}},
\code{\link[=espn_college_baseball_season_ranking]{espn_college_baseball_season_ranking()}},
\code{\link[=espn_college_baseball_season_rankings]{espn_college_baseball_season_rankings()}},
\code{\link[=espn_college_baseball_season_type]{espn_college_baseball_season_type()}},
\code{\link[=espn_college_baseball_season_types]{espn_college_baseball_season_types()}},
\code{\link[=espn_college_baseball_season_week]{espn_college_baseball_season_week()}},
\code{\link[=espn_college_baseball_season_weeks]{espn_college_baseball_season_weeks()}},
\code{\link[=espn_college_baseball_seasons]{espn_college_baseball_seasons()}},
\code{\link[=espn_college_baseball_standings]{espn_college_baseball_standings()}},
\code{\link[=espn_college_baseball_team]{espn_college_baseball_team()}},
\code{\link[=espn_college_baseball_team_box]{espn_college_baseball_team_box()}},
\code{\link[=espn_college_baseball_team_leaders]{espn_college_baseball_team_leaders()}},
\code{\link[=espn_college_baseball_team_news]{espn_college_baseball_team_news()}},
\code{\link[=espn_college_baseball_team_record]{espn_college_baseball_team_record()}},
\code{\link[=espn_college_baseball_team_record_detail]{espn_college_baseball_team_record_detail()}},
\code{\link[=espn_college_baseball_team_roster]{espn_college_baseball_team_roster()}},
\code{\link[=espn_college_baseball_team_schedule]{espn_college_baseball_team_schedule()}},
\code{\link[=espn_college_baseball_team_season_profile]{espn_college_baseball_team_season_profile()}},
\code{\link[=espn_college_baseball_team_season_roster]{espn_college_baseball_team_season_roster()}},
\code{\link[=espn_college_baseball_team_season_statistics]{espn_college_baseball_team_season_statistics()}},
\code{\link[=espn_college_baseball_teams]{espn_college_baseball_teams()}},
\code{\link[=espn_college_baseball_tournament]{espn_college_baseball_tournament()}},
\code{\link[=espn_college_baseball_tournament_season]{espn_college_baseball_tournament_season()}},
\code{\link[=espn_college_baseball_tournament_seasons]{espn_college_baseball_tournament_seasons()}},
\code{\link[=espn_college_baseball_tournaments]{espn_college_baseball_tournaments()}},
\code{\link[=espn_college_baseball_venues]{espn_college_baseball_venues()}},
\code{\link[=espn_college_baseball_week_ranking]{espn_college_baseball_week_ranking()}},
\code{\link[=espn_college_baseball_week_rankings]{espn_college_baseball_week_rankings()}}
}
\author{
Saiem Gilani
}
\concept{ESPN College Baseball Functions}
\keyword{Baseball}
\keyword{College}
\keyword{Conferences}
