% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_college_baseball_data.R
\name{espn_college_baseball_scoreboard}
\alias{espn_college_baseball_scoreboard}
\title{\strong{Get ESPN College Baseball Scoreboard}}
\usage{
espn_college_baseball_scoreboard(season)
}
\arguments{
\item{season}{Either numeric or character (YYYYMMDD)}
}
\value{
Returns a tibble with scoreboard data\tabular{lll}{
   col_name \tab types \tab description \cr
   matchup \tab character \tab Full game matchup name. \cr
   matchup_short \tab character \tab Short matchup name. \cr
   season \tab integer \tab Season (4-digit year). \cr
   season_type \tab integer \tab ESPN season type (1=pre, 2=regular, 3=postseason, 4=off-season). \cr
   season_slug \tab character \tab Season slug. \cr
   game_id \tab integer \tab Unique ESPN game/event identifier. \cr
   game_uid \tab character \tab ESPN game UID. \cr
   game_date \tab Date \tab Game date (YYYY-MM-DD). \cr
   attendance \tab integer \tab Game attendance. \cr
   play_by_play_available \tab logical \tab TRUE if play-by-play is available. \cr
   was_suspended \tab logical \tab Was suspended. \cr
   notes \tab logical \tab Notes. \cr
   status_name \tab character \tab Game status (e.g. 'STATUS_FINAL'). \cr
   broadcast_market \tab character \tab Broadcast market ('national'/'home'/'away'). \cr
   broadcast_name \tab character \tab Broadcast network name. \cr
   batting_leader_value \tab numeric \tab Team batting: batting leader value. \cr
   batting_leader_stat \tab character \tab Team batting: batting leader stat. \cr
   batting_leader_name \tab character \tab Team batting: batting leader name. \cr
   batting_leader_shortname \tab character \tab Team batting: batting leader shortname. \cr
   batting_leader_headshot \tab character \tab Team batting: batting leader headshot. \cr
   batting_leader_team_id \tab character \tab Team batting: batting leader team id. \cr
   batting_leader_pos \tab character \tab Team batting: batting leader pos. \cr
   home_run_leader_value \tab logical \tab Home run leader value. \cr
   home_run_leader_stat \tab logical \tab Home run leader stat. \cr
   home_run_leader_name \tab logical \tab Home run leader name. \cr
   home_run_leader_shortname \tab logical \tab Home run leader shortname. \cr
   home_run_leader_headshot \tab logical \tab Home run leader headshot. \cr
   home_run_leader_team_id \tab logical \tab Home run leader team id. \cr
   home_run_leader_pos \tab logical \tab Home run leader pos. \cr
   rbi_leader_value \tab logical \tab Rbi leader value. \cr
   rbi_leader_stat \tab logical \tab Rbi leader stat. \cr
   rbi_leader_name \tab logical \tab Rbi leader name. \cr
   rbi_leader_shortname \tab logical \tab Rbi leader shortname. \cr
   rbi_leader_headshot \tab logical \tab Rbi leader headshot. \cr
   rbi_leader_team_id \tab logical \tab Rbi leader team id. \cr
   rbi_leader_pos \tab logical \tab Rbi leader pos. \cr
   start_date \tab character \tab Start date. \cr
   broadcast \tab character \tab Broadcast network. \cr
   highlights \tab logical \tab Highlights. \cr
   game_date_time \tab POSIXct \tab Game start date/time (US/Eastern). \cr
   home_team_name \tab character \tab Home team name. \cr
   home_team_logo \tab character \tab Home team logo. \cr
   home_team_abb \tab character \tab Home team abb. \cr
   home_team_id \tab integer \tab Home team id. \cr
   home_team_location \tab character \tab Home team location. \cr
   home_team_full_name \tab character \tab Home team full name. \cr
   home_team_color \tab character \tab Home team color. \cr
   home_score \tab integer \tab Home team run total after the play. \cr
   home_win \tab integer \tab Home win. \cr
   home_record \tab character \tab Home record. \cr
   away_team_name \tab character \tab Away team name. \cr
   away_team_logo \tab character \tab Away team logo. \cr
   away_team_abb \tab character \tab Away team abb. \cr
   away_team_id \tab integer \tab Away team id. \cr
   away_team_location \tab character \tab Away team location. \cr
   away_team_full_name \tab character \tab Away team full name. \cr
   away_team_color \tab character \tab Away team color. \cr
   away_score \tab integer \tab Away team run total after the play. \cr
   away_win \tab integer \tab Away win. \cr
   away_record \tab character \tab Away record. \cr
}
}
\description{
\strong{Get ESPN College Baseball schedule for a specific year}
}
\examples{

# Get scoreboard from a College World Series date (2025-06-15).
\donttest{
try(espn_college_baseball_scoreboard(season = "20250615"))
}
}
\seealso{
Other ESPN College Baseball Functions:
\code{\link[=espn_college_baseball_athletes_index]{espn_college_baseball_athletes_index()}},
\code{\link[=espn_college_baseball_calendar]{espn_college_baseball_calendar()}},
\code{\link[=espn_college_baseball_coach]{espn_college_baseball_coach()}},
\code{\link[=espn_college_baseball_coach_record]{espn_college_baseball_coach_record()}},
\code{\link[=espn_college_baseball_coach_season]{espn_college_baseball_coach_season()}},
\code{\link[=espn_college_baseball_coaches]{espn_college_baseball_coaches()}},
\code{\link[=espn_college_baseball_conferences]{espn_college_baseball_conferences()}},
\code{\link[=espn_college_baseball_game_all]{espn_college_baseball_game_all()}},
\code{\link[=espn_college_baseball_game_broadcasts]{espn_college_baseball_game_broadcasts()}},
\code{\link[=espn_college_baseball_game_official_detail]{espn_college_baseball_game_official_detail()}},
\code{\link[=espn_college_baseball_game_officials]{espn_college_baseball_game_officials()}},
\code{\link[=espn_college_baseball_game_play]{espn_college_baseball_game_play()}},
\code{\link[=espn_college_baseball_game_play_personnel]{espn_college_baseball_game_play_personnel()}},
\code{\link[=espn_college_baseball_game_player_box]{espn_college_baseball_game_player_box()}},
\code{\link[=espn_college_baseball_game_rosters]{espn_college_baseball_game_rosters()}},
\code{\link[=espn_college_baseball_game_situation]{espn_college_baseball_game_situation()}},
\code{\link[=espn_college_baseball_game_team_leaders]{espn_college_baseball_game_team_leaders()}},
\code{\link[=espn_college_baseball_game_team_linescores]{espn_college_baseball_game_team_linescores()}},
\code{\link[=espn_college_baseball_game_team_records]{espn_college_baseball_game_team_records()}},
\code{\link[=espn_college_baseball_game_team_roster]{espn_college_baseball_game_team_roster()}},
\code{\link[=espn_college_baseball_game_team_roster_entry]{espn_college_baseball_game_team_roster_entry()}},
\code{\link[=espn_college_baseball_game_team_score]{espn_college_baseball_game_team_score()}},
\code{\link[=espn_college_baseball_game_team_statistics]{espn_college_baseball_game_team_statistics()}},
\code{\link[=espn_college_baseball_leaders]{espn_college_baseball_leaders()}},
\code{\link[=espn_college_baseball_news]{espn_college_baseball_news()}},
\code{\link[=espn_college_baseball_pbp]{espn_college_baseball_pbp()}},
\code{\link[=espn_college_baseball_player_awards]{espn_college_baseball_player_awards()}},
\code{\link[=espn_college_baseball_player_box]{espn_college_baseball_player_box()}},
\code{\link[=espn_college_baseball_player_career_stats]{espn_college_baseball_player_career_stats()}},
\code{\link[=espn_college_baseball_player_eventlog]{espn_college_baseball_player_eventlog()}},
\code{\link[=espn_college_baseball_player_gamelog]{espn_college_baseball_player_gamelog()}},
\code{\link[=espn_college_baseball_player_info]{espn_college_baseball_player_info()}},
\code{\link[=espn_college_baseball_player_overview]{espn_college_baseball_player_overview()}},
\code{\link[=espn_college_baseball_player_seasons]{espn_college_baseball_player_seasons()}},
\code{\link[=espn_college_baseball_player_splits]{espn_college_baseball_player_splits()}},
\code{\link[=espn_college_baseball_player_statisticslog]{espn_college_baseball_player_statisticslog()}},
\code{\link[=espn_college_baseball_season_group]{espn_college_baseball_season_group()}},
\code{\link[=espn_college_baseball_season_group_children]{espn_college_baseball_season_group_children()}},
\code{\link[=espn_college_baseball_season_group_teams]{espn_college_baseball_season_group_teams()}},
\code{\link[=espn_college_baseball_season_groups]{espn_college_baseball_season_groups()}},
\code{\link[=espn_college_baseball_season_info]{espn_college_baseball_season_info()}},
\code{\link[=espn_college_baseball_season_leaders]{espn_college_baseball_season_leaders()}},
\code{\link[=espn_college_baseball_season_ranking]{espn_college_baseball_season_ranking()}},
\code{\link[=espn_college_baseball_season_rankings]{espn_college_baseball_season_rankings()}},
\code{\link[=espn_college_baseball_season_type]{espn_college_baseball_season_type()}},
\code{\link[=espn_college_baseball_season_types]{espn_college_baseball_season_types()}},
\code{\link[=espn_college_baseball_season_week]{espn_college_baseball_season_week()}},
\code{\link[=espn_college_baseball_season_weeks]{espn_college_baseball_season_weeks()}},
\code{\link[=espn_college_baseball_seasons]{espn_college_baseball_seasons()}},
\code{\link[=espn_college_baseball_standings]{espn_college_baseball_standings()}},
\code{\link[=espn_college_baseball_team]{espn_college_baseball_team()}},
\code{\link[=espn_college_baseball_team_box]{espn_college_baseball_team_box()}},
\code{\link[=espn_college_baseball_team_leaders]{espn_college_baseball_team_leaders()}},
\code{\link[=espn_college_baseball_team_news]{espn_college_baseball_team_news()}},
\code{\link[=espn_college_baseball_team_record]{espn_college_baseball_team_record()}},
\code{\link[=espn_college_baseball_team_record_detail]{espn_college_baseball_team_record_detail()}},
\code{\link[=espn_college_baseball_team_roster]{espn_college_baseball_team_roster()}},
\code{\link[=espn_college_baseball_team_schedule]{espn_college_baseball_team_schedule()}},
\code{\link[=espn_college_baseball_team_season_profile]{espn_college_baseball_team_season_profile()}},
\code{\link[=espn_college_baseball_team_season_roster]{espn_college_baseball_team_season_roster()}},
\code{\link[=espn_college_baseball_team_season_statistics]{espn_college_baseball_team_season_statistics()}},
\code{\link[=espn_college_baseball_teams]{espn_college_baseball_teams()}},
\code{\link[=espn_college_baseball_tournament]{espn_college_baseball_tournament()}},
\code{\link[=espn_college_baseball_tournament_season]{espn_college_baseball_tournament_season()}},
\code{\link[=espn_college_baseball_tournament_seasons]{espn_college_baseball_tournament_seasons()}},
\code{\link[=espn_college_baseball_tournaments]{espn_college_baseball_tournaments()}},
\code{\link[=espn_college_baseball_venues]{espn_college_baseball_venues()}},
\code{\link[=espn_college_baseball_week_ranking]{espn_college_baseball_week_ranking()}},
\code{\link[=espn_college_baseball_week_rankings]{espn_college_baseball_week_rankings()}}
}
\author{
Saiem Gilani
}
\concept{ESPN College Baseball Functions}
