% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mlb_data.R
\name{espn_mlb_pbp}
\alias{espn_mlb_pbp}
\title{\strong{Get ESPN MLB PBP data}}
\usage{
espn_mlb_pbp(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
A play-by-play data frame.\tabular{lll}{
   col_name \tab types \tab description \cr
   id \tab character \tab Play / record identifier. \cr
   sequence_number \tab character \tab Play sequence number within the game. \cr
   text \tab character \tab Text description of the play. \cr
   away_score \tab integer \tab Away team run total after the play. \cr
   home_score \tab integer \tab Home team run total after the play. \cr
   scoring_play \tab logical \tab TRUE if the play scored a run. \cr
   score_value \tab integer \tab Runs scored on the play. \cr
   wallclock \tab character \tab Wall-clock timestamp of the play (ISO 8601). \cr
   at_bat_id \tab character \tab Identifier of the at-bat the play belongs to. \cr
   summary_type \tab character \tab Play summary type. \cr
   outs \tab integer \tab Outs in the inning after the play. \cr
   type \tab character \tab Type. \cr
   bat_order \tab integer \tab Spot in the batting order (1-9; NA if not applicable). \cr
   at_bat_pitch_number \tab integer \tab Pitch number within the at-bat. \cr
   pitch_velocity \tab integer \tab Pitch velocity (mph). \cr
   trajectory \tab character \tab Batted-ball trajectory. \cr
   alternative_play \tab character \tab Alternative play. \cr
   type_id \tab integer \tab Type id. \cr
   type_text \tab character \tab Type text. \cr
   type_type \tab character \tab Type type. \cr
   type_alternative_text \tab character \tab Type alternative text. \cr
   type_abbreviation \tab character \tab Type abbreviation. \cr
   period_type \tab character \tab Period type ('inning'). \cr
   period_number \tab integer \tab Inning number. \cr
   period_display_value \tab character \tab Inning display value (e.g. 'Top 1st'). \cr
   team_id \tab integer \tab Unique ESPN team identifier. \cr
   pitch_count_balls \tab integer \tab Balls in the count when the pitch was thrown. \cr
   pitch_count_strikes \tab integer \tab Strikes in the count when the pitch was thrown. \cr
   result_count_balls \tab integer \tab Balls in the count after the pitch. \cr
   result_count_strikes \tab integer \tab Strikes in the count after the pitch. \cr
   bats_type \tab character \tab Bats type. \cr
   bats_abbreviation \tab character \tab Bats abbreviation. \cr
   bats_display_value \tab character \tab Bats display value. \cr
   pitch_coordinate_x \tab integer \tab Pitch location x-coordinate. \cr
   pitch_coordinate_y \tab integer \tab Pitch location y-coordinate. \cr
   pitch_type_id \tab character \tab Pitch type identifier. \cr
   pitch_type_text \tab character \tab Pitch type description (e.g. 'Four-seam FB'). \cr
   pitch_type_abbreviation \tab character \tab Pitch type abbreviation. \cr
   hit_coordinate_x \tab integer \tab Batted-ball location x-coordinate. \cr
   hit_coordinate_y \tab integer \tab Batted-ball location y-coordinate. \cr
   alternative_type_id \tab character \tab Alternative type id. \cr
   alternative_type_text \tab character \tab Alternative type text. \cr
   alternative_type_abbreviation \tab character \tab Alternative type abbreviation. \cr
   alternative_type_alternative_text \tab character \tab Alternative type alternative text. \cr
   alternative_type_type \tab character \tab Alternative type type. \cr
   on_first_athlete_id \tab character \tab Athlete id of the runner on first base. \cr
   on_second_athlete_id \tab character \tab Athlete id of the runner on second base. \cr
   play_id \tab character \tab Play id. \cr
   athlete_id_1 \tab integer \tab Primary participating athlete id (batter). \cr
   athlete_id_2 \tab integer \tab Second participating athlete id (pitcher). \cr
   athlete_id_3 \tab integer \tab Third participating athlete id (fielder). \cr
   home_team_id \tab integer \tab Home team id. \cr
   home_team_mascot \tab character \tab Home team mascot. \cr
   home_team_name \tab character \tab Home team name. \cr
   home_team_abbrev \tab character \tab Home team abbrev. \cr
   home_team_logo \tab character \tab Home team logo. \cr
   home_team_logo_dark \tab character \tab Home team logo dark. \cr
   home_team_full_name \tab character \tab Home team full name. \cr
   home_team_color \tab character \tab Home team color. \cr
   home_team_alternate_color \tab character \tab Home team alternate color. \cr
   home_team_score \tab integer \tab Home team score. \cr
   home_team_winner \tab logical \tab Home team winner. \cr
   home_team_record \tab character \tab Home team record. \cr
   away_team_id \tab integer \tab Away team id. \cr
   away_team_mascot \tab character \tab Away team mascot. \cr
   away_team_name \tab character \tab Away team name. \cr
   away_team_abbrev \tab character \tab Away team abbrev. \cr
   away_team_logo \tab character \tab Away team logo. \cr
   away_team_logo_dark \tab character \tab Away team logo dark. \cr
   away_team_full_name \tab character \tab Away team full name. \cr
   away_team_color \tab character \tab Away team color. \cr
   away_team_alternate_color \tab character \tab Away team alternate color. \cr
   away_team_score \tab integer \tab Away team score. \cr
   away_team_winner \tab logical \tab Away team winner. \cr
   away_team_record \tab character \tab Away team record. \cr
   game_id \tab integer \tab Unique ESPN game/event identifier. \cr
   season \tab integer \tab Season (4-digit year). \cr
   season_type \tab integer \tab ESPN season type (1=pre, 2=regular, 3=postseason, 4=off-season). \cr
   game_date \tab Date \tab Game date (YYYY-MM-DD). \cr
   game_date_time \tab POSIXct \tab Game start date/time (US/Eastern). \cr
}
}
\description{
\strong{Get ESPN MLB PBP data}
}
\examples{
\donttest{
try(espn_mlb_pbp(game_id = 401071880))
}

}
\seealso{
Other ESPN MLB Functions:
\code{\link{espn_mlb}},
\code{\link[=espn_mlb_athletes_index]{espn_mlb_athletes_index()}},
\code{\link[=espn_mlb_award]{espn_mlb_award()}},
\code{\link[=espn_mlb_betting]{espn_mlb_betting()}},
\code{\link[=espn_mlb_calendar]{espn_mlb_calendar()}},
\code{\link[=espn_mlb_coach]{espn_mlb_coach()}},
\code{\link[=espn_mlb_coach_record]{espn_mlb_coach_record()}},
\code{\link[=espn_mlb_coach_season]{espn_mlb_coach_season()}},
\code{\link[=espn_mlb_coaches]{espn_mlb_coaches()}},
\code{\link[=espn_mlb_conferences]{espn_mlb_conferences()}},
\code{\link[=espn_mlb_draft]{espn_mlb_draft()}},
\code{\link[=espn_mlb_draft_athlete_detail]{espn_mlb_draft_athlete_detail()}},
\code{\link[=espn_mlb_draft_athletes]{espn_mlb_draft_athletes()}},
\code{\link[=espn_mlb_draft_pick]{espn_mlb_draft_pick()}},
\code{\link[=espn_mlb_draft_rounds]{espn_mlb_draft_rounds()}},
\code{\link[=espn_mlb_draft_status]{espn_mlb_draft_status()}},
\code{\link[=espn_mlb_franchise]{espn_mlb_franchise()}},
\code{\link[=espn_mlb_franchises]{espn_mlb_franchises()}},
\code{\link[=espn_mlb_freeagents]{espn_mlb_freeagents()}},
\code{\link[=espn_mlb_futures]{espn_mlb_futures()}},
\code{\link[=espn_mlb_game_all]{espn_mlb_game_all()}},
\code{\link[=espn_mlb_game_broadcasts]{espn_mlb_game_broadcasts()}},
\code{\link{espn_mlb_game_endpoints}},
\code{\link[=espn_mlb_game_info]{espn_mlb_game_info()}},
\code{\link[=espn_mlb_game_odds]{espn_mlb_game_odds()}},
\code{\link[=espn_mlb_game_official_detail]{espn_mlb_game_official_detail()}},
\code{\link[=espn_mlb_game_officials]{espn_mlb_game_officials()}},
\code{\link[=espn_mlb_game_play]{espn_mlb_game_play()}},
\code{\link[=espn_mlb_game_play_personnel]{espn_mlb_game_play_personnel()}},
\code{\link[=espn_mlb_game_player_box]{espn_mlb_game_player_box()}},
\code{\link[=espn_mlb_game_powerindex]{espn_mlb_game_powerindex()}},
\code{\link[=espn_mlb_game_predictor]{espn_mlb_game_predictor()}},
\code{\link[=espn_mlb_game_probabilities]{espn_mlb_game_probabilities()}},
\code{\link[=espn_mlb_game_probables]{espn_mlb_game_probables()}},
\code{\link[=espn_mlb_game_propbets]{espn_mlb_game_propbets()}},
\code{\link[=espn_mlb_game_rosters]{espn_mlb_game_rosters()}},
\code{\link[=espn_mlb_game_situation]{espn_mlb_game_situation()}},
\code{\link[=espn_mlb_game_team_leaders]{espn_mlb_game_team_leaders()}},
\code{\link[=espn_mlb_game_team_linescores]{espn_mlb_game_team_linescores()}},
\code{\link[=espn_mlb_game_team_records]{espn_mlb_game_team_records()}},
\code{\link[=espn_mlb_game_team_roster]{espn_mlb_game_team_roster()}},
\code{\link[=espn_mlb_game_team_roster_entry]{espn_mlb_game_team_roster_entry()}},
\code{\link[=espn_mlb_game_team_score]{espn_mlb_game_team_score()}},
\code{\link[=espn_mlb_game_team_statistics]{espn_mlb_game_team_statistics()}},
\code{\link[=espn_mlb_injuries]{espn_mlb_injuries()}},
\code{\link[=espn_mlb_leaders]{espn_mlb_leaders()}},
\code{\link[=espn_mlb_news]{espn_mlb_news()}},
\code{\link[=espn_mlb_player_awards]{espn_mlb_player_awards()}},
\code{\link[=espn_mlb_player_box]{espn_mlb_player_box()}},
\code{\link[=espn_mlb_player_career_stats]{espn_mlb_player_career_stats()}},
\code{\link[=espn_mlb_player_contract]{espn_mlb_player_contract()}},
\code{\link[=espn_mlb_player_contracts]{espn_mlb_player_contracts()}},
\code{\link{espn_mlb_player_endpoints}},
\code{\link[=espn_mlb_player_eventlog]{espn_mlb_player_eventlog()}},
\code{\link[=espn_mlb_player_eventlog_v2]{espn_mlb_player_eventlog_v2()}},
\code{\link[=espn_mlb_player_gamelog]{espn_mlb_player_gamelog()}},
\code{\link[=espn_mlb_player_info]{espn_mlb_player_info()}},
\code{\link[=espn_mlb_player_overview]{espn_mlb_player_overview()}},
\code{\link[=espn_mlb_player_seasons]{espn_mlb_player_seasons()}},
\code{\link[=espn_mlb_player_splits]{espn_mlb_player_splits()}},
\code{\link[=espn_mlb_player_statisticslog]{espn_mlb_player_statisticslog()}},
\code{\link[=espn_mlb_player_stats]{espn_mlb_player_stats()}},
\code{\link[=espn_mlb_player_stats_v3]{espn_mlb_player_stats_v3()}},
\code{\link[=espn_mlb_position]{espn_mlb_position()}},
\code{\link[=espn_mlb_positions]{espn_mlb_positions()}},
\code{\link[=espn_mlb_powerindex]{espn_mlb_powerindex()}},
\code{\link[=espn_mlb_scoreboard]{espn_mlb_scoreboard()}},
\code{\link[=espn_mlb_season_awards]{espn_mlb_season_awards()}},
\code{\link[=espn_mlb_season_draft]{espn_mlb_season_draft()}},
\code{\link[=espn_mlb_season_group]{espn_mlb_season_group()}},
\code{\link[=espn_mlb_season_group_children]{espn_mlb_season_group_children()}},
\code{\link[=espn_mlb_season_group_teams]{espn_mlb_season_group_teams()}},
\code{\link[=espn_mlb_season_groups]{espn_mlb_season_groups()}},
\code{\link[=espn_mlb_season_info]{espn_mlb_season_info()}},
\code{\link[=espn_mlb_season_leaders]{espn_mlb_season_leaders()}},
\code{\link[=espn_mlb_season_ranking]{espn_mlb_season_ranking()}},
\code{\link[=espn_mlb_season_rankings]{espn_mlb_season_rankings()}},
\code{\link[=espn_mlb_season_type]{espn_mlb_season_type()}},
\code{\link[=espn_mlb_season_types]{espn_mlb_season_types()}},
\code{\link[=espn_mlb_season_week]{espn_mlb_season_week()}},
\code{\link[=espn_mlb_season_weeks]{espn_mlb_season_weeks()}},
\code{\link[=espn_mlb_seasons]{espn_mlb_seasons()}},
\code{\link[=espn_mlb_standings]{espn_mlb_standings()}},
\code{\link[=espn_mlb_team]{espn_mlb_team()}},
\code{\link[=espn_mlb_team_box]{espn_mlb_team_box()}},
\code{\link[=espn_mlb_team_current_roster]{espn_mlb_team_current_roster()}},
\code{\link[=espn_mlb_team_depthchart]{espn_mlb_team_depthchart()}},
\code{\link{espn_mlb_team_endpoints}},
\code{\link[=espn_mlb_team_injuries]{espn_mlb_team_injuries()}},
\code{\link[=espn_mlb_team_leaders]{espn_mlb_team_leaders()}},
\code{\link[=espn_mlb_team_news]{espn_mlb_team_news()}},
\code{\link[=espn_mlb_team_odds_records]{espn_mlb_team_odds_records()}},
\code{\link[=espn_mlb_team_record]{espn_mlb_team_record()}},
\code{\link[=espn_mlb_team_record_detail]{espn_mlb_team_record_detail()}},
\code{\link[=espn_mlb_team_roster]{espn_mlb_team_roster()}},
\code{\link[=espn_mlb_team_schedule]{espn_mlb_team_schedule()}},
\code{\link[=espn_mlb_team_season_profile]{espn_mlb_team_season_profile()}},
\code{\link[=espn_mlb_team_season_roster]{espn_mlb_team_season_roster()}},
\code{\link[=espn_mlb_team_season_statistics]{espn_mlb_team_season_statistics()}},
\code{\link[=espn_mlb_team_stats]{espn_mlb_team_stats()}},
\code{\link[=espn_mlb_teams]{espn_mlb_teams()}},
\code{\link[=espn_mlb_tournament]{espn_mlb_tournament()}},
\code{\link[=espn_mlb_tournament_season]{espn_mlb_tournament_season()}},
\code{\link[=espn_mlb_tournament_seasons]{espn_mlb_tournament_seasons()}},
\code{\link[=espn_mlb_tournaments]{espn_mlb_tournaments()}},
\code{\link[=espn_mlb_transactions]{espn_mlb_transactions()}},
\code{\link[=espn_mlb_venues]{espn_mlb_venues()}},
\code{\link[=espn_mlb_week_ranking]{espn_mlb_week_ranking()}},
\code{\link[=espn_mlb_week_rankings]{espn_mlb_week_rankings()}},
\code{\link[=espn_mlb_wp]{espn_mlb_wp()}}
}
\author{
Saiem Gilani
}
\concept{ESPN MLB Functions}
\keyword{MLB}
\keyword{PBP}
