% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mlb_overviews.R
\name{espn_mlb_player_endpoints}
\alias{espn_mlb_player_endpoints}
\alias{espn_mlb_athlete}
\alias{espn_mlb_athletes_overview}
\title{\strong{ESPN MLB Player / Athlete Endpoint Overview}}
\description{
\itemize{
\item \code{espn_mlb_athletes_index()}: Paginated index of MLB athletes.
\item \code{espn_mlb_player_info()}: Athlete bio / identity.
\item \code{espn_mlb_player_overview()}: Athlete overview (stats, next game, news).
\item \code{espn_mlb_player_gamelog()}: Game-by-game log for a season.
\item \code{espn_mlb_player_splits()}: Statistical splits.
\item \code{espn_mlb_player_stats()}, \code{espn_mlb_player_stats_v3()}: Season statistics.
\item \code{espn_mlb_player_career_stats()}, \code{espn_mlb_player_seasons()}: Career / season history.
\item \code{espn_mlb_player_eventlog()}, \code{espn_mlb_player_eventlog_v2()}, \code{espn_mlb_player_statisticslog()}: Event / statistics logs.
\item \code{espn_mlb_player_contract()}, \code{espn_mlb_player_contracts()}: Contract data.
\item \code{espn_mlb_player_awards()}: Athlete awards.
\item \code{espn_mlb_draft_athletes()}, \code{espn_mlb_draft_athlete_detail()}: Draft prospects.
}
}
\details{
\subsection{\strong{ESPN MLB Players}}{

Athlete-level ESPN endpoints across \code{sports.core.api.espn.com} and
\code{site.web.api.espn.com}: identity, overviews, game logs, splits, season /
career statistics, contracts, and awards.
}
}
\seealso{
Other ESPN MLB Functions:
\code{\link{espn_mlb}},
\code{\link[=espn_mlb_athletes_index]{espn_mlb_athletes_index()}},
\code{\link[=espn_mlb_award]{espn_mlb_award()}},
\code{\link[=espn_mlb_betting]{espn_mlb_betting()}},
\code{\link[=espn_mlb_calendar]{espn_mlb_calendar()}},
\code{\link[=espn_mlb_coach]{espn_mlb_coach()}},
\code{\link[=espn_mlb_coach_record]{espn_mlb_coach_record()}},
\code{\link[=espn_mlb_coach_season]{espn_mlb_coach_season()}},
\code{\link[=espn_mlb_coaches]{espn_mlb_coaches()}},
\code{\link[=espn_mlb_conferences]{espn_mlb_conferences()}},
\code{\link[=espn_mlb_draft]{espn_mlb_draft()}},
\code{\link[=espn_mlb_draft_athlete_detail]{espn_mlb_draft_athlete_detail()}},
\code{\link[=espn_mlb_draft_athletes]{espn_mlb_draft_athletes()}},
\code{\link[=espn_mlb_draft_pick]{espn_mlb_draft_pick()}},
\code{\link[=espn_mlb_draft_rounds]{espn_mlb_draft_rounds()}},
\code{\link[=espn_mlb_draft_status]{espn_mlb_draft_status()}},
\code{\link[=espn_mlb_franchise]{espn_mlb_franchise()}},
\code{\link[=espn_mlb_franchises]{espn_mlb_franchises()}},
\code{\link[=espn_mlb_freeagents]{espn_mlb_freeagents()}},
\code{\link[=espn_mlb_futures]{espn_mlb_futures()}},
\code{\link[=espn_mlb_game_all]{espn_mlb_game_all()}},
\code{\link[=espn_mlb_game_broadcasts]{espn_mlb_game_broadcasts()}},
\code{\link{espn_mlb_game_endpoints}},
\code{\link[=espn_mlb_game_info]{espn_mlb_game_info()}},
\code{\link[=espn_mlb_game_odds]{espn_mlb_game_odds()}},
\code{\link[=espn_mlb_game_official_detail]{espn_mlb_game_official_detail()}},
\code{\link[=espn_mlb_game_officials]{espn_mlb_game_officials()}},
\code{\link[=espn_mlb_game_play]{espn_mlb_game_play()}},
\code{\link[=espn_mlb_game_play_personnel]{espn_mlb_game_play_personnel()}},
\code{\link[=espn_mlb_game_player_box]{espn_mlb_game_player_box()}},
\code{\link[=espn_mlb_game_powerindex]{espn_mlb_game_powerindex()}},
\code{\link[=espn_mlb_game_predictor]{espn_mlb_game_predictor()}},
\code{\link[=espn_mlb_game_probabilities]{espn_mlb_game_probabilities()}},
\code{\link[=espn_mlb_game_probables]{espn_mlb_game_probables()}},
\code{\link[=espn_mlb_game_propbets]{espn_mlb_game_propbets()}},
\code{\link[=espn_mlb_game_rosters]{espn_mlb_game_rosters()}},
\code{\link[=espn_mlb_game_situation]{espn_mlb_game_situation()}},
\code{\link[=espn_mlb_game_team_leaders]{espn_mlb_game_team_leaders()}},
\code{\link[=espn_mlb_game_team_linescores]{espn_mlb_game_team_linescores()}},
\code{\link[=espn_mlb_game_team_records]{espn_mlb_game_team_records()}},
\code{\link[=espn_mlb_game_team_roster]{espn_mlb_game_team_roster()}},
\code{\link[=espn_mlb_game_team_roster_entry]{espn_mlb_game_team_roster_entry()}},
\code{\link[=espn_mlb_game_team_score]{espn_mlb_game_team_score()}},
\code{\link[=espn_mlb_game_team_statistics]{espn_mlb_game_team_statistics()}},
\code{\link[=espn_mlb_injuries]{espn_mlb_injuries()}},
\code{\link[=espn_mlb_leaders]{espn_mlb_leaders()}},
\code{\link[=espn_mlb_news]{espn_mlb_news()}},
\code{\link[=espn_mlb_pbp]{espn_mlb_pbp()}},
\code{\link[=espn_mlb_player_awards]{espn_mlb_player_awards()}},
\code{\link[=espn_mlb_player_box]{espn_mlb_player_box()}},
\code{\link[=espn_mlb_player_career_stats]{espn_mlb_player_career_stats()}},
\code{\link[=espn_mlb_player_contract]{espn_mlb_player_contract()}},
\code{\link[=espn_mlb_player_contracts]{espn_mlb_player_contracts()}},
\code{\link[=espn_mlb_player_eventlog]{espn_mlb_player_eventlog()}},
\code{\link[=espn_mlb_player_eventlog_v2]{espn_mlb_player_eventlog_v2()}},
\code{\link[=espn_mlb_player_gamelog]{espn_mlb_player_gamelog()}},
\code{\link[=espn_mlb_player_info]{espn_mlb_player_info()}},
\code{\link[=espn_mlb_player_overview]{espn_mlb_player_overview()}},
\code{\link[=espn_mlb_player_seasons]{espn_mlb_player_seasons()}},
\code{\link[=espn_mlb_player_splits]{espn_mlb_player_splits()}},
\code{\link[=espn_mlb_player_statisticslog]{espn_mlb_player_statisticslog()}},
\code{\link[=espn_mlb_player_stats]{espn_mlb_player_stats()}},
\code{\link[=espn_mlb_player_stats_v3]{espn_mlb_player_stats_v3()}},
\code{\link[=espn_mlb_position]{espn_mlb_position()}},
\code{\link[=espn_mlb_positions]{espn_mlb_positions()}},
\code{\link[=espn_mlb_powerindex]{espn_mlb_powerindex()}},
\code{\link[=espn_mlb_scoreboard]{espn_mlb_scoreboard()}},
\code{\link[=espn_mlb_season_awards]{espn_mlb_season_awards()}},
\code{\link[=espn_mlb_season_draft]{espn_mlb_season_draft()}},
\code{\link[=espn_mlb_season_group]{espn_mlb_season_group()}},
\code{\link[=espn_mlb_season_group_children]{espn_mlb_season_group_children()}},
\code{\link[=espn_mlb_season_group_teams]{espn_mlb_season_group_teams()}},
\code{\link[=espn_mlb_season_groups]{espn_mlb_season_groups()}},
\code{\link[=espn_mlb_season_info]{espn_mlb_season_info()}},
\code{\link[=espn_mlb_season_leaders]{espn_mlb_season_leaders()}},
\code{\link[=espn_mlb_season_ranking]{espn_mlb_season_ranking()}},
\code{\link[=espn_mlb_season_rankings]{espn_mlb_season_rankings()}},
\code{\link[=espn_mlb_season_type]{espn_mlb_season_type()}},
\code{\link[=espn_mlb_season_types]{espn_mlb_season_types()}},
\code{\link[=espn_mlb_season_week]{espn_mlb_season_week()}},
\code{\link[=espn_mlb_season_weeks]{espn_mlb_season_weeks()}},
\code{\link[=espn_mlb_seasons]{espn_mlb_seasons()}},
\code{\link[=espn_mlb_standings]{espn_mlb_standings()}},
\code{\link[=espn_mlb_team]{espn_mlb_team()}},
\code{\link[=espn_mlb_team_box]{espn_mlb_team_box()}},
\code{\link[=espn_mlb_team_current_roster]{espn_mlb_team_current_roster()}},
\code{\link[=espn_mlb_team_depthchart]{espn_mlb_team_depthchart()}},
\code{\link{espn_mlb_team_endpoints}},
\code{\link[=espn_mlb_team_injuries]{espn_mlb_team_injuries()}},
\code{\link[=espn_mlb_team_leaders]{espn_mlb_team_leaders()}},
\code{\link[=espn_mlb_team_news]{espn_mlb_team_news()}},
\code{\link[=espn_mlb_team_odds_records]{espn_mlb_team_odds_records()}},
\code{\link[=espn_mlb_team_record]{espn_mlb_team_record()}},
\code{\link[=espn_mlb_team_record_detail]{espn_mlb_team_record_detail()}},
\code{\link[=espn_mlb_team_roster]{espn_mlb_team_roster()}},
\code{\link[=espn_mlb_team_schedule]{espn_mlb_team_schedule()}},
\code{\link[=espn_mlb_team_season_profile]{espn_mlb_team_season_profile()}},
\code{\link[=espn_mlb_team_season_roster]{espn_mlb_team_season_roster()}},
\code{\link[=espn_mlb_team_season_statistics]{espn_mlb_team_season_statistics()}},
\code{\link[=espn_mlb_team_stats]{espn_mlb_team_stats()}},
\code{\link[=espn_mlb_teams]{espn_mlb_teams()}},
\code{\link[=espn_mlb_tournament]{espn_mlb_tournament()}},
\code{\link[=espn_mlb_tournament_season]{espn_mlb_tournament_season()}},
\code{\link[=espn_mlb_tournament_seasons]{espn_mlb_tournament_seasons()}},
\code{\link[=espn_mlb_tournaments]{espn_mlb_tournaments()}},
\code{\link[=espn_mlb_transactions]{espn_mlb_transactions()}},
\code{\link[=espn_mlb_venues]{espn_mlb_venues()}},
\code{\link[=espn_mlb_week_ranking]{espn_mlb_week_ranking()}},
\code{\link[=espn_mlb_week_rankings]{espn_mlb_week_rankings()}},
\code{\link[=espn_mlb_wp]{espn_mlb_wp()}}
}
\concept{ESPN MLB Functions}
\keyword{ESPN}
