% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mlb_data.R
\name{espn_mlb_scoreboard}
\alias{espn_mlb_scoreboard}
\title{\strong{Get ESPN MLB schedule for a specific year}}
\usage{
espn_mlb_scoreboard(season)
}
\arguments{
\item{season}{Either numeric or character (YYYYMMDD)}
}
\value{
Returns a tibble with scoreboard data\tabular{lll}{
   col_name \tab types \tab description \cr
   matchup \tab character \tab Full game matchup name. \cr
   matchup_short \tab character \tab Short matchup name. \cr
   season \tab integer \tab Season (4-digit year). \cr
   season_type \tab integer \tab ESPN season type (1=pre, 2=regular, 3=postseason, 4=off-season). \cr
   season_slug \tab character \tab Season slug. \cr
   game_id \tab integer \tab Unique ESPN game/event identifier. \cr
   game_uid \tab character \tab ESPN game UID. \cr
   game_date \tab Date \tab Game date (YYYY-MM-DD). \cr
   attendance \tab integer \tab Game attendance. \cr
   play_by_play_available \tab logical \tab TRUE if play-by-play is available. \cr
   was_suspended \tab logical \tab Was suspended. \cr
   notes \tab logical \tab Notes. \cr
   status_name \tab character \tab Game status (e.g. 'STATUS_FINAL'). \cr
   broadcast_market \tab character \tab Broadcast market ('national'/'home'/'away'). \cr
   broadcast_name \tab character \tab Broadcast network name. \cr
   batting_leader_value \tab numeric \tab Team batting: batting leader value. \cr
   batting_leader_stat \tab character \tab Team batting: batting leader stat. \cr
   batting_leader_name \tab character \tab Team batting: batting leader name. \cr
   batting_leader_shortname \tab character \tab Team batting: batting leader shortname. \cr
   batting_leader_headshot \tab character \tab Team batting: batting leader headshot. \cr
   batting_leader_team_id \tab character \tab Team batting: batting leader team id. \cr
   batting_leader_pos \tab character \tab Team batting: batting leader pos. \cr
   home_run_leader_value \tab logical \tab Home run leader value. \cr
   home_run_leader_stat \tab logical \tab Home run leader stat. \cr
   home_run_leader_name \tab logical \tab Home run leader name. \cr
   home_run_leader_shortname \tab logical \tab Home run leader shortname. \cr
   home_run_leader_headshot \tab logical \tab Home run leader headshot. \cr
   home_run_leader_team_id \tab logical \tab Home run leader team id. \cr
   home_run_leader_pos \tab logical \tab Home run leader pos. \cr
   rbi_leader_value \tab logical \tab Rbi leader value. \cr
   rbi_leader_stat \tab logical \tab Rbi leader stat. \cr
   rbi_leader_name \tab logical \tab Rbi leader name. \cr
   rbi_leader_shortname \tab logical \tab Rbi leader shortname. \cr
   rbi_leader_headshot \tab logical \tab Rbi leader headshot. \cr
   rbi_leader_team_id \tab logical \tab Rbi leader team id. \cr
   rbi_leader_pos \tab logical \tab Rbi leader pos. \cr
   start_date \tab character \tab Start date. \cr
   broadcast \tab character \tab Broadcast network. \cr
   highlights \tab logical \tab Highlights. \cr
   game_date_time \tab POSIXct \tab Game start date/time (US/Eastern). \cr
   home_team_name \tab character \tab Home team name. \cr
   home_team_logo \tab character \tab Home team logo. \cr
   home_team_abb \tab character \tab Home team abb. \cr
   home_team_id \tab integer \tab Home team id. \cr
   home_team_location \tab character \tab Home team location. \cr
   home_team_full_name \tab character \tab Home team full name. \cr
   home_team_color \tab character \tab Home team color. \cr
   home_score \tab integer \tab Home team run total after the play. \cr
   home_win \tab integer \tab Home win. \cr
   home_record \tab character \tab Home record. \cr
   away_team_name \tab character \tab Away team name. \cr
   away_team_logo \tab character \tab Away team logo. \cr
   away_team_abb \tab character \tab Away team abb. \cr
   away_team_id \tab integer \tab Away team id. \cr
   away_team_location \tab character \tab Away team location. \cr
   away_team_full_name \tab character \tab Away team full name. \cr
   away_team_color \tab character \tab Away team color. \cr
   away_score \tab integer \tab Away team run total after the play. \cr
   away_win \tab integer \tab Away win. \cr
   away_record \tab character \tab Away record. \cr
}
}
\description{
\strong{Get ESPN MLB schedule for a specific year}
}
\examples{

# Get schedule from date 2022-11-17 (returns 1000 results, max allowable.)
\donttest{
try(espn_mlb_scoreboard(season = 20230423))
}
}
\seealso{
Other ESPN MLB Functions:
\code{\link{espn_mlb}},
\code{\link[=espn_mlb_athletes_index]{espn_mlb_athletes_index()}},
\code{\link[=espn_mlb_award]{espn_mlb_award()}},
\code{\link[=espn_mlb_betting]{espn_mlb_betting()}},
\code{\link[=espn_mlb_calendar]{espn_mlb_calendar()}},
\code{\link[=espn_mlb_coach]{espn_mlb_coach()}},
\code{\link[=espn_mlb_coach_record]{espn_mlb_coach_record()}},
\code{\link[=espn_mlb_coach_season]{espn_mlb_coach_season()}},
\code{\link[=espn_mlb_coaches]{espn_mlb_coaches()}},
\code{\link[=espn_mlb_conferences]{espn_mlb_conferences()}},
\code{\link[=espn_mlb_draft]{espn_mlb_draft()}},
\code{\link[=espn_mlb_draft_athlete_detail]{espn_mlb_draft_athlete_detail()}},
\code{\link[=espn_mlb_draft_athletes]{espn_mlb_draft_athletes()}},
\code{\link[=espn_mlb_draft_pick]{espn_mlb_draft_pick()}},
\code{\link[=espn_mlb_draft_rounds]{espn_mlb_draft_rounds()}},
\code{\link[=espn_mlb_draft_status]{espn_mlb_draft_status()}},
\code{\link[=espn_mlb_franchise]{espn_mlb_franchise()}},
\code{\link[=espn_mlb_franchises]{espn_mlb_franchises()}},
\code{\link[=espn_mlb_freeagents]{espn_mlb_freeagents()}},
\code{\link[=espn_mlb_futures]{espn_mlb_futures()}},
\code{\link[=espn_mlb_game_all]{espn_mlb_game_all()}},
\code{\link[=espn_mlb_game_broadcasts]{espn_mlb_game_broadcasts()}},
\code{\link{espn_mlb_game_endpoints}},
\code{\link[=espn_mlb_game_info]{espn_mlb_game_info()}},
\code{\link[=espn_mlb_game_odds]{espn_mlb_game_odds()}},
\code{\link[=espn_mlb_game_official_detail]{espn_mlb_game_official_detail()}},
\code{\link[=espn_mlb_game_officials]{espn_mlb_game_officials()}},
\code{\link[=espn_mlb_game_play]{espn_mlb_game_play()}},
\code{\link[=espn_mlb_game_play_personnel]{espn_mlb_game_play_personnel()}},
\code{\link[=espn_mlb_game_player_box]{espn_mlb_game_player_box()}},
\code{\link[=espn_mlb_game_powerindex]{espn_mlb_game_powerindex()}},
\code{\link[=espn_mlb_game_predictor]{espn_mlb_game_predictor()}},
\code{\link[=espn_mlb_game_probabilities]{espn_mlb_game_probabilities()}},
\code{\link[=espn_mlb_game_probables]{espn_mlb_game_probables()}},
\code{\link[=espn_mlb_game_propbets]{espn_mlb_game_propbets()}},
\code{\link[=espn_mlb_game_rosters]{espn_mlb_game_rosters()}},
\code{\link[=espn_mlb_game_situation]{espn_mlb_game_situation()}},
\code{\link[=espn_mlb_game_team_leaders]{espn_mlb_game_team_leaders()}},
\code{\link[=espn_mlb_game_team_linescores]{espn_mlb_game_team_linescores()}},
\code{\link[=espn_mlb_game_team_records]{espn_mlb_game_team_records()}},
\code{\link[=espn_mlb_game_team_roster]{espn_mlb_game_team_roster()}},
\code{\link[=espn_mlb_game_team_roster_entry]{espn_mlb_game_team_roster_entry()}},
\code{\link[=espn_mlb_game_team_score]{espn_mlb_game_team_score()}},
\code{\link[=espn_mlb_game_team_statistics]{espn_mlb_game_team_statistics()}},
\code{\link[=espn_mlb_injuries]{espn_mlb_injuries()}},
\code{\link[=espn_mlb_leaders]{espn_mlb_leaders()}},
\code{\link[=espn_mlb_news]{espn_mlb_news()}},
\code{\link[=espn_mlb_pbp]{espn_mlb_pbp()}},
\code{\link[=espn_mlb_player_awards]{espn_mlb_player_awards()}},
\code{\link[=espn_mlb_player_box]{espn_mlb_player_box()}},
\code{\link[=espn_mlb_player_career_stats]{espn_mlb_player_career_stats()}},
\code{\link[=espn_mlb_player_contract]{espn_mlb_player_contract()}},
\code{\link[=espn_mlb_player_contracts]{espn_mlb_player_contracts()}},
\code{\link{espn_mlb_player_endpoints}},
\code{\link[=espn_mlb_player_eventlog]{espn_mlb_player_eventlog()}},
\code{\link[=espn_mlb_player_eventlog_v2]{espn_mlb_player_eventlog_v2()}},
\code{\link[=espn_mlb_player_gamelog]{espn_mlb_player_gamelog()}},
\code{\link[=espn_mlb_player_info]{espn_mlb_player_info()}},
\code{\link[=espn_mlb_player_overview]{espn_mlb_player_overview()}},
\code{\link[=espn_mlb_player_seasons]{espn_mlb_player_seasons()}},
\code{\link[=espn_mlb_player_splits]{espn_mlb_player_splits()}},
\code{\link[=espn_mlb_player_statisticslog]{espn_mlb_player_statisticslog()}},
\code{\link[=espn_mlb_player_stats]{espn_mlb_player_stats()}},
\code{\link[=espn_mlb_player_stats_v3]{espn_mlb_player_stats_v3()}},
\code{\link[=espn_mlb_position]{espn_mlb_position()}},
\code{\link[=espn_mlb_positions]{espn_mlb_positions()}},
\code{\link[=espn_mlb_powerindex]{espn_mlb_powerindex()}},
\code{\link[=espn_mlb_season_awards]{espn_mlb_season_awards()}},
\code{\link[=espn_mlb_season_draft]{espn_mlb_season_draft()}},
\code{\link[=espn_mlb_season_group]{espn_mlb_season_group()}},
\code{\link[=espn_mlb_season_group_children]{espn_mlb_season_group_children()}},
\code{\link[=espn_mlb_season_group_teams]{espn_mlb_season_group_teams()}},
\code{\link[=espn_mlb_season_groups]{espn_mlb_season_groups()}},
\code{\link[=espn_mlb_season_info]{espn_mlb_season_info()}},
\code{\link[=espn_mlb_season_leaders]{espn_mlb_season_leaders()}},
\code{\link[=espn_mlb_season_ranking]{espn_mlb_season_ranking()}},
\code{\link[=espn_mlb_season_rankings]{espn_mlb_season_rankings()}},
\code{\link[=espn_mlb_season_type]{espn_mlb_season_type()}},
\code{\link[=espn_mlb_season_types]{espn_mlb_season_types()}},
\code{\link[=espn_mlb_season_week]{espn_mlb_season_week()}},
\code{\link[=espn_mlb_season_weeks]{espn_mlb_season_weeks()}},
\code{\link[=espn_mlb_seasons]{espn_mlb_seasons()}},
\code{\link[=espn_mlb_standings]{espn_mlb_standings()}},
\code{\link[=espn_mlb_team]{espn_mlb_team()}},
\code{\link[=espn_mlb_team_box]{espn_mlb_team_box()}},
\code{\link[=espn_mlb_team_current_roster]{espn_mlb_team_current_roster()}},
\code{\link[=espn_mlb_team_depthchart]{espn_mlb_team_depthchart()}},
\code{\link{espn_mlb_team_endpoints}},
\code{\link[=espn_mlb_team_injuries]{espn_mlb_team_injuries()}},
\code{\link[=espn_mlb_team_leaders]{espn_mlb_team_leaders()}},
\code{\link[=espn_mlb_team_news]{espn_mlb_team_news()}},
\code{\link[=espn_mlb_team_odds_records]{espn_mlb_team_odds_records()}},
\code{\link[=espn_mlb_team_record]{espn_mlb_team_record()}},
\code{\link[=espn_mlb_team_record_detail]{espn_mlb_team_record_detail()}},
\code{\link[=espn_mlb_team_roster]{espn_mlb_team_roster()}},
\code{\link[=espn_mlb_team_schedule]{espn_mlb_team_schedule()}},
\code{\link[=espn_mlb_team_season_profile]{espn_mlb_team_season_profile()}},
\code{\link[=espn_mlb_team_season_roster]{espn_mlb_team_season_roster()}},
\code{\link[=espn_mlb_team_season_statistics]{espn_mlb_team_season_statistics()}},
\code{\link[=espn_mlb_team_stats]{espn_mlb_team_stats()}},
\code{\link[=espn_mlb_teams]{espn_mlb_teams()}},
\code{\link[=espn_mlb_tournament]{espn_mlb_tournament()}},
\code{\link[=espn_mlb_tournament_season]{espn_mlb_tournament_season()}},
\code{\link[=espn_mlb_tournament_seasons]{espn_mlb_tournament_seasons()}},
\code{\link[=espn_mlb_tournaments]{espn_mlb_tournaments()}},
\code{\link[=espn_mlb_transactions]{espn_mlb_transactions()}},
\code{\link[=espn_mlb_venues]{espn_mlb_venues()}},
\code{\link[=espn_mlb_week_ranking]{espn_mlb_week_ranking()}},
\code{\link[=espn_mlb_week_rankings]{espn_mlb_week_rankings()}},
\code{\link[=espn_mlb_wp]{espn_mlb_wp()}}
}
\concept{ESPN MLB Functions}
