% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_batter_leaders.R, R/fg_fielder_leaders.R
\name{fg_bat_leaders}
\alias{fg_bat_leaders}
\title{\strong{(legacy) Scrape Batter Leaderboards from FanGraphs}}
\usage{
fg_bat_leaders(
  age = "",
  pos = "all",
  stats = "bat",
  lg = "all",
  qual = "0",
  startseason = "2023",
  endseason = "2023",
  startdate = "",
  enddate = "",
  month = "0",
  hand = "",
  team = "0",
  pageitems = "10000",
  pagenum = "1",
  ind = "0",
  rost = "0",
  players = "",
  type = "8",
  postseason = "",
  sortdir = "default",
  sortstat = "WAR"
)

fg_bat_leaders(
  age = "",
  pos = "all",
  stats = "bat",
  lg = "all",
  qual = "0",
  startseason = "2023",
  endseason = "2023",
  startdate = "",
  enddate = "",
  month = "0",
  hand = "",
  team = "0",
  pageitems = "10000",
  pagenum = "1",
  ind = "0",
  rost = "0",
  players = "",
  type = "8",
  postseason = "",
  sortdir = "default",
  sortstat = "WAR"
)
}
\arguments{
\item{age}{(integer) Age of players}

\item{pos}{(character) Position of players, defaults to "all". To exclude pitchers, use "np".}

\item{stats}{(character) Statistic to return. Defaults to "bat".}

\item{lg}{(character) League to return. Defaults to "all". Options are "al", "nl", or "all".}

\item{qual}{(character) Whether you want only batters/pitchers that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimum number of plate appearaces/innings pitched, use the number desired. Defaults to "y".}

\item{startseason}{(character) Season for which you want to scrape the data.}

\item{endseason}{(character) Last season for which you want data.}

\item{startdate}{(character) Start date for which you want data.}

\item{enddate}{(character) End date for which you want data.}

\item{month}{(character) Month for which you want data. Defaults to "0"
(full season). Special values: "13" (vs LHP split), "14" (vs RHP split),
"1000" (custom date range). When \code{startdate}/\code{enddate} are supplied and
\code{month} is left at its default, it is set to "1000" automatically so the
date filters take effect.}

\item{hand}{(character) Handedness of batter. Options are "L", "R", or "B". Empty string returns all.}

\item{team}{(character) Teams for which you want data, comma separated.}

\item{pageitems}{(character) Number of items per page.}

\item{pagenum}{(character) Page number.}

\item{ind}{(character) Whether or not to break the seasons out individual, or roll them up together. 1 = split seasons, 0 = aggregate seasons.}

\item{rost}{(character) Whether or not to include players on the roster. 1 = include, 0 = exclude.}

\item{players}{(character) Whether or not to include players on the roster. 1 = include only active roster players, 0 = exclude.}

\item{type}{(character) Defaults to 8, which is the standard leaderboard. The values for the leaderboards appear to go to from type = 0 to 48+, which correspond to links on the leaderboard page.}

\item{postseason}{(logical) Whether or not to include postseason data. TRUE = include postseason, FALSE = exclude postseason.}

\item{sortdir}{(character) Sort direction. Options are "asc" or "desc" or "default".}

\item{sortstat}{(character) Sort by stat. Default is "WAR".}
}
\value{
A data frame of batter data.

A data frame of batter data.
}
\description{
\strong{(legacy) Scrape Batter Leaderboards from FanGraphs}

\strong{(legacy) Scrape Batter Leaderboards from FanGraphs}
}
\keyword{legacy}
