% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_guts.R
\name{fg_guts}
\alias{fg_guts}
\title{\strong{Scrape FanGraphs.com Guts!}}
\usage{
fg_guts()
}
\value{
Returns a tibble of seasonal constants from FanGraphs\tabular{lll}{
   col_name \tab types \tab description \cr
   season \tab integer \tab Season (YYYY). \cr
   lg_woba \tab numeric \tab League-average wOBA for the season. \cr
   woba_scale \tab numeric \tab wOBA scale factor (converts wOBA to runs). \cr
   wBB \tab numeric \tab Linear weight (runs) for an unintentional walk. \cr
   wHBP \tab numeric \tab Linear weight (runs) for a hit-by-pitch. \cr
   w1B \tab numeric \tab Linear weight (runs) for a single. \cr
   w2B \tab numeric \tab Linear weight (runs) for a double. \cr
   w3B \tab numeric \tab Linear weight (runs) for a triple. \cr
   wHR \tab numeric \tab Linear weight (runs) for a home run. \cr
   runSB \tab numeric \tab Run value of a stolen base. \cr
   runCS \tab numeric \tab Run value of a caught stealing. \cr
   lg_r_pa \tab numeric \tab League runs per plate appearance. \cr
   lg_r_w \tab numeric \tab League runs per win. \cr
   cFIP \tab numeric \tab FIP constant for the season. \cr
}
}
\description{
Scrape historical FanGraphs Guts! table, wOBA, FIP coefficients and constants
}
\examples{
\donttest{
  try(fg_guts())
}
}
