% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_fip_plus.R
\name{fip_plus}
\alias{fip_plus}
\title{\strong{Calculate FIP and related metrics for any set of data}}
\usage{
fip_plus(df)
}
\arguments{
\item{df}{A data frame of statistics that includes, at a minimum, the following columns: IP (innings pitched), BF (batters faced), uBB (unintentional walks), HBP (Hit By Pitch), x1B (singles), x2B (doubles), x3B (triples), HR (home runs), AB (at-bats), SH (sacrifice hits), SO (strike outs), and season.}
}
\value{
Returns a tibble with the following columns:\tabular{lll}{
   col_name \tab types \tab description \cr
   bbref_id \tab character \tab Baseball-Reference player id \cr
   season \tab integer \tab Season (4-digit year) \cr
   Name \tab character \tab Player name \cr
   Age \tab numeric \tab Player age \cr
   Level \tab character \tab Level of play (e.g. MLB or minor-league level) \cr
   Team \tab character \tab Team name \cr
   G \tab numeric \tab Games \cr
   GS \tab numeric \tab Games started \cr
   W \tab numeric \tab Wins \cr
   L \tab numeric \tab Losses \cr
   SV \tab numeric \tab Saves \cr
   IP \tab numeric \tab Innings pitched \cr
   H \tab numeric \tab Hits allowed \cr
   R \tab numeric \tab Runs allowed \cr
   ER \tab numeric \tab Earned runs allowed \cr
   uBB \tab numeric \tab Unintentional walks allowed \cr
   BB \tab numeric \tab Walks (bases on balls) allowed \cr
   SO \tab numeric \tab Strikeouts \cr
   HR \tab numeric \tab Home runs allowed \cr
   HBP \tab numeric \tab Hit batters \cr
   ERA \tab numeric \tab Earned run average \cr
   AB \tab numeric \tab At-bats against \cr
   X1B \tab numeric \tab Singles allowed \cr
   X2B \tab numeric \tab Doubles allowed \cr
   X3B \tab numeric \tab Triples allowed \cr
   IBB \tab numeric \tab Intentional walks allowed \cr
   GDP \tab numeric \tab Grounded into double plays induced \cr
   SF \tab numeric \tab Sacrifice flies allowed \cr
   SB \tab numeric \tab Stolen bases allowed \cr
   CS \tab numeric \tab Caught stealing \cr
   PO \tab numeric \tab Pickoffs \cr
   BF \tab numeric \tab Batters faced \cr
   Pit \tab numeric \tab Pitches thrown \cr
   Str \tab numeric \tab Strikes thrown \cr
   StL \tab numeric \tab Strikes looking (called) \cr
   StS \tab numeric \tab Strikes swinging (whiffs) \cr
   GB.FB \tab numeric \tab Ground ball to fly ball ratio \cr
   LD \tab numeric \tab Line drives allowed \cr
   PU \tab numeric \tab Pop ups (infield fly) induced \cr
   WHIP \tab numeric \tab Walks plus hits per inning pitched \cr
   BAbip \tab numeric \tab Batting average on balls in play allowed \cr
   SO9 \tab numeric \tab Strikeouts per nine innings \cr
   SO.W \tab numeric \tab Strikeout-to-walk ratio \cr
   SO_perc \tab numeric \tab Strikeout percentage \cr
   uBB_perc \tab numeric \tab Unintentional walk percentage \cr
   SO_uBB \tab numeric \tab Strikeout percentage minus unintentional walk percentage \cr
   FIP \tab numeric \tab Fielding independent pitching \cr
   wOBA_against \tab numeric \tab Weighted on-base average against (based on batters faced) \cr
   wOBA_CON_against \tab numeric \tab Weighted on-base average against per instance of fair contact \cr
}
}
\description{
This function allows you to calculate FIP and related metrics for any given set of data, provided the right variables are in the data set. The function currently returns both FIP per inning pitched, wOBA against (based on batters faced), and wOBA against per instance of fair contact.
}
\examples{
\donttest{
  try({
    df <- bref_daily_pitcher("2015-04-05", "2015-04-30")
    fip_plus(df)
  })
}
}
