% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_mlb.R
\name{fox_mlb_team_gamelog}
\alias{fox_mlb_team_gamelog}
\title{\strong{Get Fox Sports MLB team game log}}
\usage{
fox_mlb_team_gamelog(team_id)
}
\arguments{
\item{team_id}{Fox Bifrost team id.}
}
\value{
A \code{baseballr_data} tibble (long): \code{team_id}, \code{season_type},
\code{category}, \code{game_id}, \code{game_date}, \code{opponent}, \code{stat}, \code{value}.
}
\description{
\strong{Get Fox Sports MLB team game log}
}
\examples{
\donttest{ try(fox_mlb_team_gamelog("1")) }
}
