% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_mlb.R
\name{fox_mlb_team_stats}
\alias{fox_mlb_team_stats}
\title{\strong{Get Fox Sports MLB team stat leaders}}
\usage{
fox_mlb_team_stats(team_id)
}
\arguments{
\item{team_id}{Fox Bifrost team id.}
}
\value{
A \code{baseballr_data} tibble: \code{team_id}, \code{category}, \code{stat},
\code{stat_abbreviation}, \code{player}, \code{value}.
}
\description{
\strong{Get Fox Sports MLB team stat leaders}
}
\examples{
\donttest{ try(fox_mlb_team_stats("1")) }
}
