% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_batting_orders.R
\name{get_batting_orders}
\alias{get_batting_orders}
\title{\strong{(legacy) Retrieve batting orders for a given MLB game}}
\usage{
get_batting_orders(game_pk, type = "starting")
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}

\item{type}{Whether to just return the starting lineup ('starting') or all
batters that appeared ('all')}
}
\value{
Returns a tibble that includes probable starting pitchers and
the home plate umpire for the \code{game_pk} requested
}
\description{
\strong{(legacy) Retrieve batting orders for a given MLB game}
}
\keyword{legacy}
