% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chadwick_lu.R
\name{get_chadwick_lu}
\alias{get_chadwick_lu}
\title{Directly download the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.}
\usage{
get_chadwick_lu()
}
\description{
Directly download the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.
}
\examples{
\dontrun{get_chadwick_lu()}
}
\keyword{MLB,}
\keyword{sabermetrics}
