% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_draft.R
\name{get_draft_mlb}
\alias{get_draft_mlb}
\title{\strong{(legacy) Retrieve draft pick information by year}}
\usage{
get_draft_mlb(year)
}
\arguments{
\item{year}{The year for which to return data}
}
\value{
Returns a data frame with information for every draft pick in every round for the year requested
}
\description{
\strong{(legacy) Retrieve draft pick information by year}
}
\keyword{legacy}
