% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_info_mlb.R
\name{get_game_info_mlb}
\alias{get_game_info_mlb}
\title{Retrieve additional game information via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
get_game_info_mlb(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame that includes game_pk values for the date
requested
}
\description{
Retrieve additional game information via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
get_probables_mlb(566001)
}
\keyword{MLB,}
\keyword{sabermetrics}
