% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_info_sup_petti.R
\name{get_game_info_sup_petti}
\alias{get_game_info_sup_petti}
\title{\strong{Download a data frame of supplemental data about MLB games since 2008.}}
\usage{
get_game_info_sup_petti()
}
\value{
Function returns a data frame with various columns, including:
\itemize{
\item game_pk
\item game_date
\item venue id
\item attendance
\item game temperature
\item wind speed
\item direction
\item start time
\item end time
}
}
\description{
\strong{Download a data frame of supplemental data about MLB games since 2008.}
}
\examples{
\donttest{
  get_game_info_sup_petti()
}
}
