% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_park_factor.R
\name{get_ncaa_park_factor}
\alias{get_ncaa_park_factor}
\title{Get Park Effects for NCAA Baseball Teams}
\usage{
get_ncaa_park_factor(teamid, years)
}
\arguments{
\item{teamid}{The team's unique NCAA id.}

\item{years}{The season or seasons (i.e. use 2016 for the 2015-2016 season,
etc., limited to just 2013-2020 seasons)}
}
\value{
A dataframe with the following fields: school, home_games,
away_games, runs_scored_home (the total runs scored in home games by the selected team over the selected period of time),
runs_allowed_home (the total runs allowed in home games by the selected team over the selected period of time),
run_scored_away (the total runs scored in road games by the selected team over the selected period of time),
runs_allowed_away (the total runs allowed in road games by the selected team over the selected period of time),
base_pf (base park factor, comparing the run scoring environment of home games v. non-home games),
final_pf (park factor after adjustments, like adjusting how regression should be applied for how many years are selected,
ideology taken from this article: https://library.fangraphs.com/park-factors-5-year-regressed/)
}
\description{
Get Park Effects for NCAA Baseball Teams
}
\examples{
\dontrun{get_ncaa_park_factor(736, c(2017:2019)}
}
