% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_probables.R
\name{get_probables_mlb}
\alias{get_probables_mlb}
\title{\strong{(legacy) Retrieve probable starters for a given MLB game}}
\usage{
get_probables_mlb(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a tibble that includes probable starting pitchers and
the home plate umpire for the \code{game_pk} requested
}
\description{
\strong{(legacy) Retrieve probable starters for a given MLB game}
}
\keyword{legacy}
