% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpitchchart.R
\name{ggpitchchart}
\alias{ggpitchchart}
\title{\strong{Pitch charts with ggplot2}}
\usage{
ggpitchchart(
  data,
  x_value = "plate_x",
  z_value = "plate_z",
  sz_top = NULL,
  sz_bot = NULL,
  lim = NA,
  color_value = NULL,
  density = FALSE,
  bin_size = 8,
  scall_fill_palette = "Blues",
  point_size = 10,
  point_alpha = 0.75,
  div9 = TRUE,
  base = FALSE,
  title = NULL
)
}
\arguments{
\item{data}{A data that includes ball coordinates.}

\item{x_value}{The x coordinates. Typically plate_x.}

\item{z_value}{The y coordinates. Typically plate_z.}

\item{sz_top}{The top of strike zone. Typically sz_top. If left blank, defaults to average height.}

\item{sz_bot}{The top of strike zone. Typically sz_bot. If left blank, defaults to average height.}

\item{lim}{Length from center to edge. Defaults to NA.}

\item{color_value}{The categorical variable that you want the geom_points to base the color on. Defaults to blue.}

\item{density}{Chooses between a 2d density plot or a point plot. Defaults to FALSE.}

\item{bin_size}{Size of bins used if use a density plot. Defaults to 8.}

\item{scall_fill_palette}{You can use RColorBrewer's palette if use a density plot. Defaults to Blues.}

\item{point_size}{Set the size of geom_point if used. Defaults to 10.}

\item{point_alpha}{Set the alpha of geom_point if used. Defaults to 0.75.}

\item{div9}{Whether to divide the strike zone into 9. Defaults to TRUE.}

\item{base}{Whether to show the home base. Defaults to FALSE}

\item{title}{The prot title.}
}
\value{
The pitch chart.
}
\description{
This function allows you to create pitch charts with ggplots given a data ball coordinates and The top and bottom of the strike zone.(The limits of the strike zone can be omitted)
}
\examples{
\dontrun{
ggpitchchart(df, x_value = "plate_x", z_value = "plate_z", sz_top = "sz_top", sz_bot = "sz_bot", color_value = "pitch_type", base = TRUE, title = "Ohtani, 9.24")
ggpitchchart(df, x_value = "plate_x", z_value = "plate_z", sz_top = "sz_top", sz_bot = "sz_bot", lim = 2, density = TRUE, div9 = FALSE, title = "Ohtani, 9.24")
}
}
