% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mlb_data.R
\name{helper_espn_mlb_pbp}
\alias{helper_espn_mlb_pbp}
\title{\strong{Parse ESPN MLB PBP, helper function}}
\usage{
helper_espn_mlb_pbp(resp)
}
\arguments{
\item{resp}{Response object from the ESPN MLB game summary endpoint}
}
\value{
Returns a tibble
}
\description{
\strong{Parse ESPN MLB PBP, helper function}
}
