% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_baseball.R
\name{load_baseball}
\alias{load_baseball}
\alias{load_data}
\alias{baseballr_data}
\title{\strong{baseballr Data Loader Overview}}
\description{
\itemize{
\item \code{load_umpire_ids()}: Download a data frame of all umpires and their mlbamids for games since 2008.
\item \code{load_game_info_sup()}: Download a data frame of supplemental data about MLB games since 2008.
\item \code{load_ncaa_baseball_pbp()}: Load cleaned NCAA baseball play-by-play data from the baseballr data repo.
\item \code{load_ncaa_baseball_schedule()}: Load cleaned NCAA baseball schedule from the baseballr data repo.
\item \code{load_ncaa_baseball_season_ids()}: Load cleaned NCAA men's college baseball season IDs from the baseballr data repo.
\item \code{load_ncaa_baseball_teams()}: Load cleaned NCAA men's college baseball teams from the baseballr data repo.
}
}
\details{
\subsection{\strong{baseballr Data}}{

These functions load cleaned MLB and NCAA baseball data from the baseballr data repository.
}
}
\concept{baseballr Data}
