% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_from_url}
\alias{load_from_url}
\title{Load rds/parquet/csv files from remote URL}
\usage{
load_from_url(url, ..., seasons = TRUE, baseballr = FALSE)
}
\arguments{
\item{url}{a vector of URLs to load into memory. If more than one URL provided, will row-bind them.}

\item{...}{named arguments that will be added as attributes to the data}

\item{seasons}{a numeric vector of years that will be used to filter the dataframe's \code{season} column. If \code{TRUE} (default), does not filter.}

\item{baseballr}{TRUE to add baseballr_data classing and attributes.}
}
\value{
a dataframe
}
\description{
Load rds/parquet/csv files from remote URL
}
