% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_game_info_sup.R
\name{load_game_info_sup}
\alias{load_game_info_sup}
\title{\strong{Download a data frame of supplemental data about MLB games since 2008.}}
\usage{
load_game_info_sup()
}
\value{
Function returns a tibble with various columns, including:
\itemize{
\item game_pk
\item game_date
\item venue id
\item attendance
\item game temperature
\item wind speed
\item direction
\item start time
\item end time
}
}
\description{
\strong{Download a data frame of supplemental data about MLB games since 2008.}
}
\examples{
\donttest{
  try(load_game_info_sup())
}
}
