% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_baseball_pbp.R
\name{load_ncaa_baseball_pbp}
\alias{load_ncaa_baseball_pbp}
\title{\strong{Load baseballr NCAA baseball play-by-play}}
\usage{
load_ncaa_baseball_pbp(
  seasons = most_recent_ncaa_baseball_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)
}
\arguments{
\item{seasons}{A vector of 4-digit years associated with given NCAA college baseball seasons. (Min: 2022)}

\item{...}{Additional arguments passed to an underlying function that writes
the season data into a database (used by \code{update_ncaa_baseball_db()}).}

\item{dbConnection}{A \code{DBIConnection} object, as returned by}

\item{tablename}{The name of the schedule data table within the database}
}
\value{
Returns a tibble
}
\description{
helper that loads multiple seasons from the data repo either into memory
or writes it into a db using some forwarded arguments in the dots
}
\examples{
\donttest{
  load_ncaa_baseball_pbp(seasons = 2022)
}
}
