% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_umpire_ids.R
\name{load_umpire_ids}
\alias{load_umpire_ids}
\title{\strong{Download a data frame of all umpires and their mlbamids for games since 2008}}
\usage{
load_umpire_ids()
}
\value{
Function returns a tibble with the following columns:
\itemize{
\item id
\item position
\item name
\item game_pk
\item game_date
}
}
\description{
\strong{Download a data frame of all umpires and their mlbamids for games since 2008}
}
\examples{
\donttest{
  try(load_umpire_ids())
}
}
