% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{master_ncaa_team_lu}
\alias{master_ncaa_team_lu}
\title{A data set colleges and their athletic conferences and divisions}
\format{
A data frame with 4,774 rows and 6 variables
}
\usage{
data(master_ncaa_team_lu)
}
\description{
A data set colleges and their athletic conferences and divisions keyed by school and year.

\itemize{
  \item \code{school}: School name
  \item \code{conference}: Conference name
  \item \code{school_id}: School id
  \item \code{year}: Year date
  \item \code{division}: School division
  \item \code{conference_id}: School conference id
}
}
\details{
Dataset
}
\keyword{datasets}
