% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_batter_game_logs.R
\name{milb_batter_game_logs_fg}
\alias{milb_batter_game_logs_fg}
\title{\strong{(legacy) Scrape MiLB game logs for batters from FanGraphs}}
\usage{
milb_batter_game_logs_fg(playerid, year)
}
\arguments{
\item{playerid}{The batter's minor league ID from FanGraphs.}

\item{year}{The season for which game logs should be returned.}
}
\value{
Returns a data frame of Minor League batter game logs.
}
\description{
\strong{(legacy) Scrape MiLB game logs for batters from FanGraphs}
}
\keyword{legacy}
