% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_pitcher_game_logs.R
\name{milb_pitcher_game_logs_fg}
\alias{milb_pitcher_game_logs_fg}
\title{\strong{(legacy) Scrape MiLB game logs for pitchers from FanGraphs}}
\usage{
milb_pitcher_game_logs_fg(playerid, year)
}
\arguments{
\item{playerid}{The pitcher's minor league ID from FanGraphs.com.}

\item{year}{The season for which game logs should be returned.}
}
\value{
Returns a tibble of Minor League pitcher game logs.
}
\description{
\strong{(legacy) Scrape MiLB game logs for pitchers from FanGraphs}
}
\keyword{legacy}
