% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_all_star_ballots.R
\name{mlb_all_star}
\alias{mlb_all_star}
\alias{all_star}
\alias{allstar}
\title{\strong{MLB All-Star Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_all_star_ballots()}: Find MLB All-Star Ballots.
\item \code{mlb_all_star_final_vote()}: Find MLB All-Star Final Vote.
\item \code{mlb_all_star_write_ins()}: Find MLB All-Star Write-ins.
}
}
\details{
\subsection{\strong{MLB All-Star}}{

These functions retrieve MLB All-Star ballot, final vote, and write-in information from the MLB Stats API.
}
}
\concept{MLB All-Star}
