% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_all_star_write_ins.R
\name{mlb_all_star_write_ins}
\alias{mlb_all_star_write_ins}
\title{\strong{Find MLB All-Star Write-ins}}
\usage{
mlb_all_star_write_ins(league_id = NULL, season = NULL)
}
\arguments{
\item{league_id}{League ID for league all-star ballot of interest.}

\item{season}{The season of the all-star ballot.}
}
\value{
Returns a data frame with the following columns:\tabular{ll}{
   col_name \tab types \cr
   player_id \tab integer \cr
   full_name \tab character \cr
   link \tab character \cr
   first_name \tab character \cr
   last_name \tab character \cr
   primary_number \tab character \cr
   birth_date \tab character \cr
   current_age \tab integer \cr
   birth_city \tab character \cr
   birth_state_province \tab character \cr
   birth_country \tab character \cr
   height \tab character \cr
   weight \tab integer \cr
   active \tab logical \cr
   use_name \tab character \cr
   middle_name \tab character \cr
   boxscore_name \tab character \cr
   nick_name \tab character \cr
   gender \tab character \cr
   is_player \tab logical \cr
   is_verified \tab logical \cr
   draft_year \tab integer \cr
   mlb_debut_date \tab character \cr
   name_first_last \tab character \cr
   name_slug \tab character \cr
   first_last_name \tab character \cr
   last_first_name \tab character \cr
   last_init_name \tab character \cr
   init_last_name \tab character \cr
   full_fml_name \tab character \cr
   full_lfm_name \tab character \cr
   strike_zone_top \tab numeric \cr
   strike_zone_bottom \tab numeric \cr
   pronunciation \tab character \cr
   name_matrilineal \tab character \cr
   name_title \tab character \cr
   primary_position_code \tab character \cr
   primary_position_name \tab character \cr
   primary_position_type \tab character \cr
   primary_position_abbreviation \tab character \cr
   bat_side_code \tab character \cr
   bat_side_description \tab character \cr
   pitch_hand_code \tab character \cr
   pitch_hand_description \tab character \cr
   league_id \tab numeric \cr
   season \tab numeric \cr
}
}
\description{
\strong{Find MLB All-Star Write-ins}
}
\examples{
\donttest{
 try(mlb_all_star_write_ins(league_id = 103, season = 2021))
}
}
