% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_award.R
\name{mlb_award}
\alias{mlb_award}
\title{\strong{MLB All-Star, Awards, Home Run Derby Functions}}
\description{
\describe{
\item{\code{mlb_all_star_ballots()}:}{ Find MLB All-Star Ballots.}
\item{\code{mlb_all_star_final_vote()}:}{ Find MLB All-Star Final Vote.}
\item{\code{mlb_all_star_write_ins()}:}{ Find MLB All-Star Write-ins.}
\item{\code{mlb_awards()}:}{ Find MLB Awards.}
\item{\code{mlb_awards_recipient()}:}{ Find MLB Award Recipients.}
\item{\code{mlb_homerun_derby()}:}{ Retrieve MLB Home Run Derby Data.}
\item{\code{mlb_homerun_derby_bracket()}:}{ Retrieve MLB Home Run Derby Bracket.}
\item{\code{mlb_homerun_derby_players()}:}{ Retrieve MLB Home Run Derby Players.}
}
}
\details{
\subsection{\strong{Find MLB All-Star Ballots}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(mlb_all_star_ballots(league_id = 103, season = 2021))
}\if{html}{\out{</div>}}
}

\subsection{\strong{Find MLB All-Star Final Vote}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(mlb_all_star_final_vote(league_id = 103, season = 2021))
}\if{html}{\out{</div>}}
}

\subsection{\strong{Find MLB All-Star Write-ins}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(mlb_all_star_write_ins(league_id = 103, season = 2021))
}\if{html}{\out{</div>}}
}

\subsection{\strong{Find MLB Awards}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(mlb_awards())
}\if{html}{\out{</div>}}
}

\subsection{\strong{Find MLB Award Recipients}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(mlb_awards_recipient(award_id = 'MLBHOF', season = 2020))
}\if{html}{\out{</div>}}
}

\subsection{\strong{Retrieve MLB Home Run Derby Data}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(mlb_homerun_derby(game_pk = 511101))
}\if{html}{\out{</div>}}
}

\subsection{\strong{Retrieve MLB Home Run Derby Bracket}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{   try(mlb_homerun_derby_bracket(game_pk = 511101))
}\if{html}{\out{</div>}}
}

\subsection{\strong{Retrieve MLB Home Run Derby Players}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{   try(mlb_homerun_derby_players(game_pk = 511101))
}\if{html}{\out{</div>}}
}
}
