% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_awards_recipient.R
\name{mlb_awards_recipient}
\alias{mlb_awards_recipient}
\title{\strong{MLB Award Recipients}}
\usage{
mlb_awards_recipient(
  award_id = NULL,
  sport_id = NULL,
  league_id = NULL,
  season = NULL
)
}
\arguments{
\item{award_id}{award_id to return a directory of players for a given award.}

\item{sport_id}{sport_id to return a directory of players for a given aware in a specific sport.}

\item{league_id}{league_id(s) to return a directory of players for a given award in a specific league. Format '103,104'}

\item{season}{Year(s) to return a directory of players for a given award in a given season.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   award_id \tab character \tab Award identifier code. \cr
   award_name \tab character \tab Award name. \cr
   date \tab character \tab Date the award was given (YYYY-MM-DD). \cr
   season \tab character \tab Season the award was given (YYYY). \cr
   votes \tab integer \tab Number of votes received. \cr
   notes \tab character \tab Additional notes about the recipient. \cr
   player_id \tab integer \tab MLB player ID of the recipient. \cr
   player_link \tab character \tab MLB Stats API relative player link. \cr
   player_name_first_last \tab character \tab Recipient name in first-last order. \cr
   player_primary_position_code \tab character \tab Recipient primary fielding position code. \cr
   player_primary_position_name \tab character \tab Recipient primary fielding position name. \cr
   player_primary_position_type \tab character \tab Recipient primary position type. \cr
   player_primary_position_abbreviation \tab character \tab Recipient primary position abbreviation. \cr
   team_id \tab integer \tab MLB team ID of the recipient. \cr
   team_link \tab character \tab MLB Stats API relative team link. \cr
}
}
\description{
\strong{MLB Award Recipients}
}
\examples{
\donttest{
  try(mlb_awards_recipient(award_id = 'MLBHOF', season = 2020))
}
}
