% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_baseball_stats.R
\name{mlb_baseball_stats}
\alias{mlb_baseball_stats}
\title{\strong{MLB Baseball Stats}}
\usage{
mlb_baseball_stats()
}
\value{
Returns a tibble with the following columns:\tabular{lll}{
   col_name \tab types \tab description \cr
   stat_name \tab character \tab Internal stat name. \cr
   stat_lookup_param \tab character \tab Lookup parameter/abbreviation for the stat. \cr
   is_counting \tab logical \tab Whether the stat is a counting stat. \cr
   stat_label \tab character \tab Human-readable stat label. \cr
   stat_group \tab character \tab Stat group (e.g. hitting, pitching, fielding). \cr
}
}
\description{
\strong{MLB Baseball Stats}
}
\examples{
\donttest{
  try(mlb_baseball_stats())
}
}
