% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_batting_orders.R
\name{mlb_batting_orders}
\alias{mlb_batting_orders}
\title{\strong{Retrieve batting orders for a given MLB game}}
\usage{
mlb_batting_orders(game_pk, type = "starting")
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}

\item{type}{Whether to just return the starting lineup ('starting') or all
batters that appeared ('all')}
}
\value{
Returns a tibble that includes probable starting pitchers and
the home plate umpire for the \code{game_pk} requested\tabular{lll}{
   col_name \tab types \tab description \cr
   id \tab integer \tab MLB player ID. \cr
   fullName \tab character \tab Player full name. \cr
   abbreviation \tab character \tab Fielding position abbreviation. \cr
   batting_order \tab character \tab Spot in the batting order (1-9). \cr
   batting_position_num \tab character \tab Sub-position within the lineup spot (0 = starter). \cr
   team \tab character \tab Side of the matchup ('home' or 'away'). \cr
   teamName \tab character \tab Team name. \cr
   teamID \tab integer \tab MLB team ID. \cr
}
}
\description{
\strong{Retrieve batting orders for a given MLB game}
}
\examples{
\donttest{
  try(mlb_batting_orders(game_pk=566001))
}
}
