% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_conferences.R
\name{mlb_conferences}
\alias{mlb_conferences}
\title{\strong{View all PCL conferences}}
\usage{
mlb_conferences(conference_id = NULL, season = NULL)
}
\arguments{
\item{conference_id}{Conference ID to return information for.}

\item{season}{Year to return to return conference information for.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   conference_id \tab integer \tab MLB conference ID. \cr
   conference_name \tab character \tab Conference name. \cr
   link \tab character \tab MLB Stats API relative conference link. \cr
   conference_abbreviation \tab character \tab Conference abbreviation. \cr
   has_wildcard \tab logical \tab Whether the conference has a wild card. \cr
   conference_name_short \tab character \tab Short conference name. \cr
   league_id \tab integer \tab MLB league ID. \cr
   league_link \tab character \tab MLB Stats API relative league link. \cr
   sport_id \tab integer \tab MLB sport ID. \cr
   sport_link \tab character \tab MLB Stats API relative sport link. \cr
}
}
\description{
\strong{View all PCL conferences}
}
\examples{
\donttest{
  try(mlb_conferences())
  try(mlb_conferences(conference_id =  301, season = 2020))
}
}
