% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_divisions.R
\name{mlb_divisions}
\alias{mlb_divisions}
\title{\strong{MLB Divisions}}
\usage{
mlb_divisions(division_id = NULL, league_id = NULL, sport_id = NULL)
}
\arguments{
\item{division_id}{Return division(s) data for a specific division}

\item{league_id}{Return division(s) data for all divisions in a specific league}

\item{sport_id}{Return division(s) for all divisions in a specific sport.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   division_id \tab integer \tab MLB division ID. \cr
   division_name \tab character \tab Division name. \cr
   season \tab character \tab Season (YYYY). \cr
   division_name_short \tab character \tab Short division name. \cr
   division_link \tab character \tab MLB Stats API relative division link. \cr
   division_abbreviation \tab character \tab Division abbreviation. \cr
   has_wildcard \tab logical \tab Whether the division has a wild card. \cr
   sort_order \tab integer \tab Display sort order for the division. \cr
   num_playoff_teams \tab integer \tab Number of playoff teams from the division. \cr
   active \tab logical \tab Whether the division is currently active. \cr
   league_id \tab integer \tab MLB league ID. \cr
   league_link \tab character \tab MLB Stats API relative league link. \cr
   sport_id \tab integer \tab MLB sport ID. \cr
   sport_link \tab character \tab MLB Stats API relative sport link. \cr
}
}
\description{
\strong{MLB Divisions}
}
\examples{
\donttest{
  try(mlb_divisions(sport_id = 1))
}
}
