% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_draft.R
\name{mlb_draft_endpoints}
\alias{mlb_draft_endpoints}
\alias{mlb_draft}
\alias{draft}
\title{\strong{MLB Draft Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_draft()}: Retrieve draft pick information by year.
\item \code{mlb_draft_latest()}: Retrieve latest draft information by year.
\item \code{mlb_draft_prospects()}: Retrieve draft prospect information by year.
}
}
\details{
\subsection{\strong{MLB Draft}}{

These functions retrieve MLB Draft pick, latest pick, and prospect information from the MLB Stats API.
}
}
\concept{MLB Draft}
