% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_event_types.R
\name{mlb_event_types}
\alias{mlb_event_types}
\title{\strong{MLB Event Types}}
\usage{
mlb_event_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   plate_appearance \tab logical \tab Whether the event counts as a plate appearance. \cr
   hit \tab logical \tab Whether the event is a hit. \cr
   event_code \tab character \tab Event type code. \cr
   base_running_event \tab logical \tab Whether the event is a base-running event. \cr
   event_description \tab character \tab Human-readable event description. \cr
}
}
\description{
\strong{MLB Event Types}
}
\examples{
\donttest{
  try(mlb_event_types())
}
}
