% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_fielder_detail_types.R
\name{mlb_fielder_detail_types}
\alias{mlb_fielder_detail_types}
\title{\strong{MLB Fielder Detail Types}}
\usage{
mlb_fielder_detail_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   stat_name \tab character \tab Internal fielder detail stat name. \cr
   code \tab character \tab Fielder detail type code. \cr
   names \tab list \tab Associated detail names for the type. \cr
   chance \tab logical \tab Whether the detail counts as a fielding chance. \cr
   error \tab logical \tab Whether the detail counts as an error. \cr
}
}
\description{
\strong{MLB Fielder Detail Types}
}
\examples{
\donttest{
  try(mlb_fielder_detail_types())
}
}
