% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_info.R
\name{mlb_game_endpoints}
\alias{mlb_game_endpoints}
\alias{game}
\alias{pbp}
\title{\strong{MLB Game Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_game_info()}: Retrieve additional game information for major and minor league games.
\item \code{mlb_game_pks()}: Get MLB Game Info by Date and Level.
\item \code{mlb_game_content()}: Retrieve additional game content for major and minor league games.
\item \code{mlb_game_linescore()}: Retrieve game linescores for major and minor league games.
\item \code{mlb_game_wp()}: Acquire win probability for Major and Minor League games.
\item \code{mlb_game_pace()}: Retrieve game pace metrics for major and minor league.
\item \code{mlb_game_changes()}: Acquire time codes for Major and Minor League games.
\item \code{mlb_game_context_metrics()}: Acquire game context metrics for Major and Minor League games.
\item \code{mlb_pbp()}: Acquire pitch-by-pitch data for Major and Minor League games.
\item \code{mlb_pbp_diff()}: Acquire pitch-by-pitch data between two timecodes for Major and Minor League games.
}
}
\details{
\subsection{\strong{MLB Game}}{

These functions retrieve MLB game information, content, linescores, win probability, pace, changes, context metrics, and pitch-by-pitch data from the MLB Stats API.
}
}
\concept{MLB Game}
