% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_pace.R
\name{mlb_game_pace}
\alias{mlb_game_pace}
\title{\strong{Retrieve game pace metrics for major and minor league}}
\usage{
mlb_game_pace(
  season,
  league_ids = NULL,
  sport_ids = NULL,
  team_ids = NULL,
  game_type = NULL,
  venue_ids = NULL,
  org_type = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{season}{Year for which to return information (\emph{Required}).}

\item{league_ids}{The league_id(s) for which to return information.}

\item{sport_ids}{The sport_id(s) for which to return information.}

\item{team_ids}{The team_id(s) for which to return information.}

\item{game_type}{The game_type for which to return information.}

\item{venue_ids}{Venue directorial information based venue_id.}

\item{org_type}{pace of game metrics based on team ('T'), league ('L') or sport('S')}

\item{start_date}{Date of first game for which you want data.
Format must be in MM/DD/YYYY format.}

\item{end_date}{Date of last game for which you want data.
Format must be in MM/DD/YYYY format.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   hits_per9inn \tab numeric \tab Hits per 9 innings. \cr
   runs_per9inn \tab numeric \tab Runs per 9 innings. \cr
   pitches_per9inn \tab numeric \tab Pitches per 9 innings. \cr
   plate_appearances_per9inn \tab numeric \tab Plate appearances per 9 innings. \cr
   hits_per_game \tab numeric \tab Hits per game. \cr
   runs_per_game \tab numeric \tab Runs per game. \cr
   innings_played_per_game \tab numeric \tab Innings played per game. \cr
   pitches_per_game \tab numeric \tab Pitches per game. \cr
   pitchers_per_game \tab numeric \tab Pitchers used per game. \cr
   plate_appearances_per_game \tab numeric \tab Plate appearances per game. \cr
   total_game_time \tab character \tab Total game time (HHH:MM:SS). \cr
   total_innings_played \tab numeric \tab Total innings played. \cr
   total_hits \tab integer \tab Total hits. \cr
   total_runs \tab integer \tab Total runs. \cr
   total_plate_appearances \tab integer \tab Total plate appearances. \cr
   total_pitchers \tab integer \tab Total pitchers used. \cr
   total_pitches \tab integer \tab Total pitches thrown. \cr
   total_games \tab integer \tab Total games. \cr
   total7inn_games \tab integer \tab Total 7-inning games. \cr
   total9inn_games \tab integer \tab Total 9-inning games. \cr
   total_extra_inn_games \tab integer \tab Total extra-inning games. \cr
   time_per_game \tab character \tab Average time per game (HH:MM:SS). \cr
   time_per_pitch \tab character \tab Average time per pitch (HH:MM:SS). \cr
   time_per_hit \tab character \tab Average time per hit (HH:MM:SS). \cr
   time_per_run \tab character \tab Average time per run (HH:MM:SS). \cr
   time_per_plate_appearance \tab character \tab Average time per plate appearance (HH:MM:SS). \cr
   time_per9inn \tab character \tab Average time per 9 innings (HH:MM:SS). \cr
   time_per77plate_appearances \tab character \tab Average time per 77 plate appearances. \cr
   total_extra_inn_time \tab character \tab Total extra-inning time (HHH:MM:SS). \cr
   time_per7inn_game \tab character \tab Average time per 7-inning game (HH:MM:SS). \cr
   time_per7inn_game_without_extra_inn \tab character \tab Average time per 7-inning game excl. extras. \cr
   total7inn_games_scheduled \tab integer \tab Total 7-inning games scheduled. \cr
   total7inn_games_without_extra_inn \tab integer \tab Total 7-inning games without extra innings. \cr
   total9inn_games_without_extra_inn \tab integer \tab Total 9-inning games without extra innings. \cr
   total9inn_games_scheduled \tab integer \tab Total 9-inning games scheduled. \cr
   hits_per_run \tab numeric \tab Hits per run. \cr
   pitches_per_pitcher \tab numeric \tab Pitches per pitcher. \cr
   season \tab character \tab Season (YYYY). \cr
   total9inn_games_completed_early \tab integer \tab 9-inning games completed early. \cr
   total7inn_games_completed_early \tab integer \tab 7-inning games completed early. \cr
   sport_id \tab integer \tab MLB sport ID. \cr
   sport_code \tab character \tab Sport code (e.g. mlb, aaa). \cr
   sport_link \tab character \tab MLB Stats API relative sport link. \cr
   pr_portal_calculated_fields_total7inn_games \tab integer \tab Portal-calculated total 7-inning games. \cr
   pr_portal_calculated_fields_total9inn_games \tab integer \tab Portal-calculated total 9-inning games. \cr
   pr_portal_calculated_fields_total_extra_inn_games \tab integer \tab Portal-calculated total extra-inning games. \cr
   pr_portal_calculated_fields_time_per7inn_game \tab character \tab Portal-calculated time per 7-inning game. \cr
   pr_portal_calculated_fields_time_per9inn_game \tab character \tab Portal-calculated time per 9-inning game. \cr
   pr_portal_calculated_fields_time_per_extra_inn_game \tab character \tab Portal-calculated time per extra-inning game. \cr
}
}
\description{
\strong{Retrieve game pace metrics for major and minor league}
}
\examples{
\donttest{
  try(mlb_game_pace(season = 2021, start_date = "09/14/2021", end_date = "09/16/2021"))
}
}
