% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_types.R
\name{mlb_game_types}
\alias{mlb_game_types}
\title{\strong{MLB Game Types}}
\usage{
mlb_game_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   game_type_id \tab character \tab Single-letter game type code (e.g. 'S', 'R', 'P', 'W'). \cr
   game_type_description \tab character \tab Game type description (e.g. 'Spring Training', 'Regular Season'). \cr
}
}
\description{
\strong{MLB Game Types}
}
\examples{
\donttest{
  try(mlb_game_types())
}
}
