% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_hit_trajectories.R
\name{mlb_hit_trajectories}
\alias{mlb_hit_trajectories}
\title{\strong{MLB Hit Trajectories}}
\usage{
mlb_hit_trajectories()
}
\value{
Returns a tibble with the following columns
|col_name                   |types     |
|:--------------------------|:---------|
|hit_trajectory_code        |character |
|hit_trajectory_description |character |
}
\description{
\strong{MLB Hit Trajectories}
}
\examples{
\donttest{
  mlb_hit_trajectories()
}
}
