% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_homerun_derby.R
\name{mlb_homerun_derby_endpoints}
\alias{mlb_homerun_derby_endpoints}
\alias{homerun_derby}
\alias{hr_derby}
\title{\strong{MLB Home Run Derby Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_homerun_derby()}: Retrieve Homerun Derby data.
\item \code{mlb_homerun_derby_bracket()}: Retrieve Homerun Derby Bracket.
\item \code{mlb_homerun_derby_players()}: Retrieve Homerun Derby Players.
}
}
\details{
\subsection{\strong{MLB Home Run Derby}}{

These functions retrieve MLB Home Run Derby data, bracket, and player information from the MLB Stats API.
}
}
\concept{MLB Home Run Derby}
