% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_job_types.R
\name{mlb_job_types}
\alias{mlb_job_types}
\title{\strong{MLB Job Types}}
\usage{
mlb_job_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   job_code \tab character \tab Four-letter job type code (e.g. 'UMPR', 'UDIR'). \cr
   job \tab character \tab Job title (e.g. 'Umpire', 'Director of Instant Replay'). \cr
   sort_order \tab integer \tab Display sort order for the job type. \cr
}
}
\description{
\strong{MLB Job Types}
}
\examples{
\donttest{
  try(mlb_job_types())
}
}
