% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_jobs.R
\name{mlb_jobs}
\alias{mlb_jobs}
\title{\strong{MLB Jobs}}
\usage{
mlb_jobs(job_type = "UMPR", sport_id = NULL, date = NULL)
}
\arguments{
\item{job_type}{Return information for a given job_type. See \code{mlb_job_types()}}

\item{sport_id}{Return information for a given sport_id.}

\item{date}{Return information for a given date.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   jersey_number \tab character \tab Jersey number worn (often blank for non-uniformed roles). \cr
   job \tab character \tab Job title (e.g. 'Umpire'). \cr
   job_code \tab character \tab Four-letter job type code (e.g. 'UMPR'). \cr
   title \tab character \tab Specific role title for the assignment. \cr
   person_id \tab integer \tab MLB person id for the individual. \cr
   person_full_name \tab character \tab Full name of the individual. \cr
   person_link \tab character \tab API relative link to the person. \cr
}
}
\description{
\strong{MLB Jobs}
}
\examples{
\donttest{
  try(mlb_jobs(job_type='UMPR'))
}
}
