% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_jobs_datacasters.R
\name{mlb_jobs_datacasters}
\alias{mlb_jobs_datacasters}
\title{\strong{MLB Jobs Datacasters}}
\usage{
mlb_jobs_datacasters(sport_id = NULL, date = NULL)
}
\arguments{
\item{sport_id}{Return information for a given sport_id.}

\item{date}{Return information for a given date.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   jersey_number \tab character \tab Jersey number (typically blank for datacasters). \cr
   job \tab character \tab Job title (e.g. 'Stringer'). \cr
   job_code \tab character \tab Four-letter job type code (e.g. 'MSTR'). \cr
   title \tab character \tab Specific role title for the assignment. \cr
   person_id \tab integer \tab MLB person id for the datacaster. \cr
   person_full_name \tab character \tab Full name of the datacaster. \cr
   person_link \tab character \tab API relative link to the person. \cr
}
}
\description{
\strong{MLB Jobs Datacasters}
}
\examples{
\donttest{
  try(mlb_jobs_datacasters(sport_id=1))
}
}
