% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_jobs.R
\name{mlb_jobs_endpoints}
\alias{mlb_jobs_endpoints}
\alias{jobs}
\title{\strong{MLB Jobs Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_jobs()}: MLB Jobs.
\item \code{mlb_jobs_datacasters()}: MLB Jobs Datacasters.
\item \code{mlb_jobs_official_scorers()}: MLB Jobs Official Scorers.
\item \code{mlb_jobs_umpires()}: MLB Jobs Umpires.
\item \code{mlb_job_types()}: MLB Job Types.
}
}
\details{
\subsection{\strong{MLB Jobs}}{

These functions retrieve MLB job, datacaster, official scorer, umpire, and job-type information from the MLB Stats API.
}
}
\concept{MLB Jobs}
