% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_languages.R
\name{mlb_languages}
\alias{mlb_languages}
\title{\strong{MLB API Language Options}}
\usage{
mlb_languages()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   language_id \tab integer \tab Numeric language identifier. \cr
   language_code \tab character \tab Short language code (e.g. 'en', 'xe'). \cr
   language_name \tab character \tab Language display name (e.g. 'English', 'Elias'). \cr
   locale \tab character \tab Locale string for the language (e.g. 'en_US'). \cr
}
}
\description{
\strong{MLB API Language Options}
}
\examples{
\donttest{
  try(mlb_languages())
}
}
