% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_metrics.R
\name{mlb_metrics}
\alias{mlb_metrics}
\title{\strong{MLB Metrics}}
\usage{
mlb_metrics()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   metric_name \tab character \tab Metric name (e.g. 'releaseSpinRate'). \cr
   metric_id \tab integer \tab Numeric metric identifier. \cr
   stat_group \tab character \tab Stat group the metric belongs to (e.g. 'pitching'). \cr
   metric_unit \tab character \tab Unit of measure for the metric (e.g. 'RPM'). \cr
}
}
\description{
\strong{MLB Metrics}
}
\examples{
\donttest{
  try(mlb_metrics())
}
}
