% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pbp_context_metrics.R
\name{mlb_pbp_context_metrics}
\alias{mlb_pbp_context_metrics}
\title{\strong{Acquire game context metrics for Major and Minor League games}}
\usage{
mlb_pbp_context_metrics(game_pk, timecode)
}
\arguments{
\item{game_pk}{The game_pk for the game requested}

\item{timecode}{The time code for the MLB game (format: MMDDYYYY_HHMMSS)}
}
\value{
Returns a data frame that includes time codes from the game_pk requested\tabular{ll}{
   col_name \tab types \cr
   game_pk \tab integer \cr
   link \tab character \cr
   game_type \tab character \cr
   season \tab character \cr
   game_date \tab character \cr
   official_date \tab character \cr
   status_abstract_game_state \tab character \cr
   status_coded_game_state \tab character \cr
   status_detailed_state \tab character \cr
   status_status_code \tab character \cr
   status_start_time_tbd \tab logical \cr
   status_abstract_game_code \tab character \cr
   teams_away_league_record_wins \tab integer \cr
   teams_away_league_record_losses \tab integer \cr
   teams_away_league_record_pct \tab character \cr
   teams_away_score \tab integer \cr
   teams_away_team_id \tab integer \cr
   teams_away_team_name \tab character \cr
   teams_away_team_link \tab character \cr
   teams_away_is_winner \tab logical \cr
   teams_away_probable_pitcher_id \tab integer \cr
   teams_away_probable_pitcher_full_name \tab character \cr
   teams_away_probable_pitcher_link \tab character \cr
   teams_away_split_squad \tab logical \cr
   teams_away_series_number \tab integer \cr
   teams_home_league_record_wins \tab integer \cr
   teams_home_league_record_losses \tab integer \cr
   teams_home_league_record_pct \tab character \cr
   teams_home_score \tab integer \cr
   teams_home_team_id \tab integer \cr
   teams_home_team_name \tab character \cr
   teams_home_team_link \tab character \cr
   teams_home_is_winner \tab logical \cr
   teams_home_probable_pitcher_id \tab integer \cr
   teams_home_probable_pitcher_full_name \tab character \cr
   teams_home_probable_pitcher_link \tab character \cr
   teams_home_split_squad \tab logical \cr
   teams_home_series_number \tab integer \cr
   venue_id \tab integer \cr
   venue_name \tab character \cr
   venue_link \tab character \cr
   link_1 \tab character \cr
   is_tie \tab logical \cr
   game_number \tab integer \cr
   public_facing \tab logical \cr
   double_header \tab character \cr
   gameday_type \tab character \cr
   tiebreaker \tab character \cr
   calendar_event_id \tab character \cr
   season_display \tab character \cr
   day_night \tab character \cr
   scheduled_innings \tab integer \cr
   reverse_home_away_status \tab logical \cr
   inning_break_length \tab integer \cr
   games_in_series \tab integer \cr
   series_game_number \tab integer \cr
   series_description \tab character \cr
   record_source \tab character \cr
   if_necessary \tab character \cr
   if_necessary_description \tab character \cr
   game_id \tab character \cr
   home_win_probability \tab numeric \cr
   away_win_probability \tab numeric \cr
}
}
\description{
\strong{Acquire game context metrics for Major and Minor League games}
}
\examples{
\donttest{
  mlb_pbp_context_metrics(game_pk = 531060, timecode = "20180803_182458")
}
}
