% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pbp_timecodes.R
\name{mlb_pbp_timecodes}
\alias{mlb_pbp_timecodes}
\title{\strong{Acquire time codes for Major and Minor League games}}
\usage{
mlb_pbp_timecodes(game_pk)
}
\arguments{
\item{game_pk}{The game_pk for the game requested}
}
\value{
Returns a data frame that includes time codes from the game_pk requested\tabular{ll}{
   col_name \tab types \cr
   timecodes  (MMDDYYYY_HHMMSS) \tab numeric \cr
}
}
\description{
\strong{Acquire time codes for Major and Minor League games}
}
\examples{
\donttest{
  mlb_pbp_timecodes(game_pk = 632970)
}
}
