% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pitch_types.R
\name{mlb_pitch_types}
\alias{mlb_pitch_types}
\title{\strong{MLB Pitch Types}}
\usage{
mlb_pitch_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   pitch_type_code \tab character \tab Short code identifying the pitch type (e.g. 'FA'). \cr
   pitch_type_description \tab character \tab Full name of the pitch type (e.g. 'Fastball'). \cr
}
}
\description{
\strong{MLB Pitch Types}
}
\examples{
\donttest{
  try(mlb_pitch_types())
}
}
