% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_player_status_codes.R
\name{mlb_player_status_codes}
\alias{mlb_player_status_codes}
\title{\strong{MLB Player Status Codes}}
\usage{
mlb_player_status_codes()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   player_status_code \tab character \tab Short code for the player status (e.g. 'A'). \cr
   player_status_description \tab character \tab Description of the player status (e.g. 'Active'). \cr
}
}
\description{
\strong{MLB Player Status Codes}
}
\examples{
\donttest{
  try(mlb_player_status_codes())
}
}
