% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_probables.R
\name{mlb_probables}
\alias{mlb_probables}
\title{\strong{Retrieve probable starters for a given MLB game}}
\usage{
mlb_probables(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a tiible that includes probable starting pitchers and
the home plate umpire for the \code{game_pk} requested including the following columns:\tabular{lll}{
   col_name \tab types \tab description \cr
   game_pk \tab integer \tab Unique game identifier. \cr
   game_date \tab character \tab Date of the game (YYYY-MM-DD). \cr
   fullName \tab character \tab Full name of the probable starting pitcher. \cr
   id \tab integer \tab MLBAM player ID of the probable starting pitcher. \cr
   team \tab character \tab Team name for the probable pitcher. \cr
   team_id \tab integer \tab MLBAM team ID for the probable pitcher. \cr
   home_plate_full_name \tab character \tab Full name of the home plate umpire. \cr
   home_plate_id \tab integer \tab MLBAM ID of the home plate umpire. \cr
}
}
\description{
\strong{Retrieve probable starters for a given MLB game}
}
\examples{
\donttest{
  try(mlb_probables(566001))
}
}
