% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_review_reasons.R
\name{mlb_review_reasons}
\alias{mlb_review_reasons}
\title{\strong{MLB Review Reasons}}
\usage{
mlb_review_reasons()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   review_reason_code \tab character \tab Short code for the replay review reason. \cr
   review_reason_description \tab character \tab Description of the review reason (e.g. 'Tag play'). \cr
}
}
\description{
\strong{MLB Review Reasons}
}
\examples{
\donttest{
  try(mlb_review_reasons())
}
}
