% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_runner_detail_types.R
\name{mlb_runner_detail_types}
\alias{mlb_runner_detail_types}
\title{\strong{MLB Runner Detail Types}}
\usage{
mlb_runner_detail_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   stat_name \tab character \tab Name of the runner detail stat (e.g. 'r_force_out'). \cr
}
}
\description{
\strong{MLB Runner Detail Types}
}
\examples{
\donttest{
  try(mlb_runner_detail_types())
}
}
