% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_schedule.R
\name{mlb_schedule_endpoints}
\alias{mlb_schedule_endpoints}
\alias{schedule}
\alias{game_pk}
\title{\strong{MLB Schedule Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_schedule()}: Find game_pk values for professional baseball games (major and minor leagues).
\item \code{mlb_schedule_postseason()}: Find game_pk values for professional baseball postseason games (major and minor leagues).
\item \code{mlb_schedule_postseason_series()}: Find game_pk values for professional baseball postseason series games (major and minor leagues).
\item \code{mlb_schedule_games_tied()}: Find game_pk values for professional baseball games (major and minor leagues) that are tied.
\item \code{mlb_schedule_event_types()}: MLB Schedule Event Types.
}
}
\details{
\subsection{\strong{MLB Schedule}}{

These functions retrieve MLB schedule, postseason, tied-game, and event-type information from the MLB Stats API.
}
}
\concept{MLB Schedule}
