% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_schedule_postseason_series.R
\name{mlb_schedule_postseason_series}
\alias{mlb_schedule_postseason_series}
\title{\strong{Find game_pk values for professional baseball postseason series games (major and minor leagues)}}
\usage{
mlb_schedule_postseason_series(
  season = 2021,
  game_type = NULL,
  series_number = NULL,
  sport_id = 1,
  team_id = NULL
)
}
\arguments{
\item{season}{The season for which you want to find game_pk values for MLB games}

\item{game_type}{game_type to return schedule information for all tied games in a particular game_type}

\item{series_number}{The Series number to return schedule information for all tied games in a particular series number}

\item{sport_id}{The sport_id to return schedule information for.}

\item{team_id}{The team_id to return schedule information for.\tabular{ll}{
   game_type_id \tab game_type_description \cr
   S \tab Spring Training \cr
   R \tab Regular Season \cr
   F \tab Wild Card Game \cr
   D \tab Division Series \cr
   L \tab League Championship Series \cr
   W \tab World Series \cr
   C \tab Championship \cr
   N \tab Nineteenth Century Series \cr
   P \tab Playoffs \cr
   A \tab All-Star Game \cr
   I \tab Intrasquad \cr
   E \tab Exhibition \cr
}}
}
\value{
Returns a data frame that includes game_pk values and additional
information for games scheduled or played
|col_name                        |types     |
|:-------------------------------|:---------|
|total_items                     |integer   |
|total_games                     |integer   |
|total_games_in_progress         |integer   |
|game_pk                         |integer   |
|link                            |character |
|game_type                       |character |
|season                          |character |
|game_date                       |character |
|official_date                   |character |
|is_tie                          |logical   |
|is_featured_game                |logical   |
|game_number                     |integer   |
|public_facing                   |logical   |
|double_header                   |character |
|gameday_type                    |character |
|tiebreaker                      |character |
|calendar_event_id               |character |
|season_display                  |character |
|day_night                       |character |
|description                     |character |
|scheduled_innings               |integer   |
|reverse_home_away_status        |logical   |
|inning_break_length             |integer   |
|games_in_series                 |integer   |
|series_game_number              |integer   |
|series_description              |character |
|record_source                   |character |
|if_necessary                    |character |
|if_necessary_description        |character |
|is_default_game                 |logical   |
|status_abstract_game_state      |character |
|status_coded_game_state         |character |
|status_detailed_state           |character |
|status_status_code              |character |
|status_start_time_tbd           |logical   |
|status_abstract_game_code       |character |
|teams_away_score                |integer   |
|teams_away_is_winner            |logical   |
|teams_away_split_squad          |logical   |
|teams_away_series_number        |integer   |
|teams_away_league_record_wins   |integer   |
|teams_away_league_record_losses |integer   |
|teams_away_league_record_pct    |character |
|teams_away_team_id              |integer   |
|teams_away_team_name            |character |
|teams_away_team_link            |character |
|teams_home_score                |integer   |
|teams_home_is_winner            |logical   |
|teams_home_split_squad          |logical   |
|teams_home_series_number        |integer   |
|teams_home_league_record_wins   |integer   |
|teams_home_league_record_losses |integer   |
|teams_home_league_record_pct    |character |
|teams_home_team_id              |integer   |
|teams_home_team_name            |character |
|teams_home_team_link            |character |
|venue_id                        |integer   |
|venue_name                      |character |
|venue_link                      |character |
|content_link                    |character |
|reschedule_date                 |character |
|reschedule_game_date            |character |
|rescheduled_from                |character |
|rescheduled_from_date           |character |
|status_reason                   |character |
|sort_order                      |integer   |
|series_id                       |character |
|series_sort_number              |integer   |
|series_is_default               |logical   |
|series_game_type                |character |
}
\description{
\strong{Find game_pk values for professional baseball postseason series games (major and minor leagues)}
}
\examples{
\donttest{
  mlb_schedule_postseason_series(season = 2021, sport_id = 1)
}
}
