% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_seasons.R
\name{mlb_seasons}
\alias{mlb_seasons}
\title{\strong{Find MLB Seasons}}
\usage{
mlb_seasons(sport_id = 1, with_game_type_dates = TRUE)
}
\arguments{
\item{sport_id}{The sport_id to return season information for.}

\item{with_game_type_dates}{with_game_type_dates to return season information}
}
\value{
Returns a tibble with the following columns:\tabular{ll}{
   col_name \tab types \cr
   season_id \tab character \cr
   has_wildcard \tab logical \cr
   pre_season_start_date \tab character \cr
   pre_season_end_date \tab character \cr
   season_start_date \tab character \cr
   spring_start_date \tab character \cr
   spring_end_date \tab character \cr
   regular_season_start_date \tab character \cr
   last_date1st_half \tab character \cr
   all_star_date \tab character \cr
   first_date2nd_half \tab character \cr
   regular_season_end_date \tab character \cr
   post_season_start_date \tab character \cr
   post_season_end_date \tab character \cr
   season_end_date \tab character \cr
   offseason_start_date \tab character \cr
   off_season_end_date \tab character \cr
   season_level_gameday_type \tab character \cr
   game_level_gameday_type \tab character \cr
   qualifier_plate_appearances \tab numeric \cr
   qualifier_outs_pitched \tab integer \cr
}
}
\description{
\strong{Find MLB Seasons}
}
\examples{
\donttest{
 mlb_seasons(sport_id = 1)
}
}
