% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_seasons_all.R
\name{mlb_seasons_all}
\alias{mlb_seasons_all}
\title{\strong{Find MLB Seasons all}}
\usage{
mlb_seasons_all(
  sport_id = 1,
  division_id = NULL,
  league_id = NULL,
  with_game_type_dates = TRUE
)
}
\arguments{
\item{sport_id}{The sport_id to return season information for.}

\item{division_id}{The division_id to return season information for.}

\item{league_id}{The league_id to return season information for.}

\item{with_game_type_dates}{with_game_type_dates to return season information for.}
}
\value{
Returns a tibble with the following columns:\tabular{lll}{
   col_name \tab types \tab description \cr
   season_id \tab character \tab Season year identifier. \cr
   has_wildcard \tab logical \tab Whether the season has a wild card round. \cr
   pre_season_start_date \tab character \tab Pre-season start date. \cr
   season_start_date \tab character \tab Season start date. \cr
   regular_season_start_date \tab character \tab Regular season start date. \cr
   regular_season_end_date \tab character \tab Regular season end date. \cr
   season_end_date \tab character \tab Season end date. \cr
   offseason_start_date \tab character \tab Off-season start date. \cr
   off_season_end_date \tab character \tab Off-season end date. \cr
   season_level_gameday_type \tab character \tab Season-level Gameday data feed type. \cr
   game_level_gameday_type \tab character \tab Game-level Gameday data feed type. \cr
   qualifier_plate_appearances \tab numeric \tab Plate appearances per team game to qualify. \cr
   qualifier_outs_pitched \tab integer \tab Outs pitched per team game to qualify. \cr
   post_season_start_date \tab character \tab Post-season start date. \cr
   post_season_end_date \tab character \tab Post-season end date. \cr
   last_date1st_half \tab character \tab Last date of the first half. \cr
   all_star_date \tab character \tab All-Star Game date. \cr
   first_date2nd_half \tab character \tab First date of the second half. \cr
   pre_season_end_date \tab character \tab Pre-season end date. \cr
   spring_start_date \tab character \tab Spring training start date. \cr
   spring_end_date \tab character \tab Spring training end date. \cr
}
}
\description{
\strong{Find MLB Seasons all}
}
\examples{
\donttest{
 try(mlb_seasons_all(sport_id = 1))
}
}
