% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_situation_codes.R
\name{mlb_situation_codes}
\alias{mlb_situation_codes}
\title{\strong{MLB Situation Codes}}
\usage{
mlb_situation_codes()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   situation_code \tab character \tab Code identifying the game situation. \cr
   sort_order \tab integer \tab Display sort order for the situation code. \cr
   navigation_menu \tab character \tab Navigation menu grouping (e.g. 'Game'). \cr
   situation_code_description \tab character \tab Description of the situation (e.g. 'Home Games'). \cr
   team \tab logical \tab Whether the situation applies to team stats. \cr
   batting \tab logical \tab Whether the situation applies to batting stats. \cr
   fielding \tab logical \tab Whether the situation applies to fielding stats. \cr
   pitching \tab logical \tab Whether the situation applies to pitching stats. \cr
}
}
\description{
\strong{MLB Situation Codes}
}
\examples{
\donttest{
  try(mlb_situation_codes())
}
}
