% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_sky.R
\name{mlb_sky}
\alias{mlb_sky}
\title{\strong{MLB Sky (Weather) Codes}}
\usage{
mlb_sky()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   sky_code \tab character \tab Code for the sky/weather condition. \cr
   sky_description \tab character \tab Description of the sky condition (e.g. 'Clear'). \cr
}
}
\description{
\strong{MLB Sky (Weather) Codes}
}
\examples{
\donttest{
  try(mlb_sky())
}
}
