% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_standings_types.R
\name{mlb_standings_types}
\alias{mlb_standings_types}
\title{\strong{MLB Standings Types}}
\usage{
mlb_standings_types()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   standings_type_name \tab character \tab Standings type identifier (e.g., wildCard). \cr
   standings_type_description \tab character \tab Human-readable description of the type. \cr
}
}
\description{
\strong{MLB Standings Types}
}
\examples{
\donttest{
  try(mlb_standings_types())
}
}
