% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_stat_groups.R
\name{mlb_stat_groups}
\alias{mlb_stat_groups}
\title{\strong{MLB Stat Groups}}
\usage{
mlb_stat_groups()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   stat_group_name \tab character \tab Stat group display name (e.g., hitting). \cr
}
}
\description{
\strong{MLB Stat Groups}
}
\examples{
\donttest{
  try(mlb_stat_groups())
}
}
