% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_stats.R
\name{mlb_stats_endpoints}
\alias{mlb_stats_endpoints}
\alias{stats}
\title{\strong{MLB Stats Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_stats()}: MLB Stats.
\item \code{mlb_stats_leaders()}: MLB Stats Leaders.
\item \code{mlb_stats_streaks()}: MLB Stats Streaks.
\item \code{mlb_stat_groups()}: MLB Stat Groups.
\item \code{mlb_stat_types()}: MLB Stat Types.
\item \code{mlb_baseball_stats()}: MLB Baseball Stats.
}
}
\details{
\subsection{\strong{MLB Stats}}{

These functions retrieve MLB statistics, leaders, streaks, stat groups, stat types, and baseball stat metadata from the MLB Stats API.
}
}
\concept{MLB Stats}
