% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_stats_leaders.R
\name{mlb_stats_leaders}
\alias{mlb_stats_leaders}
\title{\strong{MLB Stats Leaders}}
\usage{
mlb_stats_leaders(
  leader_categories = NULL,
  player_pool = NULL,
  leader_game_types = NULL,
  sit_codes = NULL,
  position = NULL,
  stat_group = NULL,
  season = NULL,
  league_id = NULL,
  sport_id = NULL,
  start_date = NULL,
  end_date = NULL,
  stat_type = NULL,
  limit = 1000
)
}
\arguments{
\item{leader_categories}{League leader category to return information and ranking for a particular statistic.}

\item{player_pool}{There are 4 different types of player pools to return statistics for a particular player pool across a sport.
Acceptable values include: All, Qualified, Rookies, or Qualified_rookies}

\item{leader_game_types}{Game type to return information and ranking for a particular statistic in a particular game type.}

\item{sit_codes}{Situation code to return information and ranking for a particular statistic in a particular game type.}

\item{position}{Position to return statistics for a given position. Default to "Qualified" player pool
Acceptable values include:
\itemize{
\item P
\item C
\item 1B
\item 2B
\item 3B
\item SS
\item LF
\item CF
\item RF
\item DH
\item PH
\item PR
\item BR
\item OF
\item IF
\item SP
\item RP
\item CP
\item UT
\item UI
\item UO
\item RHP
\item LHP
\item RHS
\item LHS
\item LHR
\item RHR
\item B
\item X
}}

\item{stat_group}{Stat group to return information and ranking for a particular statistic in a particular group.}

\item{season}{Year to return information and ranking for a particular statistic in a given year.}

\item{league_id}{League ID to return statistics for a given league. Default to "Qualified" player pool.}

\item{sport_id}{The sport_id to return information and ranking information for.}

\item{start_date}{Start date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY
\emph{start_date must be coupled with end_date and byDateRange stat_type}}

\item{end_date}{End date to return information and ranking for a particular statistic for a particular date range. Format: MM/DD/YYYY
\emph{end_date must be coupled with start_date and byDateRange stat_type}}

\item{stat_type}{The stat_type to return information and ranking for a particular statistic for a particular stat type.}

\item{limit}{A limit to limit return to a particular number of records.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   leader_category \tab character \tab Statistical leader category (e.g., homeRuns). \cr
   rank \tab integer \tab Rank within the leaderboard. \cr
   value \tab character \tab Statistic value for the leader. \cr
   season \tab character \tab Season year. \cr
   num_teams \tab integer \tab Number of teams the player appeared for. \cr
   team_id \tab integer \tab Team MLBAM ID. \cr
   team_name \tab character \tab Team name. \cr
   team_link \tab character \tab API link to the team. \cr
   league_id \tab integer \tab League MLBAM ID. \cr
   league_name \tab character \tab League name. \cr
   league_link \tab character \tab API link to the league. \cr
   person_id \tab integer \tab Player MLBAM ID. \cr
   person_full_name \tab character \tab Player full name. \cr
   person_link \tab character \tab API link to the player. \cr
   person_first_name \tab character \tab Player first name. \cr
   person_last_name \tab character \tab Player last name. \cr
   sport_id \tab integer \tab Sport MLBAM ID. \cr
   sport_link \tab character \tab API link to the sport. \cr
   sport_abbreviation \tab character \tab Sport abbreviation (e.g., MLB). \cr
   stat_group \tab character \tab Stat group (e.g., hitting). \cr
   total_splits \tab integer \tab Total number of splits in the leaderboard. \cr
   game_type_id \tab character \tab Game type code (e.g., R for regular season). \cr
   game_type_description \tab character \tab Game type description. \cr
}
}
\description{
\strong{MLB Stats Leaders}
}
\examples{
\donttest{
 try(mlb_stats_leaders(leader_categories='homeRuns',sport_id=1, season = 2021))
}
}
