% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_team_coaches.R
\name{mlb_team_coaches}
\alias{mlb_team_coaches}
\title{\strong{MLB Team Coaches}}
\usage{
mlb_team_coaches(team_id = NULL, date = NULL, season = NULL)
}
\arguments{
\item{team_id}{Team ID to return team coach information for.}

\item{date}{Date to return team coach information for.}

\item{season}{Year to return team coach information for.}
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   jersey_number \tab character \cr
   job \tab character \cr
   job_id \tab character \cr
   title \tab character \cr
   person_id \tab integer \cr
   person_full_name \tab character \cr
   person_link \tab character \cr
}
}
\description{
\strong{MLB Team Coaches}
}
\examples{
\donttest{
  try(mlb_team_coaches(team_id = 137, season = 2021))
}
}
