% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_teams.R
\name{mlb_team_endpoints}
\alias{mlb_team_endpoints}
\alias{teams}
\alias{team}
\title{\strong{MLB Teams Endpoint Overview}}
\description{
\itemize{
\item \code{mlb_teams()}: MLB Teams.
\item \code{mlb_team_info()}: MLB Team Info.
\item \code{mlb_team_affiliates()}: MLB Team Affiliates.
\item \code{mlb_team_alumni()}: MLB Team Alumni.
\item \code{mlb_team_coaches()}: MLB Team Coaches.
\item \code{mlb_team_history()}: MLB Teams History.
\item \code{mlb_team_leaders()}: MLB Team Leaders.
\item \code{mlb_team_personnel()}: MLB Team Personnel.
\item \code{mlb_team_stats()}: MLB Team Individual Stats.
\item \code{mlb_teams_stats()}: MLB Teams Stats.
\item \code{mlb_teams_stats_leaders()}: MLB Teams Stats Leaders.
}
}
\details{
\subsection{\strong{MLB Teams}}{

These functions retrieve MLB team, affiliate, alumni, coach, history, leader, personnel, and stats information from the MLB Stats API.
}
}
\concept{MLB Teams}
