% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_team_leaders.R
\name{mlb_team_leaders}
\alias{mlb_team_leaders}
\title{\strong{MLB Team Leaders}}
\usage{
mlb_team_leaders(
  team_id = NULL,
  leader_categories = NULL,
  leader_game_types = NULL,
  season = NULL,
  limit = 1000
)
}
\arguments{
\item{team_id}{Team ID to return team leader information for.}

\item{leader_categories}{Team leader category to return information and ranking for a particular statistic.}

\item{leader_game_types}{Game type to return information and ranking for a particular statistic in a particular game type.}

\item{season}{Season to return team leader information for.}

\item{limit}{A limit to limit return to a particular number of records.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   leader_category \tab character \tab Team leader category (e.g., homeRuns). \cr
   rank \tab integer \tab Rank within the team leaderboard. \cr
   value \tab character \tab Statistic value for the leader. \cr
   season \tab character \tab Season year. \cr
   team_id \tab integer \tab Team MLBAM ID. \cr
   team_name \tab character \tab Team name. \cr
   team_link \tab character \tab API link to the team. \cr
   league_id \tab integer \tab League MLBAM ID. \cr
   league_name \tab character \tab League name. \cr
   league_link \tab character \tab API link to the league. \cr
   person_id \tab integer \tab Player MLBAM ID. \cr
   person_full_name \tab character \tab Player full name. \cr
   person_link \tab character \tab API link to the player. \cr
   person_first_name \tab character \tab Player first name. \cr
   person_last_name \tab character \tab Player last name. \cr
   sport_id \tab integer \tab Sport MLBAM ID. \cr
   sport_link \tab character \tab API link to the sport. \cr
   sport_abbreviation \tab character \tab Sport abbreviation (e.g., MLB). \cr
   stat_group \tab character \tab Stat group (e.g., hitting). \cr
   total_splits \tab integer \tab Total number of splits in the leaderboard. \cr
   game_type_id \tab character \tab Game type code (e.g., R for regular season). \cr
   game_type_description \tab character \tab Game type description. \cr
}
}
\description{
\strong{MLB Team Leaders}
}
\examples{
\donttest{
  try(mlb_team_leaders(team_id = 137, leader_categories = "homeRuns", season = 2021))
}
}
