% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_team_personnel.R
\name{mlb_team_personnel}
\alias{mlb_team_personnel}
\title{\strong{MLB Team Personnel}}
\usage{
mlb_team_personnel(team_id = NULL, date = NULL)
}
\arguments{
\item{team_id}{Team ID to return team coach information for.}

\item{date}{Date to return team coach information for.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   jersey_number \tab character \tab Personnel uniform number. \cr
   job \tab character \tab Job name (e.g., Special Assistant). \cr
   job_id \tab character \tab Job code identifier. \cr
   title \tab character \tab Full personnel title. \cr
   person_id \tab integer \tab Personnel MLBAM person ID. \cr
   person_full_name \tab character \tab Personnel full name. \cr
   person_link \tab character \tab API link to the person. \cr
}
}
\description{
\strong{MLB Team Personnel}
}
\examples{
\donttest{
  try(mlb_team_personnel(team_id = 137, date = "08/28/2016"))
}
}
