% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_teams.R
\name{mlb_teams}
\alias{mlb_teams}
\title{\strong{MLB Teams}}
\usage{
mlb_teams(
  season = NULL,
  active_status = NULL,
  all_star_statuses = NULL,
  league_ids = NULL,
  sport_ids = NULL,
  game_type = NULL
)
}
\arguments{
\item{season}{Year to return to return team information for.}

\item{active_status}{The active statuses to populate teams for a given season.}

\item{all_star_statuses}{The all-star statuses to populate teams for a given season.}

\item{league_ids}{The league_id(s) to return team information for.}

\item{sport_ids}{The sport_id(s) to return team information for.}

\item{game_type}{The game_type to return team information for.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   all_star_status \tab character \tab All-star status flag. \cr
   team_id \tab integer \tab Team MLBAM ID. \cr
   team_full_name \tab character \tab Full team name. \cr
   link \tab character \tab API link to the team. \cr
   season \tab integer \tab Season year. \cr
   team_code \tab character \tab Internal team code. \cr
   file_code \tab character \tab File code abbreviation. \cr
   team_abbreviation \tab character \tab Team abbreviation. \cr
   team_name \tab character \tab Short team name. \cr
   location_name \tab character \tab Team location (city). \cr
   first_year_of_play \tab character \tab First year the franchise played. \cr
   short_name \tab character \tab Short display name. \cr
   franchise_name \tab character \tab Franchise name. \cr
   club_name \tab character \tab Club name. \cr
   active \tab logical \tab Whether the team is active. \cr
   spring_league_id \tab integer \tab Spring league MLBAM ID. \cr
   spring_league_name \tab character \tab Spring league name. \cr
   spring_league_link \tab character \tab API link to the spring league. \cr
   spring_league_abbreviation \tab character \tab Spring league abbreviation. \cr
   venue_id \tab integer \tab Home venue MLBAM ID. \cr
   venue_name \tab character \tab Home venue name. \cr
   venue_link \tab character \tab API link to the venue. \cr
   spring_venue_id \tab integer \tab Spring training venue MLBAM ID. \cr
   spring_venue_link \tab character \tab API link to the spring venue. \cr
   league_id \tab integer \tab League MLBAM ID. \cr
   league_name \tab character \tab League name. \cr
   league_link \tab character \tab API link to the league. \cr
   division_id \tab integer \tab Division MLBAM ID. \cr
   division_name \tab character \tab Division name. \cr
   division_link \tab character \tab API link to the division. \cr
   sport_id \tab integer \tab Sport MLBAM ID. \cr
   sport_link \tab character \tab API link to the sport. \cr
   sport_name \tab character \tab Sport name (e.g., Major League Baseball). \cr
}
}
\description{
\strong{MLB Teams}
}
\examples{
\donttest{
  try(mlb_teams(season = 2021, sport_ids = c(1)))
}
}
