% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_umpire_games.R
\name{mlb_umpire_games}
\alias{mlb_umpire_games}
\title{\strong{MLB Umpire Games}}
\usage{
mlb_umpire_games(umpire_id = NULL, season = 2021)
}
\arguments{
\item{umpire_id}{Return information for a given umpire_id.}

\item{season}{Return information for a given season.}
}
\value{
Returns a tibble with the following columns
|col_name                        |types     |
|:-------------------------------|:---------|
|date                            |character |
|total_items                     |integer   |
|total_events                    |integer   |
|total_games                     |integer   |
|total_games_in_progress         |integer   |
|game_pk                         |integer   |
|link                            |character |
|game_type                       |character |
|season                          |character |
|game_date                       |character |
|official_date                   |character |
|is_tie                          |logical   |
|game_number                     |integer   |
|public_facing                   |logical   |
|double_header                   |character |
|gameday_type                    |character |
|tiebreaker                      |character |
|calendar_event_id               |character |
|season_display                  |character |
|day_night                       |character |
|scheduled_innings               |integer   |
|reverse_home_away_status        |logical   |
|games_in_series                 |integer   |
|series_game_number              |integer   |
|series_description              |character |
|record_source                   |character |
|if_necessary                    |character |
|if_necessary_description        |character |
|status_abstract_game_state      |character |
|status_coded_game_state         |character |
|status_detailed_state           |character |
|status_status_code              |character |
|status_start_time_tbd           |logical   |
|status_abstract_game_code       |character |
|teams_away_score                |integer   |
|teams_away_is_winner            |logical   |
|teams_away_split_squad          |logical   |
|teams_away_series_number        |integer   |
|teams_away_league_record_wins   |integer   |
|teams_away_league_record_losses |integer   |
|teams_away_league_record_pct    |character |
|teams_away_team_id              |integer   |
|teams_away_team_name            |character |
|teams_away_team_link            |character |
|teams_home_score                |integer   |
|teams_home_is_winner            |logical   |
|teams_home_split_squad          |logical   |
|teams_home_series_number        |integer   |
|teams_home_league_record_wins   |integer   |
|teams_home_league_record_losses |integer   |
|teams_home_league_record_pct    |character |
|teams_home_team_id              |integer   |
|teams_home_team_name            |character |
|teams_home_team_link            |character |
|venue_id                        |integer   |
|venue_name                      |character |
|venue_link                      |character |
|content_link                    |character |
|inning_break_length             |integer   |
|status_reason                   |character |
|description                     |character |
|reschedule_date                 |character |
|reschedule_game_date            |character |
|rescheduled_from                |character |
|rescheduled_from_date           |character |
}
\description{
\strong{MLB Umpire Games}
}
\examples{
\donttest{
  mlb_umpire_games(umpire_id = 596809, season = 2021)
}
}
