% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_venues.R
\name{mlb_venues}
\alias{mlb_venues}
\title{\strong{Find MLB Venues}}
\usage{
mlb_venues(venue_ids = NULL, sport_ids = NULL, season = NULL)
}
\arguments{
\item{venue_ids}{Venue directorial information based venue_id.}

\item{sport_ids}{The sport_id(s) for which to return venue directorial information.}

\item{season}{Year for which to return venue directorial information for a given season.}
}
\value{
Returns a tibble with the following columns:\tabular{lll}{
   col_name \tab types \tab description \cr
   venue_id \tab integer \tab Venue MLBAM ID. \cr
   venue_name \tab character \tab Venue name. \cr
   venue_link \tab character \tab API link to the venue. \cr
   active \tab logical \tab Whether the venue is active. \cr
   season \tab character \tab Season the venue record applies to. \cr
}
}
\description{
\strong{Find MLB Venues}
}
\examples{
\donttest{
  try(mlb_venues())
  try(mlb_venues(venue_ids = 4781))
  try(mlb_venues(sport_ids = 1))
}
}
