% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_wind_direction_codes.R
\name{mlb_wind_direction_codes}
\alias{mlb_wind_direction_codes}
\title{\strong{MLB Wind Direction Codes}}
\usage{
mlb_wind_direction_codes()
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   wind_direction_code \tab character \tab Wind direction code (e.g., In From CF). \cr
   wind_direction_description \tab character \tab Wind direction description. \cr
}
}
\description{
\strong{MLB Wind Direction Codes}
}
\examples{
\donttest{
  try(mlb_wind_direction_codes())
}
}
