% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{most_recent_mlb_season}
\alias{most_recent_mlb_season}
\title{\strong{Most Recent MLB Season}}
\usage{
most_recent_mlb_season()
}
\value{
An integer indicating the year of the most recent season of Major League Baseball
}
\description{
\strong{Most Recent MLB Season}
}
