% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{most_recent_ncaa_baseball_season}
\alias{most_recent_ncaa_baseball_season}
\title{\strong{Most Recent NCAA Baseball Season}}
\usage{
most_recent_ncaa_baseball_season()
}
\description{
\strong{Most Recent NCAA Baseball Season}
}
