% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_expected_runs_matrix.R
\name{ncaa_get_expected_runs_matrix}
\alias{ncaa_get_expected_runs_matrix}
\title{Calculate expected runs matrix for NCAA PBP}
\usage{
ncaa_get_expected_runs_matrix(pbp_data_frame)
}
\arguments{
\item{season}{True or False, NCAA PBP by Season or Individual Teams.}
}
\description{
NCAA PBP can be obtained from `ncaa_get_pbp`
}
\examples{
\dontrun{
tamu_pbp_df <- ncaa_get_pbp(season=F,year=2017,division=1,teamid=697)
ncaa_get_expected_runs_matrix(tamu_pbp_df)
}
}
\keyword{NCAA,}
\keyword{baseball,}
\keyword{college}
