% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_pbp.R
\name{ncaa_get_pbp}
\alias{ncaa_get_pbp}
\title{Scrape and clean NCAA baseball play by play (Division I, II, and III)}
\usage{
ncaa_get_pbp(season = F, year, division = 1, teamid = NULL,
  conference = NULL)
}
\arguments{
\item{season}{True or False, NCAA PBP by Season or Individual Teams.}

\item{division}{Select what division of college baseball}

\item{teamid}{The numerical ID that the NCAA website uses to identify a team}

\item{conf}{Select a conference, naming convention can be found in master_ncaa_team_lu}
}
\description{
This function allows the user to obtain cleaned PBP data for individal teams, a conference, or all teams for that season
}
\examples{
\dontrun{
ncaa_get_pbp(season=F,year=2017,division=1,teamid=697)
ncaa_get_pbp(season=T,year=2017,division=1,conference="SEC")
}


}
\keyword{NCAA,}
\keyword{baseball,}
\keyword{college}
