% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_season_schedule.R
\name{ncaa_get_season_schedule}
\alias{ncaa_get_season_schedule}
\title{Scrape NCAA baseball schedule for all teams (or just one conference) (Division I, II, and III)}
\usage{
ncaa_get_season_schedule(yr, div = 1, conf = NULL)
}
\arguments{
\item{conf}{Select a conference, naming convention can be found in master_ncaa_team_lu}

\item{year}{The season for which data should be returned, in the form of "YYYY". Years currently available: 2013-2017.}

\item{division}{Select what division of college baseball}
}
\description{
This function allows the user to obtain multiple team's schedule.
}
\examples{
\dontrun{
ncaa_get_season_schedule(2017)
ncaa_get_season_schedule(2017,conf="ACC")
}


}
\keyword{NCAA,}
\keyword{baseball,}
\keyword{college}
