% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_get_team_schedule.R
\name{ncaa_get_team_schedule}
\alias{ncaa_get_team_schedule}
\title{Scrape NCAA baseball schedule for individual teams (Division I, II, and III)}
\usage{
ncaa_get_team_schedule(teamid, year, division = 1, verbose = F)
}
\arguments{
\item{teamid}{The numerical ID that the NCAA website uses to identify a team}

\item{year}{The season for which data should be returned, in the form of "YYYY". Years currently available: 2013-2017.}

\item{division}{Select what division of college baseball}
}
\description{
This function allows the user to obtain an individual team's schedule.
}
\examples{
\dontrun{
ncaa_get_team_schedule(697, 2017)
}




}
\keyword{NCAA,}
\keyword{baseball,}
\keyword{college}
