% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_park_factor.R
\name{ncaa_park_factor}
\alias{ncaa_park_factor}
\title{\strong{Get Park Effects for NCAA Baseball Teams}}
\usage{
ncaa_park_factor(team_id, years, type = "conference", ...)
}
\arguments{
\item{team_id}{The team's unique NCAA id.}

\item{years}{The season or seasons (i.e. use 2016 for the 2015-2016 season,
etc., limited to just 2013-2023 seasons).}

\item{type}{default is conference. the conference parameter adjusts for the conference
the school plays in, the division parameter calculates based on the division the school plays in 1,2,or 3.
Defaults to 'conference'.}

\item{...}{Additional arguments passed to an underlying function like httr.}
}
\value{
A data frame with the following fields: school, home_game,
away_game, runs_scored_home, runs_allowed_home, run_scored_away,
runs_allowed_away, base_pf (base park factor), home_game_adj (an adjustment for the percentage of home games played) final_pf (park factor after adjustments)\tabular{lll}{
   col_name \tab types \tab description \cr
   school \tab character \tab Team name. \cr
   home_game \tab numeric \tab Number of home games in the sample. \cr
   away_game \tab numeric \tab Number of away games in the sample. \cr
   runs_scored_home \tab numeric \tab Runs scored by the team in home games. \cr
   runs_allowed_home \tab numeric \tab Runs allowed by the team in home games. \cr
   runs_scored_away \tab numeric \tab Runs scored by the team in away games. \cr
   runs_allowed_away \tab numeric \tab Runs allowed by the team in away games. \cr
   base_pf \tab numeric \tab Base park factor (before adjustment). \cr
   home_game_adj \tab numeric \tab Adjustment for the share of home games played. \cr
   final_pf \tab numeric \tab Park factor after adjustments. \cr
}
}
\description{
\strong{Get Park Effects for NCAA Baseball Teams}
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(ncaa_park_factor(team_id = 736, years = c(2018:2019), type = "conference"))
}\if{html}{\out{</div>}}
}
