% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncaa_season_id_lu}
\alias{ncaa_season_id_lu}
\title{\strong{A data set of college baseball seasons}}
\format{
A data frame with 4 variables:
\itemize{
\item \code{season}: Year date
\item \code{id}: Season id
\item \code{batting_id}: Batting id
\item \code{pitching_id}: Pitching id
}
}
\usage{
data(ncaa_season_id_lu)
}
\description{
A data set college baseball
}
\keyword{datasets}
