% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncaa_team_lu}
\alias{ncaa_team_lu}
\title{\strong{A data set colleges and their athletic conferences and divisions}}
\format{
A data frame with 6 variables
\itemize{
\item \code{school}: School name
\item \code{conference}: Conference name
\item \code{school_id}: School id
\item \code{year}: Year date
\item \code{division}: School division
\item \code{conference_id}: School conference id
}
}
\usage{
data(ncaa_team_lu)
}
\description{
A data set colleges and their athletic conferences and divisions keyed by school and year.
}
\keyword{datasets}
