% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parquet_from_url}
\alias{parquet_from_url}
\title{\strong{Load .parquet file from a remote connection}}
\usage{
parquet_from_url(url)
}
\arguments{
\item{url}{a character url}
}
\description{
Retrieves a parquet file from URL.
}
\keyword{internal}
