% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd2_pitcher_boxscore.R
\name{pitcher_boxscore}
\alias{pitcher_boxscore}
\title{\strong{Retrieve pitcher boxscore data for a single game played}}
\usage{
pitcher_boxscore(x)
}
\arguments{
\item{x}{A boxscore.xml url for a given game from the MLBAM GameDay app data.}
}
\description{
This function allows a user to retrieve a boxscore of pitcher statistics for any game played in the PITCHf/x era (2008-current). The function takes a boxscore.xml url as it's only argument and returns boxscore data for both the home and away batters.
}
\examples{
\donttest{
# pitchers
url_base <- "http://gd2.mlb.com/components/game/mlb/"
url <- paste0(url_base,
  "year_2016/month_05/day_21/gid_2016_05_21_milmlb_nynmlb_1/boxscore.xml")
pitcher_boxscore(url)}
}
