% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pitcher_game_logs.R
\name{pitcher_game_logs_fg}
\alias{pitcher_game_logs_fg}
\title{\strong{(legacy) Scrape Pitcher Game Logs from FanGraphs}}
\usage{
pitcher_game_logs_fg(playerid, year = 2017)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\value{
A data frame of pitcher game logs.
}
\description{
\strong{(legacy) Scrape Pitcher Game Logs from FanGraphs}
}
\keyword{legacy}
