% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerid_lookup.R
\name{playerid_lookup}
\alias{playerid_lookup}
\title{Look up Baseball Player IDs}
\usage{
playerid_lookup(string = NULL)
}
\arguments{
\item{string}{A text string used to return results for players with that string in their last name.}
}
\description{
This function allows you to query the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.
}
\examples{
\dontrun{playerid_lookup("Synder")}
}
\keyword{MLB,}
\keyword{sabermetrics}

