% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playername_lookup.R
\name{playername_lookup}
\alias{playername_lookup}
\title{Look up Baseball Player Name}
\usage{
playername_lookup(id)
}
\arguments{
\item{id}{An integer or character string representing a player ID in a baseball database, cross-referenced through the Chadwick Bureau's public register of baseball players.}
}
\description{
This function allows you to query the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.
}
\examples{
\donttest{
  playername_lookup(4885)
  playername_lookup("kaaihki01")
}
}
