% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_statcast.R
\name{process_statcast_payload}
\alias{process_statcast_payload}
\title{Process Baseball Savant CSV payload}
\usage{
process_statcast_payload(payload)
}
\arguments{
\item{payload}{payload from a Baseball Savant request, e.g. from utils::read.csv}
}
\description{
Process Baseball Savant CSV payload
}
\keyword{MLB,}
\keyword{Statcast}
\keyword{sabermetrics,}
