% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progressively}
\alias{progressively}
\title{Progressively}
\usage{
progressively(f, p = NULL)
}
\arguments{
\item{f}{a function to add progressr functionality to.}

\item{p}{a progressor function as created by `progressr::progressor()`}
}
\value{
a function that does the same as `f` but it calls `p()` after iteration.
}
\description{
This function helps add progress-reporting to any function - given function `f()` and progressor `p()`, it will return a new function that calls `f()` and then (on-exiting) will call `p()` after every iteration.
}
\details{
This is inspired by purrr's `safely`, `quietly`, and `possibly` function decorators.
}
\keyword{Internal}
