% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rds_from_url}
\alias{rds_from_url}
\title{\strong{Load .rds file from a remote connection}}
\usage{
rds_from_url(url)
}
\arguments{
\item{url}{a character url}
}
\value{
a dataframe as created by \code{\link[=readRDS]{readRDS()}}
}
\description{
\strong{Load .rds file from a remote connection}
}
\keyword{internal}
