% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrosheet_data.R
\name{retrosheet_data}
\alias{retrosheet_data}
\title{\strong{Get, Parse, and Format Retrosheet Event and Roster Files}}
\usage{
retrosheet_data(
  path_to_directory,
  years_to_acquire = most_recent_mlb_season(),
  sequence_years = FALSE
)
}
\arguments{
\item{path_to_directory}{A file path that either 1) creates a new directory
or 2) a path to an existing directory}

\item{years_to_acquire}{The seasons to collect. Single, multiple, and
sequential years can be passed. If passing multiple years, enclose in a
vector (i.e. c(2017,2018)). Defaults to \code{most_recent_mlb_season()}.}

\item{sequence_years}{If the seasons passed in the years_to_acquire parameter
should be sequenced so that the function returns all years including and
between the vector passed, set the argument to TRUE. Defaults to FALSE.}
}
\value{
Returns two csv files to the unzipped directory: 1) a combined csv
of the event data for a given year and 2) a combined csv of each team's
roster for each year
}
\description{
\strong{Get, Parse, and Format Retrosheet Event and Roster Files}
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{get_retrosheet_data(path_to_directory = "/Users/Documents/retrosheet/", 
    years_to_acquire = c(1957,1959), 
    sequence_years = F)
}\if{html}{\out{</div>}}
}
