% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_statcast_savant_batter_all.R
\name{scrape_statcast_savant_batter_all}
\alias{scrape_statcast_savant_batter_all}
\title{Query Statcast and PITCHf/x Data for All Batters from baseballsavant.mlb.com}
\usage{
scrape_statcast_savant_batter_all(start_date, end_date)
}
\arguments{
\item{start_date}{Date of first game for which you want data. Format must be in YYYY-MM-DD format.}

\item{end_date}{Date of last game for which you want data. Format must be in YYYY-MM-DD format.}
}
\description{
This function allows you to query Statcast and PITCHf/x data as provided on baseballsavant.mlb.com and have that data returned as a dataframe. Query returns data for all batters over a given time frame.
}
\examples{
\dontrun{
scrape_statcast_savant_batter_all(start_date = "2016-04-06", end_date = "2016-04-15")
}
}
\keyword{MLB,}
\keyword{Statcast}
\keyword{sabermetrics,}
