% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrac.R
\name{spotrac}
\alias{spotrac}
\alias{sptrc}
\alias{payroll}
\title{\strong{Spotrac Endpoint Overview}}
\description{
\itemize{
\item \code{sptrc_team_active_payroll()}: Scrape Team Active Payroll Breakdown from Spotrac.
\item \code{sptrc_league_payrolls()}: Scrape League Payroll Breakdowns from Spotrac.
}
}
\details{
\subsection{\strong{Spotrac}}{

These functions retrieve team and league payroll breakdowns from Spotrac.
}
}
\concept{Spotrac}
