% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast_load_seasons.R
\name{statcast_load_seasons}
\alias{statcast_load_seasons}
\title{Function for stacking two columns with a third adjusting the color}
\usage{
statcast_load_seasons(
  seasons = most_recent_mlb_season(),
  file_type = "rds",
  remove_st = TRUE,
  remove_ex = TRUE
)
}
\arguments{
\item{seasons}{A numeric vector of 4-digit years associated with MLB seasons - defaults to latest season. If set to \code{TRUE}, returns all available data since 2015.}

\item{file_type}{One of \code{c("rds", "parquet")}}

\item{remove_st}{boolean. Should Spring Training data be removed? Defaults to \code{TRUE}}

\item{remove_ex}{boolean. Should Exhibition data be removed? Defaults to \code{TRUE}}
}
\value{
Returns the complete statcast dataset from supplied years
}
\description{
This function is used to load MLB data by year from the \href{https://github.com/sportsdataverse/baseballr-data}{baseballr-data repository}
}
