% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats_api_live_empty_df}
\alias{stats_api_live_empty_df}
\title{providing the column structure of stats_api_live_empty_df}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 131 columns.
}
\usage{
stats_api_live_empty_df
}
\description{
An empty tibble
}
\details{
Dataset
}
\keyword{datasets}
