% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_team_consistency.R
\name{team_consistency}
\alias{team_consistency}
\title{\strong{Calculate Team-level Consistency}}
\usage{
team_consistency(year)
}
\arguments{
\item{year}{Season consistency should be run for.}
}
\value{
Returns a tibble with the following columns\tabular{lll}{
   col_name \tab types \tab description \cr
   Team \tab character \tab Team name. \cr
   Con_R \tab numeric \tab Run-scoring consistency score for the season. \cr
   Con_RA \tab numeric \tab Run-prevention (runs allowed) consistency score. \cr
   Con_R_Ptile \tab numeric \tab Percentile rank of run-scoring consistency. \cr
   Con_RA_Ptile \tab numeric \tab Percentile rank of run-prevention consistency. \cr
}
}
\description{
This function allows you to calculate team-level consistency in run scoring and run prevention over the course of an entire season.
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{  try(team_consistency(year=2021))
}\if{html}{\out{</div>}}
}
