% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_results_bref.R
\name{team_results_bref}
\alias{team_results_bref}
\title{Scrape Team Results}
\usage{
team_results_bref(Tm, year)
}
\arguments{
\item{Tm}{The abbreviation used by Baseball-Reference.com for the team whose results you want to scrape.}

\item{year}{Season for which you want to scrape the park factors.}
}
\description{
This function allows you to scrape schedule and results for a major league team from Baseball-Reference.com
}
\examples{
team_results_bref("NYM", 2015)
returns the schedule and results for the New York Mets from the 2015 season
team_results_bref("TBR", 2008)
returns the schedule and results for the Tampa Bay Rays from the 2008 season
}
\keyword{MLB,}
\keyword{sabermetrics}

