% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams_lu_table}
\alias{teams_lu_table}
\title{\strong{A Team Lookup Table}}
\format{
An object of class \code{data.frame} with 797 rows and 31 columns.
}
\usage{
teams_lu_table
}
\description{
A Team Lookup Table
}
\keyword{datasets}
