% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_gb_on_period.R
\name{viz_gb_on_period}
\alias{viz_gb_on_period}
\title{Scrape MLB Standings on a Given Period and Visualize the Games Behind (GB) on any division or league the league}
\usage{
viz_gb_on_period(start_date, end_date, lg_div)
}
\arguments{
\item{start_date}{a date object representing the first date of the period}

\item{end_date}{a date object representing the last date of the period}

\item{div_viz}{One or more of AL East, AL Central, AL West,
AL Overall, NL East, NL Central, NL West, and NL Overall}
}
\description{
This function allows you to scrape the standings from MLB for a period you choose, and visualize the GB of teams along that period.
}
\examples{
\dontrun{
viz_gb_on_period("2017-04-02","2017-04-10", "AL East")
}
}
\keyword{MLB,}
\keyword{standings}
