% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_rd.R
\name{viz_rd}
\alias{viz_rd}
\title{Visualize MLB team's Runs Differential}
\usage{
viz_rd(Tm, year)
}
\arguments{
\item{Tm}{a character object representing: Team name (BBRef), or Division, or League, or entire MLB}

\item{year}{a numeric object representing the season to visualize}
}
\value{
This function allows to scrape MLB teams results on a specific season and visualizes the cumulative runs differential
}
\description{
Visualize MLB team's Runs Differential
}
\examples{
viz_rd("AL East", 2019)
viz_rd("BOS", 2018)
viz_rd("MLB", 2002)
\dontrun{
viz_rd("Boston Red Sox", 2019)
}
}
\keyword{BBRef}
\keyword{Baseball,}
\keyword{MLB,}
\keyword{Results,}
\keyword{Runs,}
