% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wlpct_on_period.R
\name{viz_wlpct_on_period}
\alias{viz_wlpct_on_period}
\title{Scrape MLB Standings on a Given Period and Visualize the Win-Loss percentage (WLpct) on any division or league}
\usage{
viz_wlpct_on_period(start_date, end_date, lg_div)
}
\arguments{
\item{start_date}{a date object representing the first date of the period}

\item{end_date}{a date object representing the last date of the period}

\item{lg_div}{One or more of AL East, AL Central, AL West,
AL Overall, NL East, NL Central, NL West, and NL Overall}
}
\description{
This function allows you to scrape the standings from MLB for a period you choose, and visualize the Win-Loss percentage of teams along that period.
}
\examples{
\dontrun{
viz_wlpct_on_period("2018-03-29","2018-04-25", "AL Overall")
}
}
\keyword{MLB,}
\keyword{standings}
