% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_woba_plus.R
\name{woba_plus}
\alias{woba_plus}
\title{\strong{Calculate wOBA and related metrics for any set of data}}
\usage{
woba_plus(df)
}
\arguments{
\item{df}{A data frame of statistics that includes, at a minimum, the following columns: uBB (unintentional walks), HBP (Hit By Pitch), X1B (singles), X2B (doubles), X3B (triples), HR (home runs), AB (at-bats), SH (sacrifice hits), SO (strike outs), and season.}
}
\value{
Returns a tibble with the wOBA factors calculated and the following columns:\tabular{lll}{
   col_name \tab types \tab description \cr
   bbref_id \tab character \tab Baseball-Reference player id \cr
   season \tab integer \tab Season (4-digit year) \cr
   Name \tab character \tab Player name \cr
   Age \tab numeric \tab Player age \cr
   Level \tab character \tab Level of play (e.g. MLB or minor-league level) \cr
   Team \tab character \tab Team name \cr
   G \tab numeric \tab Games played \cr
   PA \tab numeric \tab Plate appearances \cr
   AB \tab numeric \tab At-bats \cr
   R \tab numeric \tab Runs scored \cr
   H \tab numeric \tab Hits \cr
   X1B \tab numeric \tab Singles \cr
   X2B \tab numeric \tab Doubles \cr
   X3B \tab numeric \tab Triples \cr
   HR \tab numeric \tab Home runs \cr
   RBI \tab numeric \tab Runs batted in \cr
   BB \tab numeric \tab Walks (bases on balls) \cr
   IBB \tab numeric \tab Intentional walks \cr
   uBB \tab numeric \tab Unintentional walks (BB minus IBB) \cr
   SO \tab numeric \tab Strikeouts \cr
   HBP \tab numeric \tab Hit by pitch \cr
   SH \tab numeric \tab Sacrifice hits (bunts) \cr
   SF \tab numeric \tab Sacrifice flies \cr
   GDP \tab numeric \tab Grounded into double plays \cr
   SB \tab numeric \tab Stolen bases \cr
   CS \tab numeric \tab Caught stealing \cr
   BA \tab numeric \tab Batting average \cr
   OBP \tab numeric \tab On-base percentage \cr
   SLG \tab numeric \tab Slugging percentage \cr
   OPS \tab numeric \tab On-base plus slugging \cr
   wOBA \tab numeric \tab Weighted on-base average (per plate appearance) \cr
   wOBA_CON \tab numeric \tab Weighted on-base average per instance of fair contact \cr
}
}
\description{
This function allows you to calculate wOBA for any given set of data, provided the right variables are in the data set. The function currently returns both wOBA per plate appearance on wOBA per instance of fair contact.
}
\examples{
\donttest{
 try({
   df <- bref_daily_batter("2015-08-01", "2015-10-03") 
   woba_plus(df)
 })
}
}
