% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_game_logs_fg.R
\name{batter_game_logs_fg}
\alias{batter_game_logs_fg}
\title{Scrape Batter Game Logs from FanGraphs}
\usage{
batter_game_logs_fg(playerid, year = 2017)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\description{
This function allows you to scrape game logs by year for a batter from FanGraphs.com.
}
\examples{
\dontrun{batter_game_logs_fg(playerid = 6184, year = 2017)}
}
\keyword{MLB,}
\keyword{sabermetrics}
