% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_daily_batter.R
\name{bref_daily_batter}
\alias{bref_daily_batter}
\title{\strong{Scrape Batter Performance Data Over a Custom Time Frame}}
\usage{
bref_daily_batter(t1, t2)
}
\arguments{
\item{t1}{First date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}

\item{t2}{Last date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}
}
\value{
Returns a data frame of batter performance
|col_name |types     |
|:--------|:---------|
|bbref_id |character |
|season   |character |
|Name     |character |
|Age      |numeric   |
|Level    |character |
|Team     |character |
|G        |numeric   |
|PA       |numeric   |
|AB       |numeric   |
|R        |numeric   |
|H        |numeric   |
|X1B      |numeric   |
|X2B      |numeric   |
|X3B      |numeric   |
|HR       |numeric   |
|RBI      |numeric   |
|BB       |numeric   |
|IBB      |numeric   |
|uBB      |numeric   |
|SO       |numeric   |
|HBP      |numeric   |
|SH       |numeric   |
|SF       |numeric   |
|GDP      |numeric   |
|SB       |numeric   |
|CS       |numeric   |
|BA       |numeric   |
|OBP      |numeric   |
|SLG      |numeric   |
|OPS      |numeric   |
}
\description{
This function allows you to scrape basic batter statistics over a custom time frame. Data is sourced from Baseball-Reference.com.
}
\examples{
\donttest{
  bref_daily_batter("2015-05-10", "2015-06-20")
}
}
