% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_team_results.R
\name{bref_team_results}
\alias{bref_team_results}
\title{\strong{Scrape Team Results}}
\usage{
bref_team_results(Tm, year)
}
\arguments{
\item{Tm}{The abbreviation used by Baseball-Reference.com for the team whose results you want to scrape.}

\item{year}{Season for which you want to scrape the park factors.}
}
\value{
Returns a data frame of MLB team results
|col_name       |types     |
|:--------------|:---------|
|Gm             |numeric   |
|Date           |character |
|Tm             |character |
|H_A            |character |
|Opp            |character |
|Result         |character |
|R              |numeric   |
|RA             |numeric   |
|Inn            |character |
|Record         |character |
|Rank           |numeric   |
|GB             |character |
|Win            |character |
|Loss           |character |
|Save           |character |
|Time           |character |
|D/N            |character |
|Attendance     |numeric   |
|cLI            |numeric   |
|Streak         |numeric   |
|Orig_Scheduled |character |
|Year           |numeric   |
}
\description{
This function allows you to scrape schedule and results for a major league team from Baseball-Reference.com
}
\examples{
\donttest{
  bref_team_results("NYM", 2015)
  bref_team_results(Tm="TBR", year=2008)
}
}
