% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick_playerid_lookup.R
\name{playerid_lookup}
\alias{playerid_lookup}
\title{\strong{Look up Baseball Player IDs by Player Name}}
\usage{
playerid_lookup(last_name = NULL, first_name = NULL)
}
\arguments{
\item{last_name}{A text string used to return results for players with that string in their last name.}

\item{first_name}{A text string used to return results for players with that string in their first name.}
}
\value{
A data frame of baseball players and the various IDs associated with them in different systems of record.
|col_name         |types     |
|:----------------|:---------|
|first_name       |character |
|last_name        |character |
|given_name       |character |
|name_suffix      |character |
|nick_name        |character |
|birth_year       |integer   |
|mlb_played_first |integer   |
|mlbam_id         |integer   |
|retrosheet_id    |character |
|bbref_id         |character |
|fangraphs_id     |integer   |
}
\description{
This function allows you to query the Chadwick Bureau's public register of baseball players and the various IDs associated with them in different systems of record.
}
\examples{
\donttest{
  playerid_lookup("Garcia", "Karim")
}
}
