% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick_player_lu.R
\name{chadwick_player_lu}
\alias{chadwick_player_lu}
\alias{get_chadwick_lu}
\title{\strong{Download the Chadwick Bureau's public register of baseball players}}
\usage{
chadwick_player_lu()

get_chadwick_lu()
}
\value{
A data frame of baseball players and the various IDs associated with them in different systems of record.
}
\description{
\strong{Download the Chadwick Bureau's public register of baseball players}
}
\examples{
\donttest{
  chadwick_player_lu()
}
}
