% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadwick_player_lu.R
\name{chadwick_player_lu}
\alias{chadwick_player_lu}
\alias{get_chadwick_lu}
\title{\strong{Download the Chadwick Bureau's public register of baseball players}}
\usage{
chadwick_player_lu()

get_chadwick_lu()
}
\value{
A data frame of baseball players and the various IDs associated with them in different systems of record.\tabular{ll}{
   col_name \tab types \cr
   key_person \tab character \cr
   key_uuid \tab character \cr
   key_mlbam \tab integer \cr
   key_retro \tab character \cr
   key_bbref \tab character \cr
   key_bbref_minors \tab character \cr
   key_fangraphs \tab integer \cr
   key_npb \tab integer \cr
   key_sr_nfl \tab character \cr
   key_sr_nba \tab character \cr
   key_sr_nhl \tab character \cr
   key_findagrave \tab integer \cr
   name_last \tab character \cr
   name_first \tab character \cr
   name_given \tab character \cr
   name_suffix \tab character \cr
   name_matrilineal \tab character \cr
   name_nick \tab character \cr
   birth_year \tab integer \cr
   birth_month \tab integer \cr
   birth_day \tab integer \cr
   death_year \tab integer \cr
   death_month \tab integer \cr
   death_day \tab integer \cr
   pro_played_first \tab integer \cr
   pro_played_last \tab integer \cr
   mlb_played_first \tab integer \cr
   mlb_played_last \tab integer \cr
   col_played_first \tab integer \cr
   col_played_last \tab integer \cr
   pro_managed_first \tab integer \cr
   pro_managed_last \tab integer \cr
   mlb_managed_first \tab integer \cr
   mlb_managed_last \tab integer \cr
   col_managed_first \tab integer \cr
   col_managed_last \tab integer \cr
   pro_umpired_first \tab integer \cr
   pro_umpired_last \tab integer \cr
   mlb_umpired_first \tab integer \cr
   mlb_umpired_last \tab integer \cr
}

A data frame of baseball players and the various IDs associated with them in different systems of record.
}
\description{
\strong{Download the Chadwick Bureau's public register of baseball players}
}
\examples{
\donttest{
  try(chadwick_player_lu())
}
}
