% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_barrel.R
\name{code_barrel}
\alias{code_barrel}
\title{Code whether a batted ball is a "barrel" based on a Statcast dataset from baseballsavant.mlb.com}
\usage{
code_barrel(df)
}
\arguments{
\item{df}{A dataframe generated by baseballsavant.mlb.com that must include the following variables: hit_angle and hit_speed.}
}
\description{
This function allows you to code a batted ball as a barrel as defined by the Statcast research team using data as provided by baseballsavant.mlb.com.
}
\examples{
\dontrun{code_barrel(df)}
}
\keyword{MLB,}
\keyword{Statcast}
\keyword{sabermetrics,}
