% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_batter_bref.R
\name{daily_batter_bref}
\alias{daily_batter_bref}
\title{Scrape Batter Performance Data Over a Custom Time Frame}
\usage{
daily_batter_bref(t1, t2)
}
\arguments{
\item{t1}{First date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}

\item{t2}{Last date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}
}
\description{
This function allows you to scrape basic batter statistics over a custom time frame. Data is sourced from Baseball-Reference.com.
}
\examples{
\dontrun{daily_batter_bref("2015-05-10", "2015-06-20")}
}
\keyword{MLB,}
\keyword{sabermetrics}
