% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_pitcher_bref.R
\name{daily_pitcher_bref}
\alias{daily_pitcher_bref}
\title{Scrape Pitcher Performance Data Over a Custom Time Frame}
\usage{
daily_pitcher_bref(t1, t2)
}
\arguments{
\item{t1}{First date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}

\item{t2}{Last date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}
}
\description{
This function allows you to scrape basic pitcher statistics over a custom time frame. Data is sourced from Baseball-Reference.com.
}
\examples{
\dontrun{daily_pitcher_bref("2015-05-10", "2015-06-20")}
}
\keyword{MLB,}
\keyword{sabermetrics}
