% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_edge_frequency.R
\name{edge_frequency}
\alias{edge_frequency}
\title{Edge Percentage Frequency}
\usage{
edge_frequency(df, group = NULL)
}
\arguments{
\item{df}{A data frame of pitch by pitch data that has been coded using the edge_code() function.}

\item{group}{Character string indicating what column to group the frequency by. For example, "pitcher" or "batter". Defaults to NULL, which calculates the frequencies across the entire data set.}
}
\description{
This function allows you to calculate the percent of pitches thrown to different edges of the strike zone for a pitch by pitch data set that has been coded using the edge_code() function.
}
