% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_college_baseball_news.R
\name{espn_college_baseball_news}
\alias{espn_college_baseball_news}
\title{\strong{Get ESPN College Baseball News}}
\usage{
espn_college_baseball_news(limit = 50)
}
\arguments{
\item{limit}{integer. Maximum number of articles to return. Default \code{50}.}
}
\value{
Returns a tibble of news articles.\tabular{lll}{
   col_name \tab types \tab description \cr
   id \tab character \tab Id. \cr
   type \tab character \tab Record type / category. \cr
   headline \tab character \tab News headline. \cr
   description \tab character \tab Long-form description text. \cr
   published \tab character \tab Publication timestamp (ISO 8601). \cr
   premium \tab logical \tab  \cr
   byline \tab character \tab News article byline / author. \cr
   link_web \tab character \tab Web link / URL. \cr
   league_id \tab character \tab ESPN league identifier. \cr
}
}
\description{
\strong{Get ESPN College Baseball News}

\strong{Get ESPN College Baseball News}
}
\examples{
\donttest{
  try(espn_college_baseball_news(limit = 5))
}
}
\seealso{
Other ESPN College Baseball Functions: 
\code{\link{espn_college_baseball_athletes_index}()},
\code{\link{espn_college_baseball_calendar}()},
\code{\link{espn_college_baseball_coach}()},
\code{\link{espn_college_baseball_coach_record}()},
\code{\link{espn_college_baseball_coach_season}()},
\code{\link{espn_college_baseball_coaches}()},
\code{\link{espn_college_baseball_conferences}()},
\code{\link{espn_college_baseball_game_all}()},
\code{\link{espn_college_baseball_game_broadcasts}()},
\code{\link{espn_college_baseball_game_official_detail}()},
\code{\link{espn_college_baseball_game_officials}()},
\code{\link{espn_college_baseball_game_play}()},
\code{\link{espn_college_baseball_game_play_personnel}()},
\code{\link{espn_college_baseball_game_player_box}()},
\code{\link{espn_college_baseball_game_rosters}()},
\code{\link{espn_college_baseball_game_situation}()},
\code{\link{espn_college_baseball_game_team_leaders}()},
\code{\link{espn_college_baseball_game_team_linescores}()},
\code{\link{espn_college_baseball_game_team_records}()},
\code{\link{espn_college_baseball_game_team_roster}()},
\code{\link{espn_college_baseball_game_team_roster_entry}()},
\code{\link{espn_college_baseball_game_team_score}()},
\code{\link{espn_college_baseball_game_team_statistics}()},
\code{\link{espn_college_baseball_leaders}()},
\code{\link{espn_college_baseball_pbp}()},
\code{\link{espn_college_baseball_player_awards}()},
\code{\link{espn_college_baseball_player_box}()},
\code{\link{espn_college_baseball_player_career_stats}()},
\code{\link{espn_college_baseball_player_eventlog}()},
\code{\link{espn_college_baseball_player_gamelog}()},
\code{\link{espn_college_baseball_player_info}()},
\code{\link{espn_college_baseball_player_overview}()},
\code{\link{espn_college_baseball_player_seasons}()},
\code{\link{espn_college_baseball_player_splits}()},
\code{\link{espn_college_baseball_player_statisticslog}()},
\code{\link{espn_college_baseball_scoreboard}()},
\code{\link{espn_college_baseball_season_group}()},
\code{\link{espn_college_baseball_season_group_children}()},
\code{\link{espn_college_baseball_season_group_teams}()},
\code{\link{espn_college_baseball_season_groups}()},
\code{\link{espn_college_baseball_season_info}()},
\code{\link{espn_college_baseball_season_leaders}()},
\code{\link{espn_college_baseball_season_ranking}()},
\code{\link{espn_college_baseball_season_rankings}()},
\code{\link{espn_college_baseball_season_type}()},
\code{\link{espn_college_baseball_season_types}()},
\code{\link{espn_college_baseball_season_week}()},
\code{\link{espn_college_baseball_season_weeks}()},
\code{\link{espn_college_baseball_seasons}()},
\code{\link{espn_college_baseball_standings}()},
\code{\link{espn_college_baseball_team}()},
\code{\link{espn_college_baseball_team_box}()},
\code{\link{espn_college_baseball_team_leaders}()},
\code{\link{espn_college_baseball_team_news}()},
\code{\link{espn_college_baseball_team_record}()},
\code{\link{espn_college_baseball_team_record_detail}()},
\code{\link{espn_college_baseball_team_roster}()},
\code{\link{espn_college_baseball_team_schedule}()},
\code{\link{espn_college_baseball_team_season_profile}()},
\code{\link{espn_college_baseball_team_season_roster}()},
\code{\link{espn_college_baseball_team_season_statistics}()},
\code{\link{espn_college_baseball_teams}()},
\code{\link{espn_college_baseball_tournament}()},
\code{\link{espn_college_baseball_tournament_season}()},
\code{\link{espn_college_baseball_tournament_seasons}()},
\code{\link{espn_college_baseball_tournaments}()},
\code{\link{espn_college_baseball_venues}()},
\code{\link{espn_college_baseball_week_ranking}()},
\code{\link{espn_college_baseball_week_rankings}()}
}
\author{
Saiem Gilani
}
\concept{ESPN College Baseball Functions}
