% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_college_baseball_data.R
\name{espn_college_baseball_standings}
\alias{espn_college_baseball_standings}
\title{\strong{Get ESPN College Baseball Standings}}
\usage{
espn_college_baseball_standings(year)
}
\arguments{
\item{year}{Either numeric or character (YYYY)}
}
\value{
A standings data frame\tabular{lll}{
   col_name \tab types \tab description \cr
   team_id \tab integer \tab Unique ESPN team identifier. \cr
   team \tab character \tab Team. \cr
   otlosses \tab character \tab Otlosses. \cr
   otwins \tab character \tab Otwins. \cr
   avgpointsagainst \tab numeric \tab Avgpointsagainst. \cr
   avgpointsfor \tab numeric \tab Avgpointsfor. \cr
   clincher \tab numeric \tab Clincher. \cr
   differential \tab numeric \tab Differential. \cr
   divisionwinpercent \tab numeric \tab Divisionwinpercent. \cr
   gamesbehind \tab numeric \tab Gamesbehind. \cr
   gamesplayed \tab character \tab Gamesplayed. \cr
   leaguewinpercent \tab numeric \tab Leaguewinpercent. \cr
   losses \tab numeric \tab Losses. \cr
   playoffseed \tab numeric \tab Playoffseed. \cr
   pointdifferential \tab character \tab Pointdifferential. \cr
   points \tab character \tab Points. \cr
   pointsagainst \tab character \tab Pointsagainst. \cr
   pointsfor \tab character \tab Pointsfor. \cr
   streak \tab numeric \tab Streak. \cr
   ties \tab character \tab Ties. \cr
   winpercent \tab numeric \tab Winpercent. \cr
   wins \tab numeric \tab Wins. \cr
   divisiongamesbehind \tab character \tab Divisiongamesbehind. \cr
   divisionpercent \tab character \tab Divisionpercent. \cr
   divisiontied \tab character \tab Divisiontied. \cr
   homelosses \tab character \tab Homelosses. \cr
   hometies \tab character \tab Hometies. \cr
   homewins \tab character \tab Homewins. \cr
   magicnumberdivision \tab character \tab Magicnumberdivision. \cr
   magicnumberwildcard \tab character \tab Magicnumberwildcard. \cr
   playoffpercent \tab character \tab Playoffpercent. \cr
   roadlosses \tab character \tab Roadlosses. \cr
   roadties \tab character \tab Roadties. \cr
   roadwins \tab character \tab Roadwins. \cr
   wildcardpercent \tab character \tab Wildcardpercent. \cr
   total \tab character \tab Total. \cr
   home \tab character \tab Home. \cr
   road \tab character \tab Road. \cr
   intradivision \tab character \tab Intradivision. \cr
   intraleague \tab character \tab Intraleague. \cr
   lasttengames \tab character \tab Lasttengames. \cr
}
}
\description{
\strong{Get ESPN College Baseball's Standings}
}
\examples{
\donttest{
try(espn_college_baseball_standings(year = 2025))
}
}
\seealso{
Other ESPN College Baseball Functions: 
\code{\link{espn_college_baseball_athletes_index}()},
\code{\link{espn_college_baseball_calendar}()},
\code{\link{espn_college_baseball_coach}()},
\code{\link{espn_college_baseball_coach_record}()},
\code{\link{espn_college_baseball_coach_season}()},
\code{\link{espn_college_baseball_coaches}()},
\code{\link{espn_college_baseball_conferences}()},
\code{\link{espn_college_baseball_game_all}()},
\code{\link{espn_college_baseball_game_broadcasts}()},
\code{\link{espn_college_baseball_game_official_detail}()},
\code{\link{espn_college_baseball_game_officials}()},
\code{\link{espn_college_baseball_game_play}()},
\code{\link{espn_college_baseball_game_play_personnel}()},
\code{\link{espn_college_baseball_game_player_box}()},
\code{\link{espn_college_baseball_game_rosters}()},
\code{\link{espn_college_baseball_game_situation}()},
\code{\link{espn_college_baseball_game_team_leaders}()},
\code{\link{espn_college_baseball_game_team_linescores}()},
\code{\link{espn_college_baseball_game_team_records}()},
\code{\link{espn_college_baseball_game_team_roster}()},
\code{\link{espn_college_baseball_game_team_roster_entry}()},
\code{\link{espn_college_baseball_game_team_score}()},
\code{\link{espn_college_baseball_game_team_statistics}()},
\code{\link{espn_college_baseball_leaders}()},
\code{\link{espn_college_baseball_news}()},
\code{\link{espn_college_baseball_pbp}()},
\code{\link{espn_college_baseball_player_awards}()},
\code{\link{espn_college_baseball_player_box}()},
\code{\link{espn_college_baseball_player_career_stats}()},
\code{\link{espn_college_baseball_player_eventlog}()},
\code{\link{espn_college_baseball_player_gamelog}()},
\code{\link{espn_college_baseball_player_info}()},
\code{\link{espn_college_baseball_player_overview}()},
\code{\link{espn_college_baseball_player_seasons}()},
\code{\link{espn_college_baseball_player_splits}()},
\code{\link{espn_college_baseball_player_statisticslog}()},
\code{\link{espn_college_baseball_scoreboard}()},
\code{\link{espn_college_baseball_season_group}()},
\code{\link{espn_college_baseball_season_group_children}()},
\code{\link{espn_college_baseball_season_group_teams}()},
\code{\link{espn_college_baseball_season_groups}()},
\code{\link{espn_college_baseball_season_info}()},
\code{\link{espn_college_baseball_season_leaders}()},
\code{\link{espn_college_baseball_season_ranking}()},
\code{\link{espn_college_baseball_season_rankings}()},
\code{\link{espn_college_baseball_season_type}()},
\code{\link{espn_college_baseball_season_types}()},
\code{\link{espn_college_baseball_season_week}()},
\code{\link{espn_college_baseball_season_weeks}()},
\code{\link{espn_college_baseball_seasons}()},
\code{\link{espn_college_baseball_team}()},
\code{\link{espn_college_baseball_team_box}()},
\code{\link{espn_college_baseball_team_leaders}()},
\code{\link{espn_college_baseball_team_news}()},
\code{\link{espn_college_baseball_team_record}()},
\code{\link{espn_college_baseball_team_record_detail}()},
\code{\link{espn_college_baseball_team_roster}()},
\code{\link{espn_college_baseball_team_schedule}()},
\code{\link{espn_college_baseball_team_season_profile}()},
\code{\link{espn_college_baseball_team_season_roster}()},
\code{\link{espn_college_baseball_team_season_statistics}()},
\code{\link{espn_college_baseball_teams}()},
\code{\link{espn_college_baseball_tournament}()},
\code{\link{espn_college_baseball_tournament_season}()},
\code{\link{espn_college_baseball_tournament_seasons}()},
\code{\link{espn_college_baseball_tournaments}()},
\code{\link{espn_college_baseball_venues}()},
\code{\link{espn_college_baseball_week_ranking}()},
\code{\link{espn_college_baseball_week_rankings}()}
}
\author{
Saiem Gilani
}
\concept{ESPN College Baseball Functions}
