% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_college_baseball_data.R
\name{espn_college_baseball_teams}
\alias{espn_college_baseball_teams}
\title{\strong{Get ESPN College Baseball Teams}}
\usage{
espn_college_baseball_teams()
}
\value{
A teams data frame\tabular{lll}{
   col_name \tab types \tab description \cr
   abbreviation \tab character \tab Short abbreviation. \cr
   alternate_color \tab character \tab Alternate color (hex). \cr
   color \tab character \tab Primary color (hex). \cr
   display_name \tab character \tab Display name. \cr
   team_id \tab integer \tab Unique ESPN team identifier. \cr
   team \tab character \tab Team. \cr
   logo \tab character \tab Logo image URL. \cr
   logo_dark \tab character \tab Dark-mode logo image URL. \cr
   logos_href_3 \tab character \tab Logos href 3. \cr
   logos_href_4 \tab character \tab Logos href 4. \cr
   logos_href_5 \tab character \tab Logos href 5. \cr
   logos_href_6 \tab character \tab Logos href 6. \cr
   logos_href_7 \tab character \tab Logos href 7. \cr
   logos_href_8 \tab character \tab Logos href 8. \cr
   logos_href_9 \tab character \tab Logos href 9. \cr
   logos_href_10 \tab character \tab Logos href 10. \cr
   logos_href_11 \tab character \tab Logos href 11. \cr
   logos_href_12 \tab character \tab Logos href 12. \cr
   logos_href_13 \tab character \tab Logos href 13. \cr
   logos_href_14 \tab character \tab Logos href 14. \cr
   logos_href_15 \tab character \tab Logos href 15. \cr
   logos_href_16 \tab character \tab Logos href 16. \cr
   mascot \tab character \tab Team mascot. \cr
   nickname \tab character \tab Team nickname. \cr
   short_name \tab character \tab Short display name. \cr
}
}
\description{
\strong{Get ESPN College Baseball team names and IDs}
}
\examples{
\donttest{
try(espn_college_baseball_teams())
}

}
\seealso{
Other ESPN College Baseball Functions: 
\code{\link{espn_college_baseball_athletes_index}()},
\code{\link{espn_college_baseball_calendar}()},
\code{\link{espn_college_baseball_coach}()},
\code{\link{espn_college_baseball_coach_record}()},
\code{\link{espn_college_baseball_coach_season}()},
\code{\link{espn_college_baseball_coaches}()},
\code{\link{espn_college_baseball_conferences}()},
\code{\link{espn_college_baseball_game_all}()},
\code{\link{espn_college_baseball_game_broadcasts}()},
\code{\link{espn_college_baseball_game_official_detail}()},
\code{\link{espn_college_baseball_game_officials}()},
\code{\link{espn_college_baseball_game_play}()},
\code{\link{espn_college_baseball_game_play_personnel}()},
\code{\link{espn_college_baseball_game_player_box}()},
\code{\link{espn_college_baseball_game_rosters}()},
\code{\link{espn_college_baseball_game_situation}()},
\code{\link{espn_college_baseball_game_team_leaders}()},
\code{\link{espn_college_baseball_game_team_linescores}()},
\code{\link{espn_college_baseball_game_team_records}()},
\code{\link{espn_college_baseball_game_team_roster}()},
\code{\link{espn_college_baseball_game_team_roster_entry}()},
\code{\link{espn_college_baseball_game_team_score}()},
\code{\link{espn_college_baseball_game_team_statistics}()},
\code{\link{espn_college_baseball_leaders}()},
\code{\link{espn_college_baseball_news}()},
\code{\link{espn_college_baseball_pbp}()},
\code{\link{espn_college_baseball_player_awards}()},
\code{\link{espn_college_baseball_player_box}()},
\code{\link{espn_college_baseball_player_career_stats}()},
\code{\link{espn_college_baseball_player_eventlog}()},
\code{\link{espn_college_baseball_player_gamelog}()},
\code{\link{espn_college_baseball_player_info}()},
\code{\link{espn_college_baseball_player_overview}()},
\code{\link{espn_college_baseball_player_seasons}()},
\code{\link{espn_college_baseball_player_splits}()},
\code{\link{espn_college_baseball_player_statisticslog}()},
\code{\link{espn_college_baseball_scoreboard}()},
\code{\link{espn_college_baseball_season_group}()},
\code{\link{espn_college_baseball_season_group_children}()},
\code{\link{espn_college_baseball_season_group_teams}()},
\code{\link{espn_college_baseball_season_groups}()},
\code{\link{espn_college_baseball_season_info}()},
\code{\link{espn_college_baseball_season_leaders}()},
\code{\link{espn_college_baseball_season_ranking}()},
\code{\link{espn_college_baseball_season_rankings}()},
\code{\link{espn_college_baseball_season_type}()},
\code{\link{espn_college_baseball_season_types}()},
\code{\link{espn_college_baseball_season_week}()},
\code{\link{espn_college_baseball_season_weeks}()},
\code{\link{espn_college_baseball_seasons}()},
\code{\link{espn_college_baseball_standings}()},
\code{\link{espn_college_baseball_team}()},
\code{\link{espn_college_baseball_team_box}()},
\code{\link{espn_college_baseball_team_leaders}()},
\code{\link{espn_college_baseball_team_news}()},
\code{\link{espn_college_baseball_team_record}()},
\code{\link{espn_college_baseball_team_record_detail}()},
\code{\link{espn_college_baseball_team_roster}()},
\code{\link{espn_college_baseball_team_schedule}()},
\code{\link{espn_college_baseball_team_season_profile}()},
\code{\link{espn_college_baseball_team_season_roster}()},
\code{\link{espn_college_baseball_team_season_statistics}()},
\code{\link{espn_college_baseball_tournament}()},
\code{\link{espn_college_baseball_tournament_season}()},
\code{\link{espn_college_baseball_tournament_seasons}()},
\code{\link{espn_college_baseball_tournaments}()},
\code{\link{espn_college_baseball_venues}()},
\code{\link{espn_college_baseball_week_ranking}()},
\code{\link{espn_college_baseball_week_rankings}()}
}
\author{
Saiem Gilani
}
\concept{ESPN College Baseball Functions}
