% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mlb_groups.R
\name{espn_mlb_conferences}
\alias{espn_mlb_conferences}
\title{\strong{Get ESPN MLB Conferences}}
\usage{
espn_mlb_conferences()
}
\value{
A \code{baseballr_data} tibble with one row per conference:\tabular{lll}{
   col_name \tab types \tab description \cr
   group_id \tab integer \tab Group identifier (e.g. conference group_id). \cr
   conference_short_name \tab character \tab Conference short name (e.g. 'ACC'). \cr
   conference_uid \tab character \tab ESPN universal conference identifier. \cr
   conference_name \tab character \tab Full conference name. \cr
   conference_logo \tab character \tab Logo image URL for conference. \cr
   parent_group_id \tab integer \tab Unique identifier for parent group. \cr
   conference_id \tab integer \tab Conference identifier. \cr
}
}
\description{
\strong{Get ESPN MLB Conferences}

\strong{Get ESPN MLB Conferences}
}
\examples{
\donttest{
  try(espn_mlb_conferences())
}
}
\seealso{
Other ESPN MLB Functions: 
\code{\link{espn_mlb}},
\code{\link{espn_mlb_athletes_index}()},
\code{\link{espn_mlb_award}()},
\code{\link{espn_mlb_betting}()},
\code{\link{espn_mlb_calendar}()},
\code{\link{espn_mlb_coach}()},
\code{\link{espn_mlb_coach_record}()},
\code{\link{espn_mlb_coach_season}()},
\code{\link{espn_mlb_coaches}()},
\code{\link{espn_mlb_draft}()},
\code{\link{espn_mlb_draft_athlete_detail}()},
\code{\link{espn_mlb_draft_athletes}()},
\code{\link{espn_mlb_draft_pick}()},
\code{\link{espn_mlb_draft_rounds}()},
\code{\link{espn_mlb_draft_status}()},
\code{\link{espn_mlb_franchise}()},
\code{\link{espn_mlb_franchises}()},
\code{\link{espn_mlb_freeagents}()},
\code{\link{espn_mlb_futures}()},
\code{\link{espn_mlb_game_all}()},
\code{\link{espn_mlb_game_broadcasts}()},
\code{\link{espn_mlb_game_endpoints}},
\code{\link{espn_mlb_game_info}()},
\code{\link{espn_mlb_game_odds}()},
\code{\link{espn_mlb_game_official_detail}()},
\code{\link{espn_mlb_game_officials}()},
\code{\link{espn_mlb_game_play}()},
\code{\link{espn_mlb_game_play_personnel}()},
\code{\link{espn_mlb_game_player_box}()},
\code{\link{espn_mlb_game_powerindex}()},
\code{\link{espn_mlb_game_predictor}()},
\code{\link{espn_mlb_game_probabilities}()},
\code{\link{espn_mlb_game_probables}()},
\code{\link{espn_mlb_game_propbets}()},
\code{\link{espn_mlb_game_rosters}()},
\code{\link{espn_mlb_game_situation}()},
\code{\link{espn_mlb_game_team_leaders}()},
\code{\link{espn_mlb_game_team_linescores}()},
\code{\link{espn_mlb_game_team_records}()},
\code{\link{espn_mlb_game_team_roster}()},
\code{\link{espn_mlb_game_team_roster_entry}()},
\code{\link{espn_mlb_game_team_score}()},
\code{\link{espn_mlb_game_team_statistics}()},
\code{\link{espn_mlb_injuries}()},
\code{\link{espn_mlb_leaders}()},
\code{\link{espn_mlb_news}()},
\code{\link{espn_mlb_pbp}()},
\code{\link{espn_mlb_player_awards}()},
\code{\link{espn_mlb_player_box}()},
\code{\link{espn_mlb_player_career_stats}()},
\code{\link{espn_mlb_player_contract}()},
\code{\link{espn_mlb_player_contracts}()},
\code{\link{espn_mlb_player_endpoints}},
\code{\link{espn_mlb_player_eventlog}()},
\code{\link{espn_mlb_player_eventlog_v2}()},
\code{\link{espn_mlb_player_gamelog}()},
\code{\link{espn_mlb_player_info}()},
\code{\link{espn_mlb_player_overview}()},
\code{\link{espn_mlb_player_seasons}()},
\code{\link{espn_mlb_player_splits}()},
\code{\link{espn_mlb_player_statisticslog}()},
\code{\link{espn_mlb_player_stats}()},
\code{\link{espn_mlb_player_stats_v3}()},
\code{\link{espn_mlb_position}()},
\code{\link{espn_mlb_positions}()},
\code{\link{espn_mlb_powerindex}()},
\code{\link{espn_mlb_scoreboard}()},
\code{\link{espn_mlb_season_awards}()},
\code{\link{espn_mlb_season_draft}()},
\code{\link{espn_mlb_season_group}()},
\code{\link{espn_mlb_season_group_children}()},
\code{\link{espn_mlb_season_group_teams}()},
\code{\link{espn_mlb_season_groups}()},
\code{\link{espn_mlb_season_info}()},
\code{\link{espn_mlb_season_leaders}()},
\code{\link{espn_mlb_season_ranking}()},
\code{\link{espn_mlb_season_rankings}()},
\code{\link{espn_mlb_season_type}()},
\code{\link{espn_mlb_season_types}()},
\code{\link{espn_mlb_season_week}()},
\code{\link{espn_mlb_season_weeks}()},
\code{\link{espn_mlb_seasons}()},
\code{\link{espn_mlb_standings}()},
\code{\link{espn_mlb_team}()},
\code{\link{espn_mlb_team_box}()},
\code{\link{espn_mlb_team_current_roster}()},
\code{\link{espn_mlb_team_depthchart}()},
\code{\link{espn_mlb_team_endpoints}},
\code{\link{espn_mlb_team_injuries}()},
\code{\link{espn_mlb_team_leaders}()},
\code{\link{espn_mlb_team_news}()},
\code{\link{espn_mlb_team_odds_records}()},
\code{\link{espn_mlb_team_record}()},
\code{\link{espn_mlb_team_record_detail}()},
\code{\link{espn_mlb_team_roster}()},
\code{\link{espn_mlb_team_schedule}()},
\code{\link{espn_mlb_team_season_profile}()},
\code{\link{espn_mlb_team_season_roster}()},
\code{\link{espn_mlb_team_season_statistics}()},
\code{\link{espn_mlb_team_stats}()},
\code{\link{espn_mlb_teams}()},
\code{\link{espn_mlb_tournament}()},
\code{\link{espn_mlb_tournament_season}()},
\code{\link{espn_mlb_tournament_seasons}()},
\code{\link{espn_mlb_tournaments}()},
\code{\link{espn_mlb_transactions}()},
\code{\link{espn_mlb_venues}()},
\code{\link{espn_mlb_week_ranking}()},
\code{\link{espn_mlb_week_rankings}()},
\code{\link{espn_mlb_wp}()}
}
\author{
Saiem Gilani
}
\concept{ESPN MLB Functions}
\keyword{Conferences}
\keyword{MLB}
