% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_bat_leaders.R
\name{fg_bat_leaders}
\alias{fg_bat_leaders}
\title{Scrape Batter Leaderboards from FanGraphs.com}
\usage{
fg_bat_leaders(x, y, qual)
}
\arguments{
\item{x}{First season for which you want data.}

\item{y}{Last season for which you want data. If multiple years selected, data returned will be aggregate data for the date range. If y = x, function will return single-season data.}

\item{qual}{Whether you want only batters that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimumm number of plate appearaces, use the number desired.}
}
\description{
This function allows you to scrape all leaderboard statistics (basic and advanced) from FanGraphs.com.
}
\examples{
fg_bat_leaders(2013, 2015, "qual")
returns leaderboards for all qualified hitters and their performance from 2013 through 2015.

fg_bat_leaders(2015, 2015, 400)
returns leaderboards for all hitters with at least 400 plate appearances in 2015.
}
\keyword{MLB,}
\keyword{sabermetrics}

