% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{fg_bat_leaders}
\alias{fg_bat_leaders}
\title{Scrape Batter Leaderboards from FanGraphs}
\usage{
fg_bat_leaders(x, y, league = "all", qual = "y", ind = 0)
}
\arguments{
\item{x}{First season for which you want data.}

\item{y}{Last season for which you want data. If multiple years selected, data returned will be aggregate data for the date range. If y = x, function will return single-season data.}

\item{league}{Option for limiting results to different leagues or overall results. Options are "al", "nl", or "all".}

\item{qual}{Whether you want only batters/pitchers that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimumm number of plate appearaces/innings pitched, use the number desired. Defaults to "y".}

\item{ind}{Whether or not to break the seasons out individual, or roll them up together. 1 = split seasons, 0 = aggregate seasons.}
}
\description{
This function allows you to scrape all leaderboard statistics (basic and advanced) from FanGraphs.com.
}
\examples{
\dontrun{fg_bat_leaders(x = 2015, y = 2015, qual = 400)}
}
\keyword{MLB,}
\keyword{sabermetrics}
