% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{fg_guts}
\alias{fg_guts}
\title{Scrape FanGraphs.com Guts!}
\usage{
fg_guts()
}
\description{
Scrape historical FanGraphs Guts! table, wOBA, FIP coefficients and constants

This function allows you to scrape the historical, season-by-season wOBA and FIP constants and coefficients at FanGraphs.com.
}
\examples{
\donttest{
  fg_guts()
}
}
