% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_guts.R
\name{fg_guts}
\alias{fg_guts}
\title{\strong{Scrape FanGraphs.com Guts!}}
\usage{
fg_guts()
}
\value{
Returns a data frame of seasonal constants from FanGraphs
}
\description{
Scrape historical FanGraphs Guts! table, wOBA, FIP coefficients and constants
}
\examples{
\donttest{
  fg_guts()
}
}
