% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_guts.R
\name{fg_guts}
\alias{fg_guts}
\title{Scrape FanGraphs.com Guts!}
\usage{
fg_guts()
}
\description{
This function allows you to scrape the historical, season-by-season wOBA and FIP constants and coefficients at FanGraphs.com.
}
\examples{
fg_guts()
}
\keyword{MLB,}
\keyword{sabermetrics}
