% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_batter_game_logs.R
\name{fg_milb_batter_game_logs}
\alias{fg_milb_batter_game_logs}
\title{\strong{Scrape MiLB game logs for batters from FanGraphs}}
\usage{
fg_milb_batter_game_logs(playerid, year)
}
\arguments{
\item{playerid}{The batter's minor leauge ID from FanGraphs.}

\item{year}{The season for which game logs should be returned.}
}
\value{
Returns a data frame of Minor League batter game logs.
|col_name       |types     |
|:--------------|:---------|
|player_name    |character |
|minor_playerid |character |
|Date           |character |
|Team           |character |
|Level          |character |
|Opp            |character |
|G              |numeric   |
|AB             |numeric   |
|PA             |numeric   |
|H              |numeric   |
|1B             |numeric   |
|2B             |numeric   |
|3B             |numeric   |
|HR             |numeric   |
|R              |numeric   |
|RBI            |numeric   |
|BB             |numeric   |
|IBB            |numeric   |
|SO             |numeric   |
|HBP            |numeric   |
|SF             |numeric   |
|SH             |numeric   |
|GDP            |numeric   |
|SB             |numeric   |
|CS             |numeric   |
|AVG            |numeric   |
|BB\%            |numeric   |
|K\%             |numeric   |
|BB/K           |numeric   |
|OBP            |numeric   |
|SLG            |numeric   |
|OPS            |numeric   |
|ISO            |numeric   |
|Spd            |numeric   |
|BABIP          |numeric   |
|wRC            |numeric   |
|wRAA           |numeric   |
|wOBA           |numeric   |
|wRC+           |numeric   |
|wBsR           |numeric   |
|gamedate       |character |
|dh             |integer   |
}
\description{
This function allows you to scrape MiLB game logs for individual batters from FanGraphs.
}
\examples{
\donttest{
  fg_milb_batter_game_logs(playerid = "sa917940", year=2018)
}
}
