% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_batter_game_logs.R
\name{fg_milb_batter_game_logs}
\alias{fg_milb_batter_game_logs}
\alias{milb_batter_game_logs_fg}
\title{\strong{Scrape MiLB game logs for batters from Fangraphs}}
\usage{
fg_milb_batter_game_logs(playerid, year)

milb_batter_game_logs_fg(playerid, year)
}
\arguments{
\item{playerid}{The batter's minor leauge ID from FanGraphs.}

\item{year}{The season for which game logs should be returned.}
}
\description{
This function allows you to scrape MiLB game logs for individual batters from FanGraphs.
}
\examples{
\donttest{
  fg_milb_batter_game_logs(playerid = "sa917940", year=2018)
}
}
