% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_park.R
\name{fg_park}
\alias{fg_park}
\title{Scrape Park Factors from FanGraphs.com}
\usage{
fg_park(yr)
}
\arguments{
\item{yr}{Season for which you want to scrape the park factors.}
}
\description{
This function allows you to scrape park factors for a given season from FanGraphs.com.
}
\examples{
fg_park(2013)
}
\keyword{MLB,}
\keyword{sabermetrics}
