% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{fg_park}
\alias{fg_park}
\alias{fg_park_hand}
\title{Scrape Park Factors from FanGraphs.com}
\usage{
fg_park(yr)

fg_park_hand(yr)
}
\arguments{
\item{yr}{Season for which you want to scrape the park factors.}
}
\description{
This function allows you to scrape park factors for a given season from FanGraphs.com.

This function allows you to scrape park factors by handedness from FanGraphs.com for a given single year.
}
\examples{
\donttest{
  fg_park(2013)
}
\donttest{
  fg_park_hand(2013)
}
}
