% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{fg_pitch_leaders}
\alias{fg_pitch_leaders}
\title{Scrape Pitcher Leaderboards from FanGraphs}
\usage{
fg_pitch_leaders(
  x,
  y,
  league = "all",
  qual = "y",
  pitcher_type = "pit",
  ind = 0
)
}
\arguments{
\item{pitcher_type}{Whether you want only starting pitchers, relievers, or all pitchers that meet the criteria specified in the qual argument. Options include "pit", "sta", "rel".}
}
\description{
Scrape Pitcher Leaderboards from FanGraphs
}
\examples{
\dontrun{fg_pitch_leaders(x = 2015, y = 2015, qual = 150)}
}
