% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pitcher_game_logs.R
\name{fg_pitcher_game_logs}
\alias{fg_pitcher_game_logs}
\title{\strong{Scrape Pitcher Game Logs from FanGraphs}}
\usage{
fg_pitcher_game_logs(playerid, year = 2017)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\value{
A data frame of pitcher game logs.
|col_name   |types     |
|:----------|:---------|
|PlayerName |character |
|playerid   |integer   |
|Date       |character |
|Opp        |character |
|teamid     |integer   |
|season     |integer   |
|Team       |character |
|HomeAway   |character |
|Age        |integer   |
|W          |numeric   |
|L          |numeric   |
|ERA        |numeric   |
|G          |numeric   |
|GS         |numeric   |
|CG         |numeric   |
|ShO        |numeric   |
|SV         |numeric   |
|HLD        |numeric   |
|BS         |numeric   |
|IP         |numeric   |
|TBF        |numeric   |
|H          |numeric   |
|R          |numeric   |
|ER         |numeric   |
|HR         |numeric   |
|BB         |numeric   |
|IBB        |numeric   |
|HBP        |numeric   |
|WP         |numeric   |
|BK         |numeric   |
|SO         |numeric   |
|K/9        |numeric   |
|BB/9       |numeric   |
|H/9        |numeric   |
|K/BB       |numeric   |
|IFH\%       |numeric   |
|BUH\%       |numeric   |
|GB         |numeric   |
|FB         |numeric   |
|LD         |numeric   |
|IFFB       |numeric   |
|IFH        |numeric   |
|BU         |numeric   |
|BUH        |numeric   |
|K\%         |numeric   |
|BB\%        |numeric   |
|K-BB\%      |numeric   |
|SIERA      |numeric   |
|HR/9       |numeric   |
|AVG        |numeric   |
|WHIP       |numeric   |
|BABIP      |numeric   |
|LOB\%       |numeric   |
|FIP        |numeric   |
|E-F        |numeric   |
|xFIP       |numeric   |
|ERA-       |numeric   |
|FIP-       |numeric   |
|xFIP-      |numeric   |
|GB/FB      |numeric   |
|LD\%        |numeric   |
|GB\%        |numeric   |
|FB\%        |numeric   |
|IFFB\%      |numeric   |
|HR/FB      |numeric   |
|RS         |numeric   |
|RS/9       |numeric   |
|Balls      |numeric   |
|Strikes    |numeric   |
|Pitches    |numeric   |
|WPA        |numeric   |
|-WPA       |numeric   |
|+WPA       |numeric   |
|RE24       |numeric   |
|REW        |numeric   |
|pLI        |numeric   |
|inLI       |numeric   |
|gmLI       |numeric   |
|exLI       |numeric   |
|Pulls      |numeric   |
|Games      |numeric   |
|WPA/LI     |numeric   |
|Clutch     |numeric   |
|SD         |numeric   |
|MD         |numeric   |
|FB\%1       |numeric   |
|FBv        |numeric   |
|SL\%        |numeric   |
|SLv        |numeric   |
|CT\%        |numeric   |
|CTv        |numeric   |
|CB\%        |numeric   |
|CBv        |numeric   |
|CH\%        |numeric   |
|CHv        |numeric   |
|XX\%        |numeric   |
|PO\%        |numeric   |
|wFB        |numeric   |
|wSL        |numeric   |
|wCT        |numeric   |
|wCB        |numeric   |
|wCH        |numeric   |
|wFB/C      |numeric   |
|wSL/C      |numeric   |
|wCT/C      |numeric   |
|wCB/C      |numeric   |
|wCH/C      |numeric   |
|O-Swing\%   |numeric   |
|Z-Swing\%   |numeric   |
|Swing\%     |numeric   |
|O-Contact\% |numeric   |
|Z-Contact\% |numeric   |
|Contact\%   |numeric   |
|Zone\%      |numeric   |
|F-Strike\%  |numeric   |
|SwStr\%     |numeric   |
|Pull       |numeric   |
|Cent       |numeric   |
|Oppo       |numeric   |
|Soft       |numeric   |
|Med        |numeric   |
|Hard       |numeric   |
|bipCount   |numeric   |
|Pull\%      |numeric   |
|Cent\%      |numeric   |
|Oppo\%      |numeric   |
|Soft\%      |numeric   |
|Med\%       |numeric   |
|Hard\%      |numeric   |
|tERA       |numeric   |
|GSv2       |numeric   |
|Events     |numeric   |
|gamedate   |character |
|dh         |integer   |
}
\description{
This function allows you to scrape game logs by year for a pitcher from FanGraphs.com.
}
\examples{
\donttest{
  fg_pitcher_game_logs(playerid = 104, year = 2006)
}
}
