% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_pitcher_leaders.R
\name{fg_pitcher_leaders}
\alias{fg_pitcher_leaders}
\alias{fg_pitch_leaders}
\title{\strong{Scrape Pitcher Leaderboards from FanGraphs}}
\usage{
fg_pitcher_leaders(
  x,
  y,
  league = "all",
  qual = "y",
  pitcher_type = "pit",
  ind = 1
)

fg_pitch_leaders(
  x,
  y,
  league = "all",
  qual = "y",
  pitcher_type = "pit",
  ind = 1
)
}
\arguments{
\item{x}{First season for which you want data.}

\item{y}{Last season for which you want data. If multiple years selected, data returned will be aggregate data for the date range. If y = x, function will return single-season data.}

\item{league}{Option for limiting results to different leagues or overall results. Options are "al", "nl", or "all".}

\item{qual}{Whether you want only batters/pitchers that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimumm number of plate appearaces/innings pitched, use the number desired. Defaults to "y".}

\item{pitcher_type}{Whether you want only starting pitchers, relievers, or all pitchers that meet the criteria specified in the qual argument. Options include "pit", "sta", "rel".}

\item{ind}{Whether or not to break the seasons out individual, or roll them up together. 1 = split seasons, 0 = aggregate seasons.}
}
\description{
\strong{Scrape Pitcher Leaderboards from FanGraphs}
}
\examples{
\donttest{
  fg_pitcher_leaders(x = 2015, y = 2015, qual = 150)
}
}
