% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_fip_plus.R
\name{fip_plus}
\alias{fip_plus}
\title{\strong{Calculate FIP and related metrics for any set of data}}
\usage{
fip_plus(df)
}
\arguments{
\item{df}{A data frame of statistics that includes, at a minimum, the following columns: IP (innings pitched), BF (batters faced), uBB (unintentional walks), HBP (Hit By Pitch), x1B (singles), x2B (doubles), x3B (triples), HR (home runs), AB (at-bats), SH (sacrafice hits), SO (strike outs), and season.}
}
\description{
This function allows you to calculate FIP and related metrics for any given set of data, provided the right variables are in the data set. The function currently returns both FIP per inning pitched, wOBA against (based on batters faced), and wOBA against per instance of fair contact.
}
