% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_mlb.R
\name{get_draft_mlb}
\alias{get_draft_mlb}
\title{Retrieve draft pick information by year via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
get_draft_mlb(year)
}
\arguments{
\item{year}{The year for which to return data}
}
\value{
Returns a data frame with information for every draft pick in every
round for that year
requested
}
\description{
Retrieve draft pick information by year via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
\dontrun{get_draft_mlb(2009)}
}
\keyword{MLB,}
\keyword{sabermetrics}
