% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_info.R
\name{get_game_info_mlb}
\alias{get_game_info_mlb}
\title{\strong{(legacy) Retrieve additional game information for major and minor league games}}
\usage{
get_game_info_mlb(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a tibble that includes supplemental information, such as
weather, official scorer, attendance, etc., for the game_pk provided
}
\description{
\strong{(legacy) Retrieve additional game information for major and minor league games}
}
\keyword{legacy}
