% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_game_info_sup.R
\name{get_game_info_sup_petti}
\alias{get_game_info_sup_petti}
\title{\strong{(legacy) Download a data frame of supplemental data about MLB games since 2008.}}
\usage{
get_game_info_sup_petti()
}
\value{
Function returns a data frame with various columns, including:
\itemize{
\item game_pk
\item game_date
\item venue id
\item attendance
\item game temperature
\item wind speed
\item direction
\item start time
\item end time
}
}
\description{
\strong{(legacy) Download a data frame of supplemental data about MLB games since 2008.}
}
\keyword{legacy}
