% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_info_sup_petti.R
\name{get_game_info_sup_petti}
\alias{get_game_info_sup_petti}
\title{Download a data frame of supplemental data about MLB games since 2008.}
\usage{
get_game_info_sup_petti()
}
\value{
Function returns a data frame with various columns, including: game_pk,
game_date, venue id, attendance, game temperature, wind speed and direction,
and start and end time,
}
\description{
Download a data frame of supplemental data about MLB games since 2008.
}
\examples{
\dontrun{get_game_info_sup_petti()}
}
\keyword{MLB,}
\keyword{sabermetrics}
