% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_pks.R
\name{get_game_pks_mlb}
\alias{get_game_pks_mlb}
\title{\strong{(legacy) Get MLB Game Info by Date and Level}}
\usage{
get_game_pks_mlb(date, level_ids = c(1))
}
\arguments{
\item{date}{The date for which you want to find game_pk values for MLB games}

\item{level_ids}{A numeric vector with ids for each level where game_pks are
desired. See below for a reference of level ids.}
}
\value{
Returns a tibble that includes game_pk values and additional
information for games scheduled or played
}
\description{
\strong{(legacy) Get MLB Game Info by Date and Level}
}
\keyword{legacy}
