% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_pks_mlb.R
\name{get_game_pks_mlb}
\alias{get_game_pks_mlb}
\title{Find game_pk values for MLB games via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
get_game_pks_mlb(date)
}
\arguments{
\item{date}{The date for which you want to find game_pk values for MLB games}
}
\value{
Returns a data frame that includes game_pk values for the date
requested
}
\description{
Find game_pk values for MLB games via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
get_game_pks("2019-04-29")
}
\keyword{MLB,}
\keyword{sabermetrics}
