% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_baseball_pbp.R
\name{get_ncaa_baseball_pbp}
\alias{get_ncaa_baseball_pbp}
\title{Get Play-By-Play Data for NCAA Baseball Games}
\usage{
get_ncaa_baseball_pbp(game_info_url)
}
\arguments{
\item{game_info_url}{The url for the game's play-by-play data.
This can be found using the get_ncaa_schedule_info function.}
}
\value{
A dataframe with play-by-play data for an individual
game.
}
\description{
Get Play-By-Play Data for NCAA Baseball Games
}
\examples{
\dontrun {get_ncaa_schedule_info(736, 2019)}
}
