% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_baseball_pbp.R
\name{get_ncaa_baseball_pbp}
\alias{get_ncaa_baseball_pbp}
\title{\strong{(legacy) Get Play-By-Play Data for NCAA Baseball Games}}
\usage{
get_ncaa_baseball_pbp(game_info_url)
}
\arguments{
\item{game_info_url}{The url for the game's play-by-play data. This can be
found using the get_ncaa_schedule_info function.}
}
\value{
A data frame with play-by-play data for an individual game.
}
\description{
\strong{(legacy) Get Play-By-Play Data for NCAA Baseball Games}
}
\keyword{legacy}
