% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_baseball_roster.R
\name{get_ncaa_baseball_roster}
\alias{get_ncaa_baseball_roster}
\title{\strong{(legacy) Get NCAA Baseball Rosters}}
\usage{
get_ncaa_baseball_roster(teamid = NA_integer_, team_year, ...)
}
\arguments{
\item{teamid}{NCAA id for a school}

\item{team_year}{The year of interest}

\item{...}{Additional arguments passed to an underlying function like httr.}
}
\value{
A data frame containing roster information, including
IDs and urls for each player (if available)
}
\description{
\strong{(legacy) Get NCAA Baseball Rosters}
}
\keyword{legacy}
