% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_baseball_roster.R
\name{get_ncaa_baseball_roster}
\alias{get_ncaa_baseball_roster}
\title{Get NCAA Baseball Rosters}
\usage{
get_ncaa_baseball_roster(teamid = NA, team_year = 2019)
}
\arguments{
\item{teamid}{NCAA id for a school}

\item{team_year}{The year of interest}
}
\value{
A data frame containing roster information, including
IDs and urls for each player (if available)
}
\description{
Get NCAA Baseball Rosters
}
\examples{
\dontrun{
get_ncaa_baseball_roster(104, 2019)
}
}
