% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_game_logs.R
\name{get_ncaa_game_logs}
\alias{get_ncaa_game_logs}
\title{Get NCAA Baseball Game Logs}
\usage{
get_ncaa_game_logs(player_id, year = 2019, type = "batting")
}
\arguments{
\item{player_id}{A player's unique id. Can be found using the
get_ncaa_baseball_roster function.}

\item{year}{The year of interest.}

\item{type}{The kind of statistics you want to return. Current options
are 'batting' or 'pitching'.}
}
\value{
A data frame containing player and school information
as well as game by game statistics
}
\description{
Get NCAA Baseball Game Logs
}
\examples{
\dontrun{
get_ncaa_game_logs(player_id = 1879650, year = 2019), type = "batting")
}
}
