% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_game_logs.R
\name{get_ncaa_game_logs}
\alias{get_ncaa_game_logs}
\title{\strong{(legacy) Get NCAA Baseball Game Logs}}
\usage{
get_ncaa_game_logs(player_id, year, type = "batting", span = "game", ...)
}
\arguments{
\item{player_id}{A player's unique id. Can be found using the
get_ncaa_baseball_roster function.}

\item{year}{The year of interest.}

\item{type}{The kind of statistics you want to return. Current options
are 'batting' or 'pitching'.}

\item{span}{The span of time; can either be 'game' for game logs in a season, or 'career' which
returns seasonal stats for a player's career.}

\item{...}{Additional arguments passed to an underlying function like httr.}
}
\value{
A data frame containing player and school information
as well as game by game statistics
}
\description{
\strong{(legacy) Get NCAA Baseball Game Logs}
}
\keyword{legacy}
