% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_lineups.R
\name{get_ncaa_lineups}
\alias{get_ncaa_lineups}
\title{Retrieve lineups for a given NCAA game via its game_info_url}
\usage{
get_ncaa_lineups(game_info_url, year)
}
\arguments{
\item{game_info_url}{The unique game info url}

\item{year}{The year the game was played in}
}
\value{
returns a data frame of each school's starting lineup and starting pitcher
}
\description{
Retrieve lineups for a given NCAA game via its game_info_url
}
\examples{
\donttest{
  get_ncaa_lineups(game_info_url="https://stats.ncaa.org/game/index/4587474?org_id=528",year=2018)
}
}
