% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_schedule_info.R
\name{get_ncaa_schedule_info}
\alias{get_ncaa_schedule_info}
\title{\strong{(legacy) Get Schedule and Results for NCAA Baseball Teams}}
\usage{
get_ncaa_schedule_info(team_id = NULL, year = NULL, pbp_links = FALSE, ...)
}
\arguments{
\item{team_id}{The team's unique NCAA id.}

\item{year}{The season (i.e. use 2016 for the 2015-2016 season,
etc.)}

\item{pbp_links}{Logical parameter to run process for scraping play_by_play urls for each game}

\item{...}{Additional arguments passed to an underlying function like httr.}
}
\value{
A data frame with the following fields: date, opponent,
result, score, innings (if more than regulation), and the url
for the game itself.
}
\description{
\strong{(legacy) Get Schedule and Results for NCAA Baseball Teams}
}
\keyword{legacy}
