% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_schedule_info.R
\name{get_ncaa_schedule_info}
\alias{get_ncaa_schedule_info}
\title{Get Schedule and Results for NCAA Baseball Teams}
\usage{
get_ncaa_schedule_info(teamid = NULL, year = NULL)
}
\arguments{
\item{teamid}{The team's unique NCAA id.}

\item{year}{The season (i.e. use 2016 for the 2015-2016 season,
etc.)}
}
\value{
A dataframe with the following fields: date, opponent,
result, score, innings (if more than regulation), and the url
for the game itself.
}
\description{
Get Schedule and Results for NCAA Baseball Teams
}
\examples{
\dontrun{get_ncaa_schedule_info(teamid =736, year = 2017)}
}
