% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pbp.R
\name{get_pbp_mlb}
\alias{get_pbp_mlb}
\title{\strong{(legacy) Acquire pitch-by-pitch data for Major and Minor League games}}
\usage{
get_pbp_mlb(game_pk)
}
\arguments{
\item{game_pk}{The date for which you want to find game_pk values for MLB games}
}
\value{
Returns a data frame that includes over 100 columns of data provided
by the MLB Stats API at a pitch level.
}
\description{
\strong{(legacy) Acquire pitch-by-pitch data for Major and Minor League games}
}
\keyword{legacy}
