% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbp_mlb.R
\name{get_pbp_mlb}
\alias{get_pbp_mlb}
\title{Acquire pitch-by-pitch data for Major and Minor League games via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
get_pbp_mlb(game_pk)
}
\arguments{
\item{game_pk}{The date for which you want to find game_pk values for MLB games}
}
\value{
Returns a data frame that includes over 100 columns of data provided
by the MLB Stats API at a pitch level.

Some data will vary depending on the
park and the league level, as most sensor data is not availble in
minor league parks via this API. Note that the column names have mostly
been left as-is and there are likely duplicate columns in terms of the
information they provide. I plan to clean the output up down the road, but
for now I am leaving the majority as-is.

Both major and minor league pitch-by-pitch data can be pulled with this
 function.
}
\description{
Acquire pitch-by-pitch data for Major and Minor League games via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
\dontrun{get_pbp_mlb(game_pk = 575156)}
}
