% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probables_mlb.R
\name{get_probables_mlb}
\alias{get_probables_mlb}
\title{Retrieve probable starters for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
get_probables_mlb(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame that includes probable starting pitchers and
the home plate umpire for the game_pk provided
requested
}
\description{
Retrieve probable starters for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
get_probables_mlb(game_pk = 566001)
}
\keyword{MLB,}
\keyword{sabermetrics}
