% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_umpire_ids.R
\name{get_umpire_ids_petti}
\alias{get_umpire_ids_petti}
\title{\strong{(legacy) Download a data frame of all umpires and their MLBAM IDs for games since 2008}}
\usage{
get_umpire_ids_petti()
}
\value{
Function returns a data frame with the following columns:
\itemize{
\item id
\item position,
\item name
\item game_pk
\item game_date
}
}
\description{
\strong{(legacy) Download a data frame of all umpires and their MLBAM IDs for games since 2008}
}
\keyword{legacy}
