% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpitchchart.R
\name{ggpitchcharts}
\alias{ggpitchcharts}
\title{\strong{Pitch charts with ggplot2}}
\usage{
ggpitchcharts(
  data,
  x_value = "-plate_x",
  y_value = "plate_z",
  sz_top = NULL,
  sz_bot = NULL,
  color_value = NULL,
  density = FALSE,
  bin_size = 5,
  scall_fill_palette = "Blues",
  point_size = 1,
  frame = "detailed"
)
}
\arguments{
\item{data}{A data that includes ball coordinates.}

\item{x_value}{The x coordinates. Typically plate_x. (If you put '-', it will be a catcher's point of view.)}

\item{y_value}{The y coordinates. Typically plate_z.}

\item{sz_top}{The top of strike zone. Typically sz_top. If left blank, defaults to average height.}

\item{sz_bot}{The top of strike zone. Typically sz_bot. If left blank, defaults to average height.}

\item{color_value}{The categorical variable that you want the geom_points to base the color on. Pass as a string. If left blank, defaults to blue.}

\item{density}{Chooses between a 2d density plot or a point plot. Defaults to FALSE.}

\item{bin_size}{Size of bins used if use a density plot. Defaults to 5.}

\item{scall_fill_palette}{You can use RColorBrewer's palette if use a density plot. Defaults to Blues.}

\item{point_size}{Set the size of geom_point if used.}

\item{frame}{Chooses from detailed, outline or none. Defaults to detailed.}
}
\value{
The pitch charts.
}
\description{
This function allows you to create pitch charts with ggplots given a data ball coordinates and The top and bottom of the strike zone.(The limits of the strike zone can be omitted)
}
