% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_label_statcast_imputed_data.R
\name{label_statcast_imputed_data}
\alias{label_statcast_imputed_data}
\title{\strong{Label Statcast data as imputed}}
\usage{
label_statcast_imputed_data(
  statcast_df,
  impute_file = NULL,
  inverse_precision = 10000
)
}
\arguments{
\item{statcast_df}{A dataframe containing Statcast batted ball data}

\item{impute_file}{A CSV file giving the launch angle, launch speed,
\code{bb_type}, events fields to label
as imputed. if NULL then it's read from the \code{extdata} folder of the package.}

\item{inverse_precision}{inverse of how many digits to truncate the launch angle
and speed to for comparison. Default is \code{10000}, i.e. keep 4 digits of precision.}
}
\value{
A copy of the input dataframe with a new column \code{imputed} appended. imputed
is 1 if launch angle and launch speed are likely imputed, 0 otherwise.
}
\description{
Based on a series of heuristics, this function attempts to
label Statcast data for which the launch angle and speed
have been imputed.
}
\examples{
\donttest{
statcast_df <- scrape_statcast_savant("2017-05-01", "2017-05-02")
sc_df <- label_statcast_imputed_data(statcast_df)
mean(sc_df$imputed)
}
}
