% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncaa_baseball_season_ids.R
\name{load_ncaa_baseball_season_ids}
\alias{load_ncaa_baseball_season_ids}
\title{\strong{Load baseballr men's college baseball season IDs}}
\usage{
load_ncaa_baseball_season_ids(..., dbConnection = NULL, tablename = NULL)
}
\arguments{
\item{...}{Additional arguments passed to an underlying function that writes
the season data into a database (used by \code{update_ncaa_baseball_db()}).}

\item{dbConnection}{A \code{DBIConnection} object, as returned by}

\item{tablename}{The name of the data table within the database}
}
\value{
Returns a tibble
}
\description{
helper that loads multiple seasons of season IDs from the data repo either into memory
or writes it into a db using some forwarded arguments in the dots
}
\examples{
\donttest{
  load_ncaa_baseball_season_ids()
}
}
