% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_pitcher_game_logs.R
\name{milb_pitcher_game_logs_fg}
\alias{milb_pitcher_game_logs_fg}
\title{# Scrape MiLB game logs for pitchers from Fangraphs, combining 'standard' and 'advanced' tabs}
\usage{
milb_pitcher_game_logs_fg(playerid, year)
}
\arguments{
\item{playerid}{The pitcher's minor leauge ID from FanGraphs.com.}

\item{year}{The season for which game logs should be returned.}
}
\description{
This function allows you to scrape MiLB game logs for individual batters from FanGraphs.com.
}
