% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milb_pitcher_game_logs_fg.R
\name{milb_pitcher_game_logs_fg}
\alias{milb_pitcher_game_logs_fg}
\title{# Scrape MiLB game logs for pitchers from Fangraphs, combining 'standard' and 'advanced' tabs}
\usage{
milb_pitcher_game_logs_fg(playerid, year = 2017)
}
\arguments{
\item{playerid}{The pitcher's minor leauge ID from FanGraphs.com.}

\item{year}{The season for which game logs should be returned.}
}
\description{
This function allows you to scrape MILB game logs for individual batters from FanGraphs.com.
}
\examples{
\dontrun{milb_pitcher_game_logs_fg(playerid = "sa873980", 2018)}
}
\keyword{MLB,}
\keyword{sabermetrics}
