% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_attendance.R
\name{mlb_attendance}
\alias{mlb_attendance}
\title{\strong{MLB Attendance}}
\usage{
mlb_attendance(
  team_id = NULL,
  league_id = NULL,
  season = NULL,
  date = NULL,
  league_list_id = NULL
)
}
\arguments{
\item{team_id}{Return attendance information for a particular team_id(s).}

\item{league_id}{Return attendance information for a particular league_id(s). Format: '103,104'}

\item{season}{Return attendance information for particular year(s).}

\item{date}{Return attendance information on a particular date. Format: MM/DD/YYYY}

\item{league_list_id}{Unique league list identifier to return a directory of attendance for a specific league list_id
Valid values include:
\itemize{
\item milb_full
\item milb_short
\item milb_complex
\item milb_all
\item milb_all_nomex
\item milb_all_domestic
\item milb_noncomp
\item milb_noncomp_nomex
\item milb_domcomp
\item milb_intcomp
\item win_noabl
\item win_caribbean
\item win_all
\item abl
\item mlb
\item mlb_hist
\item mlb_milb
\item mlb_milb_hist
\item mlb_milb_win
\item baseball_all
}}
}
\value{
Returns a data frame with the following columns
|col_name                          |types     |
|:---------------------------------|:---------|
|openings_total                    |integer   |
|openings_total_away               |integer   |
|openings_total_home               |integer   |
|openings_total_lost               |integer   |
|games_total                       |integer   |
|games_away_total                  |integer   |
|games_home_total                  |integer   |
|year                              |character |
|attendance_average_away           |integer   |
|attendance_average_home           |integer   |
|attendance_average_ytd            |integer   |
|attendance_high                   |integer   |
|attendance_high_date              |character |
|attendance_low                    |integer   |
|attendance_low_date               |character |
|attendance_opening_average        |integer   |
|attendance_total                  |integer   |
|attendance_total_away             |integer   |
|attendance_total_home             |integer   |
|attendance_high_game_game_pk      |integer   |
|attendance_high_game_link         |character |
|attendance_high_game_day_night    |character |
|attendance_high_game_content_link |character |
|attendance_low_game_game_pk       |integer   |
|attendance_low_game_link          |character |
|attendance_low_game_day_night     |character |
|attendance_low_game_content_link  |character |
|game_type_id                      |character |
|game_type_description             |character |
|team_id                           |integer   |
|team_name                         |character |
|team_link                         |character |
}
\description{
\strong{MLB Attendance}
}
\examples{
\donttest{
  mlb_attendance(team_id = 109, season = 2021)
}
}
