% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_awards.R
\name{mlb_awards}
\alias{mlb_awards}
\title{\strong{MLB Awards}}
\usage{
mlb_awards()
}
\value{
Returns a data frame with the following columns
|col_name          |types     |
|:-----------------|:---------|
|award_id          |character |
|award_name        |character |
|award_description |character |
|sort_order        |integer   |
|notes             |character |
|sport_id          |integer   |
|sport_link        |character |
|league_id         |integer   |
|league_link       |character |
}
\description{
\strong{MLB Awards}
}
\examples{
\donttest{
  mlb_awards()
}
}
