% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_awards.R
\name{mlb_awards}
\alias{mlb_awards}
\title{\strong{MLB Awards}}
\usage{
mlb_awards()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   award_id \tab character \cr
   award_name \tab character \cr
   award_description \tab character \cr
   sort_order \tab integer \cr
   notes \tab character \cr
   sport_id \tab integer \cr
   sport_link \tab character \cr
   league_id \tab integer \cr
   league_link \tab character \cr
}
}
\description{
\strong{MLB Awards}
}
\examples{
\donttest{
  try(mlb_awards())
}
}
