% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_awards_recipient.R
\name{mlb_awards_recipient}
\alias{mlb_awards_recipient}
\title{\strong{MLB Award Recipients}}
\usage{
mlb_awards_recipient(
  award_id = NULL,
  sport_id = NULL,
  league_id = NULL,
  season = NULL
)
}
\arguments{
\item{award_id}{award_id to return a directory of players for a given award.}

\item{sport_id}{sport_id to return a directory of players for a given aware in a specific sport.}

\item{league_id}{league_id(s) to return a directory of players for a given award in a specific league. Format '103,104'}

\item{season}{Year(s) to return a directory of players for a given award in a given season.}
}
\value{
Returns a data frame with the following columns
|col_name                             |types     |
|:------------------------------------|:---------|
|award_id                             |character |
|award_name                           |character |
|date                                 |character |
|season                               |character |
|votes                                |integer   |
|notes                                |character |
|player_id                            |integer   |
|player_link                          |character |
|player_name_first_last               |character |
|player_primary_position_code         |character |
|player_primary_position_name         |character |
|player_primary_position_type         |character |
|player_primary_position_abbreviation |character |
|team_id                              |integer   |
|team_link                            |character |
}
\description{
\strong{MLB Award Recipients}
}
\examples{
\donttest{
  mlb_awards_recipient(award_id = 'MLBHOF', season = 2020)
}
}
