% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_baseball_stats.R
\name{mlb_baseball_stats}
\alias{mlb_baseball_stats}
\title{\strong{MLB Baseball Stats}}
\usage{
mlb_baseball_stats()
}
\value{
Returns a data frame with the following columns
|col_name          |types     |
|:-----------------|:---------|
|stat_name         |character |
|stat_lookup_param |character |
|is_counting       |logical   |
|stat_label        |character |
|stat_group        |character |
}
\description{
\strong{MLB Baseball Stats}
}
\examples{
\donttest{
  mlb_baseball_stats()
}
}
