% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_baseball_stats.R
\name{mlb_baseball_stats}
\alias{mlb_baseball_stats}
\title{\strong{MLB Baseball Stats}}
\usage{
mlb_baseball_stats()
}
\value{
Returns a data frame with the following columns\tabular{ll}{
   col_name \tab types \cr
   stat_name \tab character \cr
   stat_lookup_param \tab character \cr
   is_counting \tab logical \cr
   stat_label \tab character \cr
   stat_group \tab character \cr
}
}
\description{
\strong{MLB Baseball Stats}
}
\examples{
\donttest{
  try(mlb_baseball_stats())
}
}
