% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_batting_orders.R
\name{mlb_batting_orders}
\alias{mlb_batting_orders}
\alias{get_batting_orders}
\title{Retrieve batting orders for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
mlb_batting_orders(game_pk, type = "starting")

get_batting_orders(game_pk, type = "starting")
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}

\item{type}{Whether to just return the starting lineup ('starting') or all
batters that appeared ('all')}
}
\value{
Returns a data frame that includes probable starting pitchers and
the home plate umpire for the game_pk provided
requested

Returns a data frame that includes probable starting pitchers and
the home plate umpire for the game_pk provided
requested
}
\description{
Retrieve batting orders for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}

Retrieve batting orders for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
\donttest{
  mlb_batting_orders(game_pk=566001)
}
}
