% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_conferences.R
\name{mlb_conferences}
\alias{mlb_conferences}
\title{\strong{View all PCL conferences}}
\usage{
mlb_conferences(conference_id = NULL, season = NULL)
}
\arguments{
\item{conference_id}{Conference ID to return information for.}

\item{season}{Year to return to return conference information for.}
}
\value{
Returns a data frame with the following columns
|col_name                |types     |
|:-----------------------|:---------|
|conference_id           |integer   |
|conference_name         |character |
|link                    |character |
|conference_abbreviation |character |
|has_wildcard            |logical   |
|name_short              |character |
|league_id               |integer   |
|league_link             |character |
|sport_id                |integer   |
|sport_link              |character |
}
\description{
\strong{View all PCL conferences}
}
\examples{
\donttest{
  mlb_conferences()
  mlb_conferences(conference_id =  301, season = 2021)
}
}
