% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_conferences.R
\name{mlb_conferences}
\alias{mlb_conferences}
\title{\strong{View all PCL conferences}}
\usage{
mlb_conferences(conference_id = NULL, season = NULL)
}
\arguments{
\item{conference_id}{Conference ID to return information for.}

\item{season}{Year to return to return conference information for.}
}
\value{
Returns a data frame with the following columns\tabular{ll}{
   col_name \tab types \cr
   conference_id \tab integer \cr
   conference_name \tab character \cr
   link \tab character \cr
   conference_abbreviation \tab character \cr
   has_wildcard \tab logical \cr
   name_short \tab character \cr
   league_id \tab integer \cr
   league_link \tab character \cr
   sport_id \tab integer \cr
   sport_link \tab character \cr
}
}
\description{
\strong{View all PCL conferences}
}
\examples{
\donttest{
  mlb_conferences()
  mlb_conferences(conference_id =  301, season = 2020)
}
}
