% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_divisions.R
\name{mlb_divisions}
\alias{mlb_divisions}
\title{\strong{MLB Divisions}}
\usage{
mlb_divisions(division_id = NULL, league_id = NULL, sport_id = NULL)
}
\arguments{
\item{division_id}{Return division(s) data for a specific division}

\item{league_id}{Return division(s) data for all divisions in a specific league}

\item{sport_id}{Return division(s) for all divisions in a specific sport.}
}
\value{
Returns a tibble with the following columns
|col_name              |types     |
|:---------------------|:---------|
|division_id           |integer   |
|division_name         |character |
|season                |character |
|division_name_short   |character |
|division_link         |character |
|division_abbreviation |character |
|has_wildcard          |logical   |
|sort_order            |integer   |
|num_playoff_teams     |integer   |
|active                |logical   |
|league_id             |integer   |
|league_link           |character |
|sport_id              |integer   |
|sport_link            |character |
}
\description{
\strong{MLB Divisions}
}
\examples{
\donttest{
  mlb_divisions(sport_id = 1)
}
}
