% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_draft.R
\name{mlb_draft}
\alias{mlb_draft}
\alias{get_draft_mlb}
\title{Retrieve draft pick information by year via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
mlb_draft(year)

get_draft_mlb(year)
}
\arguments{
\item{year}{The year for which to return data}
}
\value{
Returns a data frame with information for every draft pick in every
round for that year
requested

Returns a data frame with information for every draft pick in every
round for the year requested
}
\description{
Retrieve draft pick information by year via the MLB api \url{http://statsapi.mlb.com/api/}

Retrieve draft pick information by year via the MLB api \url{http://statsapi.mlb.com/api/}
}
