% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_draft.R
\name{mlb_draft}
\alias{mlb_draft}
\title{\strong{Retrieve draft pick information by year}}
\usage{
mlb_draft(year)
}
\arguments{
\item{year}{The year for which to return data}
}
\value{
Returns a data frame with information for every draft pick in every round for the year requested
}
\description{
\strong{Retrieve draft pick information by year}
}
