% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_event_types.R
\name{mlb_event_types}
\alias{mlb_event_types}
\title{\strong{MLB Event Types}}
\usage{
mlb_event_types()
}
\value{
Returns a data frame with the following columns
|col_name           |types     |
|:------------------|:---------|
|plate_appearance   |logical   |
|hit                |logical   |
|event_code         |character |
|base_running_event |logical   |
|event_description  |character |
}
\description{
\strong{MLB Event Types}
}
\examples{
\donttest{
  mlb_event_types()
}
}
