% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_event_types.R
\name{mlb_event_types}
\alias{mlb_event_types}
\title{\strong{MLB Event Types}}
\usage{
mlb_event_types()
}
\value{
Returns a data frame with the following columns\tabular{ll}{
   col_name \tab types \cr
   plate_appearance \tab logical \cr
   hit \tab logical \cr
   event_code \tab character \cr
   base_running_event \tab logical \cr
   event_description \tab character \cr
}
}
\description{
\strong{MLB Event Types}
}
\examples{
\donttest{
  mlb_event_types()
}
}
