% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_fielder_detail_types.R
\name{mlb_fielder_detail_types}
\alias{mlb_fielder_detail_types}
\title{\strong{MLB Fielder Detail Types}}
\usage{
mlb_fielder_detail_types()
}
\value{
Returns a data frame with the following columns
|col_name  |types     |
|:---------|:---------|
|stat_name |character |
|code      |character |
|names     |character |
|chance    |logical   |
|error     |logical   |
}
\description{
\strong{MLB Fielder Detail Types}
}
\examples{
\donttest{
  mlb_fielder_detail_types()
}
}
