% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_fielder_detail_types.R
\name{mlb_fielder_detail_types}
\alias{mlb_fielder_detail_types}
\title{\strong{MLB Fielder Detail Types}}
\usage{
mlb_fielder_detail_types()
}
\value{
Returns a data frame with the following columns\tabular{ll}{
   col_name \tab types \cr
   stat_name \tab character \cr
   code \tab character \cr
   names \tab character \cr
   chance \tab logical \cr
   error \tab logical \cr
}
}
\description{
\strong{MLB Fielder Detail Types}
}
\examples{
\donttest{
  try(mlb_fielder_detail_types())
}
}
