% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_content.R
\name{mlb_game_content}
\alias{mlb_game_content}
\title{\strong{Retrieve additional game content for major and minor league games}}
\usage{
mlb_game_content(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame with the following columns
|col_name            |types     |
|:-------------------|:---------|
|title               |character |
|epg_id              |integer   |
|content_id          |character |
|media_id            |character |
|media_state         |character |
|media_feed_type     |character |
|media_feed_sub_type |character |
|call_letters        |character |
|fox_auth_required   |logical   |
|tbs_auth_required   |logical   |
|espn_auth_required  |logical   |
|fs1auth_required    |logical   |
|mlbn_auth_required  |logical   |
|free_game           |logical   |
|type                |character |
|description         |character |
|rendition_name      |character |
|language            |character |

Returns a data frame that includes game content information
}
\description{
\strong{Retrieve additional game content for major and minor league games}
}
\examples{
\donttest{
 mlb_game_content(game_pk = 566001)
}
}
