% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_info.R
\name{mlb_game_info}
\alias{mlb_game_info}
\alias{get_game_info_mlb}
\title{Retrieve additional game information for major and minor league games via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
mlb_game_info(game_pk)

get_game_info_mlb(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame that includes supplemental information, such as
weather, official scorer, attendance, etc., for the game_pk provided

Returns a data frame that includes supplemental information, such as
weather, official scorer, attendance, etc., for the game_pk provided
}
\description{
Retrieve additional game information for major and minor league games via the MLB api \url{http://statsapi.mlb.com/api/}

Retrieve additional game information for major and minor league games via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
\donttest{
  mlb_game_info(game_pk = 566001)
}
}
