% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_linescore.R
\name{mlb_game_linescore}
\alias{mlb_game_linescore}
\title{\strong{Retrieve game linescores for major and minor league games}}
\usage{
mlb_game_linescore(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame with the following columns
|col_name                                  |types     |
|:-----------------------------------------|:---------|
|game_pk                                   |numeric   |
|home_team_id                              |character |
|home_team_name                            |character |
|away_team_id                              |character |
|away_team_name                            |character |
|num                                       |integer   |
|ordinal_num                               |character |
|home_runs                                 |integer   |
|home_hits                                 |integer   |
|home_errors                               |integer   |
|home_left_on_base                         |integer   |
|away_runs                                 |integer   |
|away_hits                                 |integer   |
|away_errors                               |integer   |
|away_left_on_base                         |integer   |
|home_team_link                            |character |
|home_team_season                          |character |
|home_team_venue_id                        |character |
|home_team_venue_name                      |character |
|home_team_venue_link                      |character |
|home_team_team_code                       |character |
|home_team_file_code                       |character |
|home_team_abbreviation                    |character |
|home_team_team_name                       |character |
|home_team_location_name                   |character |
|home_team_first_year_of_play              |character |
|home_team_league_id                       |character |
|home_team_league_name                     |character |
|home_team_league_link                     |character |
|home_team_division_id                     |character |
|home_team_division_name                   |character |
|home_team_division_link                   |character |
|home_team_sport_id                        |character |
|home_team_sport_link                      |character |
|home_team_sport_name                      |character |
|home_team_short_name                      |character |
|home_team_record_games_played             |character |
|home_team_record_wild_card_games_back     |character |
|home_team_record_league_games_back        |character |
|home_team_record_spring_league_games_back |character |
|home_team_record_sport_games_back         |character |
|home_team_record_division_games_back      |character |
|home_team_record_conference_games_back    |character |
|home_team_record_league_record_wins       |character |
|home_team_record_league_record_losses     |character |
|home_team_record_league_record_pct        |character |
|home_team_record_division_leader          |character |
|home_team_record_wins                     |character |
|home_team_record_losses                   |character |
|home_team_record_winning_percentage       |character |
|home_team_franchise_name                  |character |
|home_team_club_name                       |character |
|home_team_all_star_status                 |character |
|home_team_active                          |character |
|away_team_link                            |character |
|away_team_season                          |character |
|away_team_venue_id                        |character |
|away_team_venue_name                      |character |
|away_team_venue_link                      |character |
|away_team_team_code                       |character |
|away_team_file_code                       |character |
|away_team_abbreviation                    |character |
|away_team_team_name                       |character |
|away_team_location_name                   |character |
|away_team_first_year_of_play              |character |
|away_team_league_id                       |character |
|away_team_league_name                     |character |
|away_team_league_link                     |character |
|away_team_division_id                     |character |
|away_team_division_name                   |character |
|away_team_division_link                   |character |
|away_team_sport_id                        |character |
|away_team_sport_link                      |character |
|away_team_sport_name                      |character |
|away_team_short_name                      |character |
|away_team_record_games_played             |character |
|away_team_record_wild_card_games_back     |character |
|away_team_record_league_games_back        |character |
|away_team_record_spring_league_games_back |character |
|away_team_record_sport_games_back         |character |
|away_team_record_division_games_back      |character |
|away_team_record_conference_games_back    |character |
|away_team_record_league_record_wins       |character |
|away_team_record_league_record_losses     |character |
|away_team_record_league_record_pct        |character |
|away_team_record_division_leader          |character |
|away_team_record_wins                     |character |
|away_team_record_losses                   |character |
|away_team_record_winning_percentage       |character |
|away_team_franchise_name                  |character |
|away_team_club_name                       |character |
|away_team_all_star_status                 |character |
|away_team_active                          |character |
}
\description{
\strong{Retrieve game linescores for major and minor league games}
}
\examples{
\donttest{
 mlb_game_linescore(game_pk = 566001)
}
}
