% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_status_codes.R
\name{mlb_game_status_codes}
\alias{mlb_game_status_codes}
\title{\strong{MLB Game Status Codes}}
\usage{
mlb_game_status_codes()
}
\value{
Returns a data frame with the following columns
|col_name            |types     |
|:-------------------|:---------|
|abstract_game_state |character |
|coded_game_state    |character |
|detailed_state      |character |
|status_code         |character |
|reason              |character |
|abstract_game_code  |character |
}
\description{
\strong{MLB Game Status Codes}
}
\examples{
\donttest{
  mlb_game_status_codes()
}
}
