% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_types.R
\name{mlb_game_types}
\alias{mlb_game_types}
\title{\strong{MLB Game Types}}
\usage{
mlb_game_types()
}
\value{
Returns a data frame with the following columns\tabular{ll}{
   col_name \tab types \cr
   game_type_id \tab character \cr
   game_type_description \tab character \cr
}
}
\description{
\strong{MLB Game Types}
}
\examples{
\donttest{
  mlb_game_types()
}
}
