% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_types.R
\name{mlb_game_types}
\alias{mlb_game_types}
\title{\strong{MLB Game Types}}
\usage{
mlb_game_types()
}
\value{
Returns a data frame with the following columns
|col_name              |types     |
|:---------------------|:---------|
|game_type_id          |character |
|game_type_description |character |
}
\description{
\strong{MLB Game Types}
}
\examples{
\donttest{
  mlb_game_types()
}
}
