% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_game_wp.R
\name{mlb_game_wp}
\alias{mlb_game_wp}
\title{\strong{Acquire win probability for Major and Minor League games}}
\usage{
mlb_game_wp(game_pk, timecode = NULL)
}
\arguments{
\item{game_pk}{The game_pk for the game requested}

\item{timecode}{The time code for the MLB game (format: MMDDYYYY_HHMMSS)}
}
\value{
Returns a data frame that includes time codes from the game_pk requested\tabular{ll}{
   col_name \tab types \cr
   home_team_win_probability \tab numeric \cr
   away_team_win_probability \tab numeric \cr
   home_team_win_probability_added \tab numeric \cr
   at_bat_index \tab integer \cr
   leverage_index \tab numeric \cr
}
}
\description{
\strong{Acquire win probability for Major and Minor League games}
}
\examples{
\donttest{
  try(mlb_game_wp(game_pk = 531060))
}
}
