% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_high_low_types.R
\name{mlb_high_low_types}
\alias{mlb_high_low_types}
\title{\strong{MLB Stat High/Low Types}}
\usage{
mlb_high_low_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   stat_name \tab character \cr
   stat_lookup_param \tab character \cr
   is_counting \tab logical \cr
   stat_label \tab character \cr
   stat_groups \tab list \cr
   org_types \tab list \cr
   high_low_types \tab list \cr
}
}
\description{
\strong{MLB Stat High/Low Types}
}
\examples{
\donttest{
  try(mlb_high_low_types())
}
}
