% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_high_low_types.R
\name{mlb_high_low_types}
\alias{mlb_high_low_types}
\title{\strong{MLB Stat High/Low Types}}
\usage{
mlb_high_low_types()
}
\value{
Returns a tibble with the following columns
|col_name          |types     |
|:-----------------|:---------|
|stat_name         |character |
|stat_lookup_param |character |
|is_counting       |logical   |
|stat_label        |character |
|stat_groups       |list      |
|org_types         |list      |
|high_low_types    |list      |
}
\description{
\strong{MLB Stat High/Low Types}
}
\examples{
\donttest{
  mlb_high_low_types()
}
}
