% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_hit_trajectories.R
\name{mlb_hit_trajectories}
\alias{mlb_hit_trajectories}
\title{\strong{MLB Hit Trajectories}}
\usage{
mlb_hit_trajectories()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   hit_trajectory_code \tab character \cr
   hit_trajectory_description \tab character \cr
}
}
\description{
\strong{MLB Hit Trajectories}
}
\examples{
\donttest{
  mlb_hit_trajectories()
}
}
