% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_homerun_derby.R
\name{mlb_homerun_derby_bracket}
\alias{mlb_homerun_derby_bracket}
\title{\strong{Retrieve Homerun Derby Bracket}}
\usage{
mlb_homerun_derby_bracket(game_pk)
}
\arguments{
\item{game_pk}{The game_pk for which you want to return data}
}
\value{
Returns a data frame with the following columns
|col_name                                      |types     |
|:---------------------------------------------|:---------|
|game_pk                                       |integer   |
|event_name                                    |character |
|event_type_code                               |character |
|event_type_name                               |character |
|event_date                                    |character |
|venue_id                                      |integer   |
|venue_name                                    |character |
|venue_link                                    |character |
|is_multi_day                                  |logical   |
|is_primary_calendar                           |logical   |
|file_code                                     |character |
|event_number                                  |integer   |
|public_facing                                 |logical   |
|round                                         |integer   |
|num_batters                                   |integer   |
|top_seed_complete                             |logical   |
|top_seed_started                              |logical   |
|top_seed_winner                               |logical   |
|top_seed_seed                                 |integer   |
|top_seed_is_winner                            |logical   |
|top_seed_is_complete                          |logical   |
|top_seed_is_started                           |logical   |
|top_seed_num_home_runs                        |integer   |
|top_seed_player_id                            |integer   |
|top_seed_player_full_name                     |character |
|top_seed_player_link                          |character |
|top_seed_top_derby_hit_data_launch_speed      |integer   |
|top_seed_top_derby_hit_data_total_distance    |integer   |
|bottom_seed_complete                          |logical   |
|bottom_seed_started                           |logical   |
|bottom_seed_winner                            |logical   |
|bottom_seed_seed                              |integer   |
|bottom_seed_is_winner                         |logical   |
|bottom_seed_is_complete                       |logical   |
|bottom_seed_is_started                        |logical   |
|bottom_seed_num_home_runs                     |integer   |
|bottom_seed_player_id                         |integer   |
|bottom_seed_player_full_name                  |character |
|bottom_seed_player_link                       |character |
|bottom_seed_top_derby_hit_data_launch_speed   |integer   |
|bottom_seed_top_derby_hit_data_total_distance |integer   |
}
\description{
\strong{Retrieve Homerun Derby Bracket}
}
\examples{
\donttest{
 mlb_homerun_derby_bracket(game_pk = 511101)
}
}
