% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_homerun_derby.R
\name{mlb_homerun_derby_players}
\alias{mlb_homerun_derby_players}
\title{\strong{Retrieve Homerun Derby Players}}
\usage{
mlb_homerun_derby_players(game_pk)
}
\arguments{
\item{game_pk}{The game_pk for which you want to return data}
}
\value{
Returns a data frame with the following columns
|col_name                                       |types     |
|:----------------------------------------------|:---------|
|game_pk                                        |integer   |
|event_name                                     |character |
|event_date                                     |character |
|event_type_code                                |character |
|event_type_name                                |character |
|venue_id                                       |integer   |
|venue_name                                     |character |
|player_id                                      |integer   |
|player_full_name                               |character |
|player_link                                    |character |
|player_first_name                              |character |
|player_last_name                               |character |
|player_primary_number                          |character |
|player_birth_date                              |character |
|player_current_age                             |integer   |
|player_birth_city                              |character |
|player_birth_state_province                    |character |
|player_birth_country                           |character |
|player_height                                  |character |
|player_weight                                  |integer   |
|player_active                                  |logical   |
|player_use_name                                |character |
|player_middle_name                             |character |
|player_boxscore_name                           |character |
|player_nick_name                               |character |
|player_gender                                  |character |
|player_is_player                               |logical   |
|player_is_verified                             |logical   |
|player_draft_year                              |integer   |
|player_pronunciation                           |character |
|player_mlb_debut_date                          |character |
|player_name_first_last                         |character |
|player_name_slug                               |character |
|player_first_last_name                         |character |
|player_last_first_name                         |character |
|player_last_init_name                          |character |
|player_init_last_name                          |character |
|player_full_fml_name                           |character |
|player_full_lfm_name                           |character |
|player_strike_zone_top                         |numeric   |
|player_strike_zone_bottom                      |numeric   |
|player_name_matrilineal                        |character |
|player_current_team_id                         |integer   |
|player_current_team_name                       |character |
|player_current_team_link                       |character |
|player_current_team_season                     |integer   |
|player_current_team_team_code                  |character |
|player_current_team_file_code                  |character |
|player_current_team_abbreviation               |character |
|player_current_team_team_name                  |character |
|player_current_team_location_name              |character |
|player_current_team_first_year_of_play         |character |
|player_current_team_short_name                 |character |
|player_current_team_franchise_name             |character |
|player_current_team_club_name                  |character |
|player_current_team_all_star_status            |character |
|player_current_team_active                     |logical   |
|player_current_team_parent_org_name            |character |
|player_current_team_parent_org_id              |integer   |
|player_current_team_venue_id                   |integer   |
|player_current_team_venue_name                 |character |
|player_current_team_venue_link                 |character |
|player_current_team_spring_venue_id            |integer   |
|player_current_team_spring_venue_link          |character |
|player_current_team_league_id                  |integer   |
|player_current_team_league_name                |character |
|player_current_team_league_link                |character |
|player_current_team_division_id                |integer   |
|player_current_team_division_name              |character |
|player_current_team_division_link              |character |
|player_current_team_sport_id                   |integer   |
|player_current_team_sport_link                 |character |
|player_current_team_sport_name                 |character |
|player_current_team_spring_league_id           |integer   |
|player_current_team_spring_league_name         |character |
|player_current_team_spring_league_link         |character |
|player_current_team_spring_league_abbreviation |character |
|player_primary_position_code                   |character |
|player_primary_position_name                   |character |
|player_primary_position_type                   |character |
|player_primary_position_abbreviation           |character |
|player_bat_side_code                           |character |
|player_bat_side_description                    |character |
|player_pitch_hand_code                         |character |
|player_pitch_hand_description                  |character |
|venue_link                                     |character |
|is_multi_day                                   |logical   |
|is_primary_calendar                            |logical   |
|file_code                                      |character |
|event_number                                   |integer   |
|public_facing                                  |logical   |
}
\description{
\strong{Retrieve Homerun Derby Players}
}
\examples{
\donttest{
 mlb_homerun_derby_players(game_pk = 511101)
}
}
