% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_job_types.R
\name{mlb_job_types}
\alias{mlb_job_types}
\title{\strong{MLB Job Types}}
\usage{
mlb_job_types()
}
\value{
Returns a tibble with the following columns
|col_name   |types     |
|:----------|:---------|
|job_code   |character |
|job        |character |
|sort_order |integer   |
}
\description{
\strong{MLB Job Types}
}
\examples{
\donttest{
  mlb_job_types()
}
}
