% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_jobs.R
\name{mlb_jobs}
\alias{mlb_jobs}
\title{\strong{MLB Jobs}}
\usage{
mlb_jobs(job_type = "UMPR", sport_id = NULL, date = NULL)
}
\arguments{
\item{job_type}{Return information for a given job_type. See \code{mlb_job_types()}}

\item{sport_id}{Return information for a given sport_id.}

\item{date}{Return information for a given date.}
}
\value{
Returns a tibble with the following columns
|col_name         |types     |
|:----------------|:---------|
|jersey_number    |character |
|job              |character |
|job_code         |character |
|title            |character |
|person_id        |integer   |
|person_full_name |character |
|person_link      |character |
}
\description{
\strong{MLB Jobs}
}
\examples{
\donttest{
  mlb_jobs(job_type='UMPR')
}
}
