% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_jobs_datacasters.R
\name{mlb_jobs_datacasters}
\alias{mlb_jobs_datacasters}
\title{\strong{MLB Jobs Datacasters}}
\usage{
mlb_jobs_datacasters(sport_id = NULL, date = NULL)
}
\arguments{
\item{sport_id}{Return information for a given sport_id.}

\item{date}{Return information for a given date.}
}
\value{
Returns a tibble with the following columns
|col_name         |types     |
|:----------------|:---------|
|jersey_number    |character |
|job              |character |
|job_code         |character |
|title            |character |
|person_id        |integer   |
|person_full_name |character |
|person_link      |character |
}
\description{
\strong{MLB Jobs Datacasters}
}
\examples{
\donttest{
  mlb_jobs_datacasters(sport_id=1)
}
}
