% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_jobs_official_scorers.R
\name{mlb_jobs_official_scorers}
\alias{mlb_jobs_official_scorers}
\title{\strong{MLB Jobs Official Scorers}}
\usage{
mlb_jobs_official_scorers(sport_id = NULL, date = NULL)
}
\arguments{
\item{sport_id}{Return information for a given sport_id.}

\item{date}{Return information for a given date.}
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   jersey_number \tab character \cr
   job \tab character \cr
   job_code \tab character \cr
   title \tab character \cr
   person_id \tab integer \cr
   person_full_name \tab character \cr
   person_link \tab character \cr
}
}
\description{
\strong{MLB Jobs Official Scorers}
}
\examples{
\donttest{
  mlb_jobs_official_scorers(sport_id=1)
}
}
