% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_languages.R
\name{mlb_languages}
\alias{mlb_languages}
\title{\strong{MLB API Language Options}}
\usage{
mlb_languages()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   language_name \tab character \cr
   language_code \tab character \cr
   locale \tab character \cr
}
}
\description{
\strong{MLB API Language Options}
}
\examples{
\donttest{
  mlb_languages()
}
}
