% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_languages.R
\name{mlb_languages}
\alias{mlb_languages}
\title{\strong{MLB API Language Options}}
\usage{
mlb_languages()
}
\value{
Returns a tibble with the following columns
|col_name      |types     |
|:-------------|:---------|
|language_name |character |
|language_code |character |
|locale        |character |
}
\description{
\strong{MLB API Language Options}
}
\examples{
\donttest{
  mlb_languages()
}
}
