% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_league.R
\name{mlb_league}
\alias{mlb_league}
\title{\strong{MLB Leagues}}
\usage{
mlb_league(seasons = NULL, sport_id = NULL, league_id = NULL)
}
\arguments{
\item{seasons}{Year(s) to return to return league information for.}

\item{sport_id}{The sport_id to return league information for.}

\item{league_id}{The league_id(s) to return league information for.}
}
\value{
Returns a tibble with the following columns
|col_name      |types     |
|:-------------|:---------|
|leader_type   |character |
}
\description{
\strong{MLB Leagues}
}
\examples{
\donttest{
  mlb_league(seasons = 2021, sport_id = 1)
}
}
