% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_league_leader_types.R
\name{mlb_league_leader_types}
\alias{mlb_league_leader_types}
\title{\strong{MLB League Leader Types}}
\usage{
mlb_league_leader_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   leader_type \tab character \cr
}
}
\description{
\strong{MLB League Leader Types}
}
\examples{
\donttest{
  mlb_league_leader_types()
}
}
