% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_logical_events.R
\name{mlb_logical_events}
\alias{mlb_logical_events}
\title{\strong{MLB Logical Events}}
\usage{
mlb_logical_events()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   event_code \tab character \cr
}
}
\description{
\strong{MLB Logical Events}
}
\examples{
\donttest{
  mlb_logical_events()
}
}
