% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_metrics.R
\name{mlb_metrics}
\alias{mlb_metrics}
\title{\strong{MLB Metrics}}
\usage{
mlb_metrics()
}
\value{
Returns a tibble with the following columns
|col_name    |types     |
|:-----------|:---------|
|metric_name |character |
|metric_id   |integer   |
|stat_group  |character |
|metric_unit |character |
}
\description{
\strong{MLB Metrics}
}
\examples{
\donttest{
  mlb_metrics()
}
}
