% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_metrics.R
\name{mlb_metrics}
\alias{mlb_metrics}
\title{\strong{MLB Metrics}}
\usage{
mlb_metrics()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   metric_name \tab character \cr
   metric_id \tab integer \cr
   stat_group \tab character \cr
   metric_unit \tab character \cr
}
}
\description{
\strong{MLB Metrics}
}
\examples{
\donttest{
  try(mlb_metrics())
}
}
