% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_people.R
\name{mlb_people}
\alias{mlb_people}
\title{\strong{Find Biographical Information for MLB Players}}
\usage{
mlb_people(person_ids = NULL)
}
\arguments{
\item{person_ids}{MLBAMIDs for players of interest. Multiple IDs should be separated by a comma.}
}
\value{
Returns a data frame with the following columns:
|col_name                       |types      |
|:------------------------------|:----------|
}
\description{
\strong{Find Biographical Information for MLB Players}
}
\examples{
\donttest{
  mlb_people(person_ids = 502671)
  mlb_people(person_ids = c(502671,605151))
}
}
