% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_people_free_agents.R
\name{mlb_people_free_agents}
\alias{mlb_people_free_agents}
\title{\strong{Find Information About MLB Free Agents}}
\usage{
mlb_people_free_agents(season = NULL)
}
\arguments{
\item{season}{Season preceding free agency}
}
\value{
Returns a data frame with the following columns:
|col_name               |types      |
|:----------------------|:----------|
}
\description{
\strong{Find Information About MLB Free Agents}
}
\examples{
\donttest{
  mlb_people_free_agents(season = 2018)
}
}
