% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pitch_codes.R
\name{mlb_pitch_codes}
\alias{mlb_pitch_codes}
\title{\strong{MLB Pitch Codes}}
\usage{
mlb_pitch_codes()
}
\value{
Returns a tibble with the following columns
|col_name           |types     |
|:------------------|:---------|
|pitch_code         |character |
|pitch_description  |character |
}
\description{
\strong{MLB Pitch Codes}
}
\examples{
\donttest{
  mlb_pitch_codes()
}
}
