% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pitch_types.R
\name{mlb_pitch_types}
\alias{mlb_pitch_types}
\title{\strong{MLB Pitch Types}}
\usage{
mlb_pitch_types()
}
\value{
Returns a tibble with the following columns
|col_name                |types     |
|:-----------------------|:---------|
|pitch_type_code         |character |
|pitch_type_description  |character |
}
\description{
\strong{MLB Pitch Types}
}
\examples{
\donttest{
  mlb_pitch_types()
}
}
