% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_pitch_types.R
\name{mlb_pitch_types}
\alias{mlb_pitch_types}
\title{\strong{MLB Pitch Types}}
\usage{
mlb_pitch_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   pitch_type_code \tab character \cr
   pitch_type_description \tab character \cr
}
}
\description{
\strong{MLB Pitch Types}
}
\examples{
\donttest{
  try(mlb_pitch_types())
}
}
