% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_player_status_codes.R
\name{mlb_player_status_codes}
\alias{mlb_player_status_codes}
\title{\strong{MLB Player Status Codes}}
\usage{
mlb_player_status_codes()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   player_status_code \tab character \cr
   player_status_description \tab character \cr
}
}
\description{
\strong{MLB Player Status Codes}
}
\examples{
\donttest{
  mlb_player_status_codes()
}
}
