% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_player_status_codes.R
\name{mlb_player_status_codes}
\alias{mlb_player_status_codes}
\title{\strong{MLB Player Status Codes}}
\usage{
mlb_player_status_codes()
}
\value{
Returns a tibble with the following columns
|col_name                   |types     |
|:--------------------------|:---------|
|player_status_code         |character |
|player_status_description  |character |
}
\description{
\strong{MLB Player Status Codes}
}
\examples{
\donttest{
  mlb_player_status_codes()
}
}
