% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_positions.R
\name{mlb_positions}
\alias{mlb_positions}
\title{\strong{MLB Positions}}
\usage{
mlb_positions()
}
\value{
Returns a tibble with the following columns
|col_name              |types     |
|:---------------------|:---------|
|position_short_name   |character |
|position_full_name    |character |
|position_abbreviation |character |
|position_code         |character |
|position_type         |character |
|position_formal_name  |character |
|game_position         |logical   |
|pitcher               |logical   |
|fielder               |logical   |
|outfield              |logical   |
|position_display_name |character |
}
\description{
\strong{MLB Positions}
}
\examples{
\donttest{
  mlb_positions()
}
}
