% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_positions.R
\name{mlb_positions}
\alias{mlb_positions}
\title{\strong{MLB Positions}}
\usage{
mlb_positions()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   position_short_name \tab character \cr
   position_full_name \tab character \cr
   position_abbreviation \tab character \cr
   position_code \tab character \cr
   position_type \tab character \cr
   position_formal_name \tab character \cr
   game_position \tab logical \cr
   pitcher \tab logical \cr
   fielder \tab logical \cr
   outfield \tab logical \cr
   position_display_name \tab character \cr
}
}
\description{
\strong{MLB Positions}
}
\examples{
\donttest{
  mlb_positions()
}
}
