% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_probables.R
\name{mlb_probables}
\alias{mlb_probables}
\title{\strong{Retrieve probable starters for a given MLB game}}
\usage{
mlb_probables(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame that includes probable starting pitchers and
the home plate umpire for the \code{game_pk} requested
|col_name             |types     |
|:--------------------|:---------|
|game_pk              |integer   |
|game_date            |character |
|fullName             |character |
|id                   |integer   |
|team                 |character |
|team_id              |integer   |
|home_plate_full_name |character |
|home_plate_id        |integer   |
}
\description{
\strong{Retrieve probable starters for a given MLB game}
}
\examples{
\donttest{
  mlb_probables(566001)
}
}
