% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_probables.R
\name{mlb_probables}
\alias{mlb_probables}
\title{\strong{Retrieve probable starters for a given MLB game}}
\usage{
mlb_probables(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame that includes probable starting pitchers and
the home plate umpire for the \code{game_pk} requested\tabular{ll}{
   col_name \tab types \cr
   game_pk \tab integer \cr
   game_date \tab character \cr
   fullName \tab character \cr
   id \tab integer \cr
   team \tab character \cr
   team_id \tab integer \cr
   home_plate_full_name \tab character \cr
   home_plate_id \tab integer \cr
}
}
\description{
\strong{Retrieve probable starters for a given MLB game}
}
\examples{
\donttest{
  try(mlb_probables(566001))
}
}
