% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_review_reasons.R
\name{mlb_review_reasons}
\alias{mlb_review_reasons}
\title{\strong{MLB Review Reasons}}
\usage{
mlb_review_reasons()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   review_reason_code \tab character \cr
   review_reason_description \tab character \cr
}
}
\description{
\strong{MLB Review Reasons}
}
\examples{
\donttest{
  try(mlb_review_reasons())
}
}
