% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_review_reasons.R
\name{mlb_review_reasons}
\alias{mlb_review_reasons}
\title{\strong{MLB Review Reasons}}
\usage{
mlb_review_reasons()
}
\value{
Returns a tibble with the following columns
|col_name                  |types     |
|:-------------------------|:---------|
|review_reason_code        |character |
|review_reason_description |character |
}
\description{
\strong{MLB Review Reasons}
}
\examples{
\donttest{
  mlb_review_reasons()
}
}
