% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_rosters.R
\name{mlb_roster_types}
\alias{mlb_roster_types}
\title{\strong{MLB Roster Types}}
\usage{
mlb_roster_types()
}
\value{
Returns a data frame with the following columns
|col_name    |types     |
|:-----------|:---------|
|description |character |
|lookup_name |character |
|parameter   |character |
}
\description{
\strong{MLB Roster Types}
}
\examples{
\donttest{
  mlb_roster_types()
}
}
