% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_roster_types.R
\name{mlb_roster_types}
\alias{mlb_roster_types}
\title{\strong{MLB Roster Types}}
\usage{
mlb_roster_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   roster_type_description \tab character \cr
   roster_type_lookup_name \tab character \cr
   roster_type_parameter \tab character \cr
}
}
\description{
\strong{MLB Roster Types}
}
\examples{
\donttest{
  try(mlb_roster_types())
}
}
