% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_rosters.R
\name{mlb_rosters}
\alias{mlb_rosters}
\title{\strong{Find MLB Rosters by Roster Type}}
\usage{
mlb_rosters(team_id = NULL, season = NULL, date = NULL, roster_type = NULL)
}
\arguments{
\item{team_id}{team_id to return team roster information for a particular club.}

\item{season}{Year to return team roster information for a particular club in a specific season.}

\item{date}{Date to return team roster and their coaching staff directorial information for a particular team.}

\item{roster_type}{roster_type to return team directorial information for. See \code{mlb_roster_types()} for more options.
Valid options include: '40Man', 'fullSeason', 'fullRoster', 'nonRosterInvitees', 'active',
'allTime', 'depthChart', 'gameday', 'coach'}
}
\value{
Returns a data frame with the following columns:
|col_name              |types     |
|:---------------------|:---------|
|jersey_number         |character |
|person_id             |integer   |
|person_full_name      |character |
|person_link           |character |
|position_code         |character |
|position_name         |character |
|position_type         |character |
|position_abbreviation |character |
|status_code           |character |
|status_description    |character |
|link                  |character |
|team_id               |integer   |
|roster_type           |character |
|season                |numeric   |
|date                  |character |
}
\description{
\strong{Find MLB Rosters by Roster Type}
}
\examples{
\donttest{
  mlb_rosters(team_id = 109, season = 2018, roster_type = 'active')
  mlb_rosters(team_id = 109, season = 2018, roster_type = 'coach')
}
}
