% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_runner_detail_types.R
\name{mlb_runner_detail_types}
\alias{mlb_runner_detail_types}
\title{\strong{MLB Runner Detail Types}}
\usage{
mlb_runner_detail_types()
}
\value{
Returns a data frame with the following columns
|col_name  |types     |
|:---------|:---------|
|stat_name |character |
}
\description{
\strong{MLB Runner Detail Types}
}
\examples{
\donttest{
  mlb_runner_detail_types()
}
}
