% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_runner_detail_types.R
\name{mlb_runner_detail_types}
\alias{mlb_runner_detail_types}
\title{\strong{MLB Runner Detail Types}}
\usage{
mlb_runner_detail_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   stat_name \tab character \cr
}
}
\description{
\strong{MLB Runner Detail Types}
}
\examples{
\donttest{
  try(mlb_runner_detail_types())
}
}
