% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_schedule.R
\name{mlb_schedule}
\alias{mlb_schedule}
\title{\strong{Find game_pk values for professional baseball games (major and minor leagues)}}
\usage{
mlb_schedule(season = 2019, level_ids = "1")
}
\arguments{
\item{season}{The season for which you want to find game_pk values for MLB games}

\item{level_ids}{A numeric vector with ids for each level where game_pks are
desired. See below for a reference of level ids.}
}
\value{
Returns a data frame that includes game_pk values and additional
information for games scheduled or played
requested
}
\description{
\strong{Find game_pk values for professional baseball games (major and minor leagues)}
}
\section{Level IDs}{


The following IDs can be passed to the level_ids argument:

1 = MLB \cr
11 = Triple-A \cr
12 = Doubl-A \cr
13 = Class A Advanced \cr
14 = Class A \cr
15 = Class A Short Season \cr
5442 = Rookie Advanced \cr
16 = Rookie \cr
17 = Winter League \cr
}

\examples{
\donttest{
  mlb_schedule(season = "2019")
}
}
