% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_schedule_event_types.R
\name{mlb_schedule_event_types}
\alias{mlb_schedule_event_types}
\title{\strong{MLB Schedule Event Types}}
\usage{
mlb_schedule_event_types()
}
\value{
Returns a tibble with the following columns
|col_name                 |types     |
|:------------------------|:---------|
|schedule_event_type_code |character |
|schedule_event_type_name |character |
}
\description{
\strong{MLB Schedule Event Types}
}
\examples{
\donttest{
  mlb_schedule_event_types()
}
}
