% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_schedule_event_types.R
\name{mlb_schedule_event_types}
\alias{mlb_schedule_event_types}
\title{\strong{MLB Schedule Event Types}}
\usage{
mlb_schedule_event_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   schedule_event_type_code \tab character \cr
   schedule_event_type_name \tab character \cr
}
}
\description{
\strong{MLB Schedule Event Types}
}
\examples{
\donttest{
  try(mlb_schedule_event_types())
}
}
