% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_seasons_all.R
\name{mlb_seasons_all}
\alias{mlb_seasons_all}
\title{\strong{Find MLB Seasons all}}
\usage{
mlb_seasons_all(
  sport_id = 1,
  division_id = NULL,
  league_id = NULL,
  with_game_type_dates = TRUE
)
}
\arguments{
\item{sport_id}{The sport_id to return season information for.}

\item{division_id}{The division_id to return season information for.}

\item{league_id}{The league_id to return season information for.}

\item{with_game_type_dates}{with_game_type_dates to return season information for.}
}
\value{
Returns a tibble with the following columns:\tabular{ll}{
   col_name \tab types \cr
   season_id \tab character \cr
   has_wildcard \tab logical \cr
   pre_season_start_date \tab character \cr
   season_start_date \tab character \cr
   regular_season_start_date \tab character \cr
   regular_season_end_date \tab character \cr
   season_end_date \tab character \cr
   offseason_start_date \tab character \cr
   off_season_end_date \tab character \cr
   season_level_gameday_type \tab character \cr
   game_level_gameday_type \tab character \cr
   qualifier_plate_appearances \tab numeric \cr
   qualifier_outs_pitched \tab integer \cr
   post_season_start_date \tab character \cr
   post_season_end_date \tab character \cr
   last_date1st_half \tab character \cr
   all_star_date \tab character \cr
   first_date2nd_half \tab character \cr
   pre_season_end_date \tab character \cr
   spring_start_date \tab character \cr
   spring_end_date \tab character \cr
}
}
\description{
\strong{Find MLB Seasons all}
}
\examples{
\donttest{
 try(mlb_seasons_all(sport_id = 1))
}
}
