% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_situation_codes.R
\name{mlb_situation_codes}
\alias{mlb_situation_codes}
\title{\strong{MLB Situation Codes}}
\usage{
mlb_situation_codes()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   situation_code \tab character \cr
   sort_order \tab integer \cr
   navigation_menu \tab character \cr
   situation_code_description \tab character \cr
   team \tab logical \cr
   batting \tab logical \cr
   fielding \tab logical \cr
   pitching \tab logical \cr
}
}
\description{
\strong{MLB Situation Codes}
}
\examples{
\donttest{
  mlb_situation_codes()
}
}
