% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_situation_codes.R
\name{mlb_situation_codes}
\alias{mlb_situation_codes}
\title{\strong{MLB Situation Codes}}
\usage{
mlb_situation_codes()
}
\value{
Returns a tibble with the following columns
|col_name                   |types     |
|:--------------------------|:---------|
|situation_code             |character |
|sort_order                 |integer   |
|navigation_menu            |character |
|situation_code_description |character |
|team                       |logical   |
|batting                    |logical   |
|fielding                   |logical   |
|pitching                   |logical   |
}
\description{
\strong{MLB Situation Codes}
}
\examples{
\donttest{
  mlb_situation_codes()
}
}
