% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_sky.R
\name{mlb_sky}
\alias{mlb_sky}
\title{\strong{MLB Sky (Weather) Codes}}
\usage{
mlb_sky()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   sky_code \tab character \cr
   sky_description \tab character \cr
}
}
\description{
\strong{MLB Sky (Weather) Codes}
}
\examples{
\donttest{
  mlb_sky()
}
}
