% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_sports_info.R
\name{mlb_sports_info}
\alias{mlb_sports_info}
\title{\strong{MLB Sport IDs Information}}
\usage{
mlb_sports_info(sport_id = 1)
}
\arguments{
\item{sport_id}{The sport_id to return information for.}
}
\value{
Returns a tibble with the following columns
|col_name           |types     |
|:------------------|:---------|
|sport_id           |integer   |
|sport_code         |character |
|sport_link         |character |
|sport_name         |character |
|sport_abbreviation |character |
|sort_order         |integer   |
|active_status      |logical   |
}
\description{
\strong{MLB Sport IDs Information}
}
\examples{
\donttest{
  mlb_sports_info(sport_id = 1)
}
}
