% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_sports_info.R
\name{mlb_sports_info}
\alias{mlb_sports_info}
\title{\strong{MLB Sport IDs Information}}
\usage{
mlb_sports_info(sport_id = 1)
}
\arguments{
\item{sport_id}{The sport_id to return information for.}
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   sport_id \tab integer \cr
   sport_code \tab character \cr
   sport_link \tab character \cr
   sport_name \tab character \cr
   sport_abbreviation \tab character \cr
   sort_order \tab integer \cr
   active_status \tab logical \cr
}
}
\description{
\strong{MLB Sport IDs Information}
}
\examples{
\donttest{
  mlb_sports_info(sport_id = 1)
}
}
