% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_standings.R
\name{mlb_standings}
\alias{mlb_standings}
\title{\strong{MLB Standings}}
\usage{
mlb_standings(
  season = NULL,
  date = NULL,
  standings_type = NULL,
  league_id = NULL
)
}
\arguments{
\item{season}{Year(s) to return to return standings information for.}

\item{date}{Date to return to return standings information for.}

\item{standings_type}{The standings_type(s) to return standings information for.
\strong{Description of all standings_types}
\enumerate{
\item regularSeason - Regular Season Standings
\item wildCard - Wild card standings
\item divisionLeaders - Division Leader standings
\item wildCardWithLeaders - Wild card standings with Division Leaders
\item firstHalf - First half standings.  Only valid for leagues with a split season (Mexican League).
\item secondHalf - Second half standings. Only valid for leagues with a split season (Mexican League).
\item springTraining - Spring Training Standings
\item postseason - Postseason Standings
\item byDivision - Standings by Division
\item byConference - Standings by Conference
\item byLeague - Standings by League
}}

\item{league_id}{The league_id(s) to return standings information for.}
}
\value{
Returns a tibble with the following columns
|col_name                                  |types     |
|:-----------------------------------------|:---------|
|standings_type                            |character |
|last_updated                              |character |
|team_records_season                       |character |
|team_records_clinch_indicator             |character |
|team_records_division_rank                |character |
|team_records_league_rank                  |character |
|team_records_sport_rank                   |character |
|team_records_games_played                 |integer   |
|team_records_games_back                   |character |
|team_records_wild_card_games_back         |character |
|team_records_league_games_back            |character |
|team_records_spring_league_games_back     |character |
|team_records_sport_games_back             |character |
|team_records_division_games_back          |character |
|team_records_conference_games_back        |character |
|team_records_last_updated                 |character |
|team_records_runs_allowed                 |integer   |
|team_records_runs_scored                  |integer   |
|team_records_division_champ               |logical   |
|team_records_division_leader              |logical   |
|team_records_has_wildcard                 |logical   |
|team_records_clinched                     |logical   |
|team_records_elimination_number           |character |
|team_records_wild_card_elimination_number |character |
|team_records_magic_number                 |character |
|team_records_wins                         |integer   |
|team_records_losses                       |integer   |
|team_records_run_differential             |integer   |
|team_records_winning_percentage           |character |
|team_records_wild_card_rank               |character |
|team_records_wild_card_leader             |logical   |
|team_records_team_id                      |integer   |
|team_records_team_name                    |character |
|team_records_team_link                    |character |
|team_records_streak_streak_type           |character |
|team_records_streak_streak_number         |integer   |
|team_records_streak_streak_code           |character |
|team_records_league_record_wins           |integer   |
|team_records_league_record_losses         |integer   |
|team_records_league_record_ties           |integer   |
|team_records_league_record_pct            |character |
|team_records_records_split_records        |list      |
|team_records_records_division_records     |list      |
|team_records_records_overall_records      |list      |
|team_records_records_league_records       |list      |
|team_records_records_expected_records     |list      |
|league_id                                 |integer   |
|league_link                               |character |
|division_id                               |integer   |
|division_link                             |character |
|sport_id                                  |integer   |
|sport_link                                |character |
}
\description{
\strong{MLB Standings}
}
\examples{
\donttest{
  mlb_standings(season = 2021, league_id = 103)
}
}
