% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_standings_types.R
\name{mlb_standings_types}
\alias{mlb_standings_types}
\title{\strong{MLB Standings Types}}
\usage{
mlb_standings_types()
}
\value{
Returns a tibble with the following columns
|col_name                   |types     |
|:--------------------------|:---------|
|standings_type_name        |character |
|standings_type_description |character |
}
\description{
\strong{MLB Standings Types}
}
\examples{
\donttest{
  mlb_standings_types()
}
}
