% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_stat_groups.R
\name{mlb_stat_groups}
\alias{mlb_stat_groups}
\title{\strong{MLB Stat Groups}}
\usage{
mlb_stat_groups()
}
\value{
Returns a tibble with the following columns
|col_name               |types     |
|:----------------------|:---------|
|stat_group_name        |character |
}
\description{
\strong{MLB Stat Groups}
}
\examples{
\donttest{
  mlb_stat_groups()
}
}
