% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_stat_types.R
\name{mlb_stat_types}
\alias{mlb_stat_types}
\title{\strong{MLB Stat Types}}
\usage{
mlb_stat_types()
}
\value{
Returns a tibble with the following columns
|col_name              |types     |
|:---------------------|:---------|
|stat_type_name        |character |
}
\description{
\strong{MLB Stat Types}
}
\examples{
\donttest{
  mlb_stat_types()
}
}
