% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_stat_types.R
\name{mlb_stat_types}
\alias{mlb_stat_types}
\title{\strong{MLB Stat Types}}
\usage{
mlb_stat_types()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   stat_type_name \tab character \cr
}
}
\description{
\strong{MLB Stat Types}
}
\examples{
\donttest{
  mlb_stat_types()
}
}
