% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_teams.R
\name{mlb_teams}
\alias{mlb_teams}
\title{\strong{MLB Teams}}
\usage{
mlb_teams(
  season = NULL,
  active_status = NULL,
  all_star_statuses = NULL,
  league_ids = NULL,
  sport_ids = NULL,
  game_type = NULL
)
}
\arguments{
\item{season}{Year to return to return team information for.}

\item{active_status}{The active statuses to populate teams for a given season.}

\item{all_star_statuses}{The all-star statuses to populate teams for a given season.}

\item{league_ids}{The league_id(s) to return team information for.}

\item{sport_ids}{The sport_id(s) to return team information for.}

\item{game_type}{The game_type to return team information for.}
}
\value{
Returns a data frame with the following columns
|col_name                   |types     |
|:--------------------------|:---------|
|team_id                    |integer   |
|team_full_name             |character |
|link                       |character |
|season                     |integer   |
|team_code                  |character |
|file_code                  |character |
|team_abbreviation          |character |
|team_name                  |character |
|location_name              |character |
|first_year_of_play         |character |
|short_name                 |character |
|franchise_name             |character |
|club_name                  |character |
|all_star_status            |character |
|active                     |logical   |
|venue_id                   |integer   |
|venue_name                 |character |
|venue_link                 |character |
|spring_venue_id            |integer   |
|spring_venue_link          |character |
|league_id                  |integer   |
|league_name                |character |
|league_link                |character |
|division_id                |integer   |
|division_name              |character |
|division_link              |character |
|sport_id                   |integer   |
|sport_link                 |character |
|sport_name                 |character |
|spring_league_id           |integer   |
|spring_league_name         |character |
|spring_league_link         |character |
|spring_league_abbreviation |character |
}
\description{
\strong{MLB Teams}
}
\examples{
\donttest{
  mlb_teams(season = 2021, sport_ids = c(1))
}
}
