% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_venues.R
\name{mlb_venues}
\alias{mlb_venues}
\title{\strong{Find MLB Venues}}
\usage{
mlb_venues(venue_ids = NULL, sport_ids = NULL, season = NULL)
}
\arguments{
\item{venue_ids}{Venue directorial information based venue_id.}

\item{sport_ids}{The sport_id(s) for which to return venue directorial information.}

\item{season}{Year for which to return venue directorial information for a given season.}
}
\value{
Returns a data frame with the following columns:
|col_name   |types     |
|:----------|:---------|
|venue_id   |integer   |
|venue_name |character |
|venue_link |character |
|active     |logical   |
}
\description{
\strong{Find MLB Venues}
}
\examples{
\donttest{
  mlb_venues()
  mlb_venues(venue_ids = 4781)
  mlb_venues(sport_ids = 1)
}
}
