% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_venues.R
\name{mlb_venues}
\alias{mlb_venues}
\title{\strong{Find MLB Venues}}
\usage{
mlb_venues(venue_ids = NULL, sport_ids = NULL, season = NULL)
}
\arguments{
\item{venue_ids}{Venue directorial information based venue_id.}

\item{sport_ids}{The sport_id(s) for which to return venue directorial information.}

\item{season}{Year for which to return venue directorial information for a given season.}
}
\value{
Returns a data frame with the following columns:\tabular{ll}{
   col_name \tab types \cr
   venue_id \tab integer \cr
   venue_name \tab character \cr
   venue_link \tab character \cr
   active \tab logical \cr
}
}
\description{
\strong{Find MLB Venues}
}
\examples{
\donttest{
  mlb_venues()
  mlb_venues(venue_ids = 4781)
  mlb_venues(sport_ids = 1)
}
}
