% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_wind_direction_codes.R
\name{mlb_wind_direction_codes}
\alias{mlb_wind_direction_codes}
\title{\strong{MLB Wind Direction Codes}}
\usage{
mlb_wind_direction_codes()
}
\value{
Returns a tibble with the following columns
|col_name                   |types     |
|:--------------------------|:---------|
|wind_direction_code        |character |
|wind_direction_description |character |
}
\description{
\strong{MLB Wind Direction Codes}
}
\examples{
\donttest{
  mlb_wind_direction_codes()
}
}
