% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_wind_direction_codes.R
\name{mlb_wind_direction_codes}
\alias{mlb_wind_direction_codes}
\title{\strong{MLB Wind Direction Codes}}
\usage{
mlb_wind_direction_codes()
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   wind_direction_code \tab character \cr
   wind_direction_description \tab character \cr
}
}
\description{
\strong{MLB Wind Direction Codes}
}
\examples{
\donttest{
  mlb_wind_direction_codes()
}
}
