% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_baseball_roster.R
\name{ncaa_baseball_roster}
\alias{ncaa_baseball_roster}
\title{\strong{Get NCAA Baseball Rosters}}
\usage{
ncaa_baseball_roster(teamid = NA, team_year)
}
\arguments{
\item{teamid}{NCAA id for a school}

\item{team_year}{The year of interest}
}
\value{
A data frame containing roster information, including
IDs and urls for each player (if available)
|col_name      |types     |
|:-------------|:---------|
|name          |character |
|class         |character |
|player_id     |character |
|season        |numeric   |
|number        |character |
|position      |character |
|player_url    |character |
|school        |character |
|conference    |character |
|school_id     |numeric   |
|division      |numeric   |
|conference_id |numeric   |
}
\description{
\strong{Get NCAA Baseball Rosters}
}
\examples{
\donttest{
  ncaa_baseball_roster(teamid = 104, team_year = 2021)
}
}
