% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_game_logs.R
\name{ncaa_game_logs}
\alias{ncaa_game_logs}
\title{\strong{Get NCAA Baseball Game Logs}}
\usage{
ncaa_game_logs(player_id, year, type = "batting", span = "game")
}
\arguments{
\item{player_id}{A player's unique id. Can be found using the
get_ncaa_baseball_roster function.}

\item{year}{The year of interest.}

\item{type}{The kind of statistics you want to return. Current options
are 'batting' or 'pitching'.}

\item{span}{The span of time; can either be 'game' for game logs in a season, or 'career' which
returns seasonal stats for a player's career.}
}
\value{
A data frame containing player and school information
as well as game by game statistics
|col_name      |types     |
|:-------------|:---------|
|player_id     |numeric   |
|player_name   |character |
|Date          |character |
|Opponent      |character |
|Result        |character |
|App           |numeric   |
|G             |numeric   |
|GS            |numeric   |
|IP            |numeric   |
|CG            |numeric   |
|H             |numeric   |
|R             |numeric   |
|ER            |numeric   |
|BB            |numeric   |
|SO            |numeric   |
|SHO           |numeric   |
|BF            |numeric   |
|P-OAB         |numeric   |
|2B-A          |numeric   |
|3B-A          |numeric   |
|Bk            |numeric   |
|HR-A          |numeric   |
|WP            |numeric   |
|HB            |numeric   |
|IBB           |numeric   |
|Inh Run       |numeric   |
|Inh Run Score |numeric   |
|SHA           |numeric   |
|SFA           |numeric   |
|Pitches       |numeric   |
|GO            |numeric   |
|FO            |numeric   |
|W             |numeric   |
|L             |numeric   |
|SV            |numeric   |
|OrdAppeared   |numeric   |
|KL            |numeric   |
|pickoffs      |character |
}
\description{
\strong{Get NCAA Baseball Game Logs}
}
\examples{
\donttest{
  ncaa_game_logs(player_id = 2113782, year = 2021, type = "pitching", span = "game")
  ncaa_game_logs(player_id = 1879650, year = 2019, type = "batting", span="career")
}
}
