% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_schedule_info.R
\name{ncaa_schedule_info}
\alias{ncaa_schedule_info}
\alias{get_ncaa_schedule_info}
\title{\strong{Get Schedule and Results for NCAA Baseball Teams}}
\usage{
ncaa_schedule_info(teamid = NULL, year = NULL)

get_ncaa_schedule_info(teamid = NULL, year = NULL)
}
\arguments{
\item{teamid}{The team's unique NCAA id.}

\item{year}{The season (i.e. use 2016 for the 2015-2016 season,
etc.)}
}
\value{
A dataframe with the following fields: date, opponent,
result, score, innings (if more than regulation), and the url
for the game itself.
}
\description{
\strong{Get Schedule and Results for NCAA Baseball Teams}
}
\examples{
\donttest{
  ncaa_schedule_info(teamid = 736, year = 2021)
}
}
