% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_teams.R
\name{ncaa_teams}
\alias{ncaa_teams}
\title{\strong{Scrape NCAA baseball Teams (Division I, II, and III)}}
\usage{
ncaa_teams(year = most_recent_ncaa_baseball_season(), division = 1, ...)
}
\arguments{
\item{year}{The season for which data should be returned, in the form of "YYYY". Years currently available: 2002 onward.}

\item{division}{Division - 1, 2, 3}

\item{...}{Additional arguments passed to an underlying function like httr.}
}
\value{
A data frame with the following variables\tabular{ll}{
   col_name \tab types \cr
   team_id \tab character \cr
   team_name \tab character \cr
   team_url \tab character \cr
   conference_id \tab character \cr
   conference \tab character \cr
   division \tab numeric \cr
   year \tab numeric \cr
   season_id \tab character \cr
}
}
\description{
This function allows the user to obtain NCAA teams by year and division
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{ncaa_teams(2023, 1)
}\if{html}{\out{</div>}}
}
