% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_game_logs_fg.R
\name{pitcher_game_logs_fg}
\alias{pitcher_game_logs_fg}
\title{Scrape Pitcher Game Logs from FanGraphs}
\usage{
pitcher_game_logs_fg(playerid, year = 2017)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\description{
This function allows you to scrape game logs by year for a pitcher from FanGraphs.com.
}
\examples{
\dontrun{pitcher_game_logs_fg(playerid = 104, year = 2006)}
}
\keyword{MLB,}
\keyword{sabermetrics}
