% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_process_statcast_payload.R
\name{process_statcast_payload}
\alias{process_statcast_payload}
\title{\strong{Process Baseball Savant CSV payload}}
\usage{
process_statcast_payload(payload)
}
\arguments{
\item{payload}{payload from a Baseball Savant request}
}
\value{
A tibble with the processed Statcast data coerced to the correct types.
}
\description{
This is a helper function for all \code{statcast_search()} functions.
The function processes the initial csv payload acquired from
Baseball Savant to ensure consistency in formatting across downloads
}
