% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_statcast_payload.R
\name{process_statcast_payload}
\alias{process_statcast_payload}
\title{Process Baseball Savant CSV payload}
\usage{
process_statcast_payload(payload)
}
\arguments{
\item{payload}{payload from a Baseball Savant request, e.g. from utils::read.csv}
}
\description{
This is a helper function for all scrape_statcast_savant functions. The function processes the initial csv payload acquired from baseballsavant to ensure consistency in formattting across downloads
}
\keyword{MLB,}
\keyword{Statcast}
\keyword{sabermetrics,}
