% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progressively}
\alias{progressively}
\title{\strong{Progressively}}
\usage{
progressively(f, p = NULL)
}
\arguments{
\item{f}{a function to add progressr functionality to.}

\item{p}{a progressor function as created by \code{progressr::progressor()}}
}
\value{
a function that does the same as \code{f} but it calls \code{p()} after iteration.
}
\description{
This function helps add progress-reporting to any function - given function \code{f()} and progressor \code{p()}, it will return a new function that calls \code{f()} and then (on-exiting) will call \code{p()} after every iteration.

This is inspired by purrr's \code{safely}, \code{quietly}, and \code{possibly} function decorators.
}
\keyword{internal}
