% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_run_expectancy_table.R
\name{run_expectancy_table}
\alias{run_expectancy_table}
\title{\strong{Generate run expectancy tables from Baseball Savant data}}
\usage{
run_expectancy_table(df, level = "plate appearance")
}
\arguments{
\item{df}{A data frame generated from Baseball Savant that has been formatted using the run_expectancy_code() function.}

\item{level}{Whether you want run expectancy calculated at the plate appearance or pitch level. Defaults to plate appearance.}
}
\description{
These functions allow a user to generate run expectancy tables from Baseball Savant data.
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{run_expectancy_table(df, level = "plate appearances")
}\if{html}{\out{</div>}}
}
