% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_school_id_lu.R
\name{school_id_lu}
\alias{school_id_lu}
\title{\strong{(legacy) Lookup NCAA baseball school IDs (Division I, II, and III)}}
\usage{
school_id_lu(team_name = NULL)
}
\arguments{
\item{team_name}{A string that will be searched for in the names of the teams.}
}
\value{
Returns a tibble with school identification data: team_id, team_name, team_url,
conference, conference_id, division, year, and season_id
}
\description{
\strong{(legacy) Lookup NCAA baseball school IDs (Division I, II, and III)}
}
\keyword{legacy}
