% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_school_id_lu.R
\name{school_id_lu}
\alias{school_id_lu}
\title{\strong{(legacy) Lookup NCAA baseball school IDs (Division I, II, and III)}}
\usage{
school_id_lu(team_name = NULL)
}
\value{
Returns a tibble with school identification data: school, conference, school_id, year, division, conference_id
}
\description{
\strong{(legacy) Lookup NCAA baseball school IDs (Division I, II, and III)}
}
\keyword{legacy}
