% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_id_lu.R
\name{school_id_lu}
\alias{school_id_lu}
\title{Lookup NCAA School IDs (Division I, II, and III)}
\usage{
school_id_lu(school_name = NULL)
}
\arguments{
\item{school_name}{A string that will be searched for in the names of the schools.}
}
\description{
This function allows the user to lookup the school_id needed for the ncaa_scrape function.
}
\examples{
\donttest{school_id_lu("Van")}
}
