% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_statcast_leaderboards.R
\name{scrape_savant_leaderboards}
\alias{scrape_savant_leaderboards}
\title{Query BaseballSavant Leaderboards from \url{http://baseballsavant.mlb.com}}
\usage{
scrape_savant_leaderboards(
  leaderboard = "exit_velocity_barrels",
  year = 2018,
  abs = 50,
  min_pa = "q",
  min_pitches = 100,
  min_field = "q",
  min_run = 0,
  player_type = "batter",
  fielding_type = "player",
  oaa_position = "",
  oaa_roles = "",
  team = "",
  arsenal_type = "n_",
  run_type = "raw",
  min2b = 5,
  min3b = 0,
  position = "",
  bats = "",
  hand = ""
)
}
\arguments{
\item{leaderboard}{The type of leaderboard to retrieve, input as a string. Current
options include exit_velocity_barrels, expected_statistics, pitch_arsenal,
outs_above_average, directional_oaa, catch_probability, pop_time, sprint_speed, and
running_splits_90_ft.}

\item{year}{The season for which you want data.}

\item{abs}{The minimum number of batted balls. Applies only to exit_velocity_barrels
leaderboards.}

\item{min_pa}{Minimum number of plate appearances. Can be a number or 'q' for qualified batters.}

\item{min_pitches}{Minimum number of pitches thrown.}

\item{min_field}{Minimum number of fieding opportunities.}

\item{min_run}{Minimum number of running opportunities.}

\item{player_type}{One of either 'batter' or pitcher. For the expected_statistics
leaderboard, 'batter-team' and 'pitcher-team' are also available.}

\item{fielding_type}{One of either 'player' or 'team'.}

\item{team}{An abbreviation for a team. Can be left blank.}

\item{arsenal_type}{One of either 'n_', 'avg_spin', or 'avg_speed'.}

\item{run_type}{One of either 'percent' or 'raw'.}

\item{min2b}{The minimum number of throwing attempts to second base.}

\item{min3b}{The minimum number of throwing attempts to third base.}

\item{position}{The numeric position of the player. For DH use 10. Can be left blank.}

\item{bats}{The handedness of the batter. One of 'R' or 'L'. Can be left blank.}

\item{hand}{The handedness of the pitcher. One of 'R' or 'L'. Can be left blank.}

\item{ooa_position}{Can be either the number position of a player or 'if' or 'of' for
position categories.}
}
\description{
This function allows you to read leaderboard data from BaseballSavant dirctly into R as #` a \code{\link{data.frame}}.
}
\section{oaa_roles argument}{


30 = 1B - Straight Up  \cr
31 = 1B - Towards 1B/2B Hole \cr
32 = 1B - Close to Line \cr
40 = 2B - Straight Up \cr
41 = 2B - Shaded Towards 2B Bag \cr
42 = 2B - Towards 1B/2B Hole \cr
43 = 2B - Behind First Basemen \cr
46 = 2B - Up the Middle \cr
60 = SS - Straight Up \cr
61 = SS - Towards 3B/SS Hole \cr
62 = SS - Shaded Towards 2B Bag \cr
64 = SS - Up the Middle \cr
50 = 3B - Straight Up \cr
51 = 3B - Close to Line \cr
52 = 3B - Towards 3B/SS Hole \cr
77 = LF - Close to Line \cr
71 = LF - Leaning Left \cr
70 = LF - Straight Up \cr
72 = LF - Leaning Right \cr
78 = LF - LF Gap \cr
87 = CF - LF Gap \cr
81 = CF - Leaning Left \cr
82 = CF - Leaning Right \cr
89 = CF - RF Gap \cr
98 = RF - RF Gap \cr
91 = RF - Leaning Left \cr
90 = RF - Straight Up \cr
92 = RF - Leaning Right \cr
99 = RF - Close to Line \cr
}

\examples{
\dontrun{
scrape_savant_leaderboards(leaderboard = "exit_velocity_barrels", year = 2018)
}
}
\keyword{MLB,}
\keyword{Statcast}
\keyword{sabermetrics,}
