% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_statcast.R
\name{scrape_statcast_savant}
\alias{scrape_statcast_savant}
\alias{scrape_statcast_savant.Date}
\alias{scrape_statcast_savant.default}
\title{Query Statcast and PITCHf/x Data for data from baseballsavant.mlb.com}
\usage{
scrape_statcast_savant(start_date = Sys.Date() - 1, end_date = Sys.Date(),
  playerid = NULL, player_type = "batter", ...)

\method{scrape_statcast_savant}{Date}(start_date = Sys.Date() - 1,
  end_date = Sys.Date(), playerid = NULL, player_type = "batter", ...)

\method{scrape_statcast_savant}{default}(start_date = Sys.Date() - 1,
  end_date = Sys.Date(), playerid = NULL, player_type = "batter", ...)
}
\arguments{
\item{start_date}{Date of first game for which you want data. Format must be in YYYY-MM-DD format.}

\item{end_date}{Date of last game for which you want data. Format must be in YYYY-MM-DD format.}

\item{playerid}{The MLBAM ID for the player whose data you want to query.}

\item{player_type}{The player type. Can be 'batter' or 'pitcher'. Default is 'batter'}

\item{...}{currently ignored}
}
\description{
This function allows you to query Statcast and PITCHf/x data as provided on baseballsavant.mlb.com and have that data returned as a dataframe.
}
\examples{
\dontrun{
correa <- scrape_statcast_savant(start_date = "2016-04-06", 
  end_date = "2016-04-15", playerid = 621043)

noah <- scrape_statcast_savant(start_date = "2016-04-06", 
  end_date = "2016-04-15", playerid = 592789, player_type='pitcher')

daily <- scrape_statcast_savant(start_date = "2016-04-06", end_date = "2016-04-06")
}
}
\keyword{MLB,}
\keyword{Statcast}
\keyword{sabermetrics,}
