% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_standings_on_date.R
\name{standings_on_date_bref}
\alias{standings_on_date_bref}
\title{\strong{(legacy) Scrape MLB Standings on a Given Date}}
\usage{
standings_on_date_bref(date, division, from = FALSE)
}
\arguments{
\item{date}{a date object}

\item{division}{One or more of AL East, AL Central, AL West,
AL Overall, NL East, NL Central, NL West, and NL Overall}

\item{from}{a logical indicating whether you want standings up to and
including the date (FALSE, default) or rather standings for games played
after the date}
}
\value{
Returns a data frame of MLB standings
}
\description{
\strong{(legacy) Scrape MLB Standings on a Given Date}
}
\keyword{legacy}
