% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings_on_date_bref.R
\name{standings_on_date_bref}
\alias{standings_on_date_bref}
\title{Scrape MLB Standings on a Given Date}
\usage{
standings_on_date_bref(y, m, d, division)
}
\arguments{
\item{y}{Year}

\item{m}{Month}

\item{d}{Day}

\item{division}{Division for which you want to scrape standings. Must be in all capitals, such as AL EAST or NL WEST}
}
\description{
This function allows you to scrape the standings from MLB for any date you choose.
}
\examples{
standings_on_date_bref()
}
\keyword{MLB,}
\keyword{standings}

