% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statcast_impute}
\alias{statcast_impute}
\title{Statcast Label Imputation}
\format{
An object of class \code{data.frame} with 44 rows and 4 columns.
}
\usage{
statcast_impute
}
\description{
Statcast Label Imputation
}
\details{
Dataset
}
\keyword{datasets}
