% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_statcast_search.R
\name{statcast_search}
\alias{statcast_search}
\alias{statcast_search.default}
\alias{statcast_search_batters}
\alias{statcast_search_pitchers}
\title{\strong{Query Statcast by Date Range and Players}}
\usage{
statcast_search(
  start_date = Sys.Date() - 1,
  end_date = Sys.Date(),
  playerid = NULL,
  player_type = "batter",
  ...
)

statcast_search.default(
  start_date = Sys.Date() - 1,
  end_date = Sys.Date(),
  playerid = NULL,
  player_type = "batter",
  ...
)

statcast_search_batters(start_date, end_date, batterid = NULL, ...)

statcast_search_pitchers(start_date, end_date, pitcherid = NULL, ...)
}
\arguments{
\item{start_date}{Date of first game for which you want data.
Format must be in YYYY-MM-DD format.}

\item{end_date}{Date of last game for which you want data.
Format must be in YYYY-MM-DD format.}

\item{playerid}{The MLBAM ID for the player whose data you want to query.}

\item{player_type}{The player type. Can be \code{batter} or \code{pitcher}.
Default is \code{batter}}

\item{...}{currently ignored}

\item{batterid}{The MLBAM ID for the batter whose data you want to query.}

\item{pitcherid}{The MLBAM ID for the pitcher whose data you want to query.}
}
\value{
Data returned as a data frame
}
\description{
This function allows you to query Statcast data as provided on \url{https://baseballsavant.mlb.com}
}
\examples{
\donttest{
  ### Correa
  try(statcast_search(start_date = "2016-04-06",
                      end_date = "2016-04-15", 
                      playerid = 621043, 
                      player_type = 'batter'))
  ### Noah
  try(statcast_search(start_date = "2016-04-06",
                      end_date = "2016-04-15", 
                      playerid = 592789, 
                      player_type = 'pitcher'))
  ### Daily
  try(statcast_search(start_date = "2016-04-06", 
                      end_date = "2016-04-06"))
}
\donttest{
  correa <- statcast_search_batters(start_date = "2016-04-06",
    end_date = "2016-04-15", batterid = 621043)
  daily <- statcast_search_batters(start_date = "2016-04-06",
    end_date = "2016-04-06", batterid = NULL)
}
\donttest{
  noah <- statcast_search_pitchers(start_date = "2016-04-06",
    end_date = "2016-04-15", pitcherid = 592789)
  daily <- statcast_search_pitchers(start_date = "2016-04-06",
    end_date = "2016-04-06", pitcherid = NULL)
}
}
