% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats_api_live_empty_df}
\alias{stats_api_live_empty_df}
\title{\strong{Column structure of MLB Stats Live Game API data frame}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 131 columns.
}
\usage{
stats_api_live_empty_df
}
\description{
An empty tibble
}
\keyword{datasets}
