% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_team_consistency.R
\name{team_consistency}
\alias{team_consistency}
\title{\strong{Calculate Team-level Consistency}}
\usage{
team_consistency(year)
}
\arguments{
\item{year}{Season consistency should be run for.}
}
\description{
This function allows you to calculate team-level consistency in run scoring and run prevention over the course of an entire season.
}
\examples{
\donttest{
  team_consistency(year=2015)
}
}
