% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_team_consistency.R
\name{team_consistency}
\alias{team_consistency}
\title{\strong{Calculate Team-level Consistency}}
\usage{
team_consistency(year)
}
\arguments{
\item{year}{Season consistency should be run for.}
}
\value{
Returns a tibble with the following columns\tabular{ll}{
   col_name \tab types \cr
   Team \tab character \cr
   Con_R \tab numeric \cr
   Con_RA \tab numeric \cr
   Con_R_Ptile \tab numeric \cr
   Con_RA_Ptile \tab numeric \cr
}
}
\description{
This function allows you to calculate team-level consistency in run scoring and run prevention over the course of an entire season.
}
\examples{
\donttest{
  try(team_consistency(year=2021))
}
}
