% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_woba_plus.R
\name{woba_plus}
\alias{woba_plus}
\title{\strong{Calcuate wOBA and related metrics for any set of data}}
\usage{
woba_plus(df)
}
\arguments{
\item{df}{A data frame of statistics that includes, at a minimum, the following columns: uBB (unintentional walks), HBP (Hit By Pitch), X1B (singles), X2B (doubles), X3B (triples), HR (home runs), AB (at-bats), SH (sacrafice hits), SO (strike outs), and season.}
}
\description{
This function allows you to calculate wOBA for any given set of data, provided the right variables are in the data set. The function currently returns both wOBA per plate appearance on wOBA per instance of fair contact.
}
\examples{
\donttest{
  df <- bref_daily_batter("2015-08-01", "2015-10-03") 
  woba_plus(df)
}
}
