% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woba_plus.R
\name{woba_plus}
\alias{woba_plus}
\title{Calcuate wOBA for any set of data}
\usage{
woba_plus(df)
}
\description{
This function allows you to calculate wOBA for any given set of data, provided the right variables are in the data set.
}
\details{
wOBA = (wBB×uBB + wHBP×HBP + w1B*x1B + w2B×x2B + w3B×x3B + wHR×HR) / PA
}
\examples{
woba_plus()
}
\keyword{MLB,}
\keyword{standings}

