% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woba_plus.R
\name{woba_plus}
\alias{woba_plus}
\title{Calcuate wOBA for any set of data}
\usage{
woba_plus(df)
}
\arguments{
\item{df}{A data frame of statistics that includes, at a minimum, the following columns: uBB (unintentional walks), HBP (Hit By Pitch), x1B (singles), x2B (doubles), x3B (triples), HR (home runs), AB (at-bats), SH (sacrafice hits), and season.}
}
\description{
This function allows you to calculate wOBA for any given set of data, provided the right variables are in the data set.
}
\keyword{MLB,}
\keyword{sabermetrics}
\keyword{woba,}

